### 2017-11-09 #156 @wlandau-lilly
drake (R package)
[The drake package](https://github.com/wlandau-lilly/drake) is an R-focused [pipeline toolkit](https://github.com/pditommaso/awesome-pipeline). It [reproducibly brings results up to date](https://github.com/wlandau-lilly/drake#reproducibility) and [automatically arranges computations into successive parallelizable stages](https://github.com/wlandau-lilly/drake#high-performance-computing). It has a [Tidyverse-friendly front-end](https://github.com/wlandau-lilly/drake/blob/master/vignettes/quickstart.Rmd#workflow-plan), [powerful interactive visuals](https://cran.r-project.org/web/packages/drake/vignettes/graph.html), and a [vast arsenal of multicore and distributed computing backends](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd).

```
Package: drake
Title: Data Frames in R for Make
Version: 4.4.1.9000
  person(
    family = "Landau",
    given = c("William", "Michael"),
    email = "will.landau@lilly.com",
    role = c("aut", "cre")),
  person(
    family = "Axthelm",
    given = "Alex",
    email = "aaxthelm@che.IN.gov",
    role = "ctb"),
  person(
    family = "Clarkberg",
    given = "Jasper",
    email = "jasper@clarkberg.org",
    role = "ctb"),
  person(
    family = "Eli Lilly and Company",
    role = "cph"))
Description: A solution for reproducible code and
  high-performance computing.
  codetools,
  crayon,
  eply,
  evaluate,
  digest,
  formatR,
  future,
  grDevices,
  igraph,
  parallel,
  plyr,
  R.utils,
  rprojroot,
  storr (>= 1.1.0),
  stringi,
  visNetwork,
  withr
  abind,
  DBI,
  future.batchtools,
  MASS,
  RSQLite,
  tibble
```

- URL: https://github.com/wlandau-lilly/drake

- Fit: [drake](https://github.com/wlandau-lilly/drake) falls easily within reproducibility and high-performance computing.

- Target audience: anyone who uses R for medium-to-long computations for which the results need to stay up to date with the dependencies.

### Similar work

#### Remake

[Drake](https://github.com/wlandau-lilly/drake) overlaps with its direct predecessor, [remake](https://github.com/richfitz/remake). In fact, [drake](https://github.com/wlandau-lilly/drake) owes its core ideas to [remake](https://github.com/richfitz/remake) and @richfitz, and [explicit acknowledgements are in the documentation](https://github.com/wlandau-lilly/drake#acknowledgements-and-related-work). However, [drake](https://github.com/wlandau-lilly/drake) surpasses [remake](https://github.com/richfitz/remake) in several important ways, including but not limited to the following.

1. **High-performance computing**. [Remake](https://github.com/richfitz/remake) has no native parallel computing support. [Drake](https://github.com/wlandau-lilly/drake), on the other hand, has a [vast arsenal](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd) of parallel computing options, from local multicore computing to serious distributed computing. Thanks to [future](github.com/HenrikBengtsson/future), [future.batchtools](github.com/HenrikBengtsson/future.batchtools), and [batchtools](github.com/mllg/batchtools), it is straightforward to configure a [drake](https://github.com/wlandau-lilly/drake) project for most popular job schedulers, such as [SLURM](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/slurm), [TORQUE](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/torque), and the [Sun/Univa Grid Engine](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/sge), as well as systems contained in [Docker images](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/Docker-psock).
1. **A friendly interface**. In [remake](https://github.com/richfitz/remake), the user must manually write a [YAML](https://github.com/richfitz/remake/blob/master/doc/remake.yml) configuration file to arrange the steps of a workflow. In [drake](https://github.com/wlandau-lilly/drake), this configuration is based on data frames that built-in [wildcard templating functionality](https://github.com/wlandau-lilly/drake/blob/master/vignettes/quickstart.Rmd#generate-the-workflow-plan)  easily generates at scale.
1. **Thorough documentation**. [Drake](https://github.com/wlandau-lilly/drake) contains [eight vignettes](https://github.com/wlandau-lilly/drake/tree/master/vignettes), a [comprehensive README](https://github.com/wlandau-lilly/drake/blob/master/README.md), examples in the help files of user-side functions, and [accessible example code](https://github.com/wlandau-lilly/drake/tree/master/inst/examples) that users can write with `drake::example_drake()`.
1. **Active maintenance**. [Drake](https://github.com/wlandau-lilly/drake) is actively developed and maintained, and [issues](https://github.com/wlandau-lilly/drake/issues) are usually solved promptly.

#### Factual's drake

[Factual's drake](https://github.com/Factual/drake) is similar in concept, but the development effort is completely unrelated to the [R package of the same name](https://github.com/wlandau-lilly/drake).

#### Other pipeline toolkits

There are many [other successful pipeline toolkits](https://github.com/pditommaso/awesome-pipeline), and [the drake package](https://github.com/wlandau-lilly/drake) distinguishes itself with its R-focused approach, Tidyverse-friendly interface, and parallel computing flexibility.

I plan to submit to JOSS in the future, but the manuscript is not currently ready.
