### 2017-11-09 57b5da6
Set verbose = FALSE in a make()
That make() is in an unevaluated code chunk.
