### 2017-11-09 d350134
Elaborate on ways around the lack of back...
...compatibility in the upcoming version 5.0.0.
For old projects, users can just
devtools::install_version("drake", "4.4.0").
Mentioned this in migrate() and the [top of the
"caution" vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/caution.Rmd#projects-built-with-drake--440-are-not-back-compatible-with-later-versions)
