### 2017-11-09 03a215a
Address some goodpractice::gp() messages
