### 2017-11-09 41dbd3b
Address more goodpractice::gp() checks
Use lapply() and vapply() instead of sapply().
