### 2017-11-09 d3a2def
Expand console log message for psock cluster.
Change "load parallel socket cluster" to
paste("load parallel socket cluster with", config$jobs, "workers")
