### 2017-11-09 #156 @wlandau-lilly
Thanks, @sckott.

Edit: I think @richfitz would be an ideal reviewer due to the similarity of [remake](https://github.com/richfitz/remake).
