### 2017-11-10 c1013f8
Progress on moving cat() to message()
