+++ R/test_scenarios.R
@@ -95,7 +95,7 @@ test_scenarios <- function(
   for (scenario_name in scenario_names){
     skip <- skip_criterion(scenario_name)
     msg <- ifelse(skip, "skip", "run")
-    cat(scenario_name, ": ", msg, "\n", sep = "")
+    message(scenario_name, ": ", msg, sep = "")
     new <- list()
     new[[test_option_name]] <- scenario_name
     if (!skip) {
