### 2017-11-10 d5f8887
Replace cat() with message()
