### 2017-11-10 ee475f1
Edit notes on reproducibility
