### 2017-11-10 1bbff3e
Change style in test
Minor formatting.
