+++ tests/testthat/test-Makefile.R
@@ -161,10 +161,12 @@ test_with_dir("packages are loaded in prework", {
   options(test_drake_option_12345 = "unset")
   expect_equal(getOption("test_drake_option_12345"), "unset")
   config <- dbug()
-  if (R.utils::isPackageLoaded("abind"))
-  detach("package:abind", unload = TRUE)
-  if (R.utils::isPackageLoaded("MASS"))
-  detach("package:MASS", unload = TRUE)
+  if (R.utils::isPackageLoaded("abind")){
+    detach("package:abind", unload = TRUE)
+  }
+  if (R.utils::isPackageLoaded("MASS")){
+    detach("package:MASS", unload = TRUE)
+  }
   expect_error(abind(1))
   expect_error(deparse(body(lda)))
 
@@ -188,10 +190,12 @@ test_with_dir("packages are loaded in prework", {
   # load packages the usual way
   options(test_drake_option_12345 = "unset")
   expect_equal(getOption("test_drake_option_12345"), "unset")
-  if (R.utils::isPackageLoaded("abind"))
-  detach("package:abind", unload = TRUE)
-  if (R.utils::isPackageLoaded("MASS"))
-  detach("package:MASS", unload = TRUE)
+  if (R.utils::isPackageLoaded("abind")){
+    detach("package:abind", unload = TRUE)
+  }
+  if (R.utils::isPackageLoaded("MASS")){
+    detach("package:MASS", unload = TRUE)
+  }
   expect_error(abind(1))
   expect_error(deparse(body(lda)))
   library(abind)
