### 2017-11-10 #156 @wlandau-lilly
I just ran [goodpractice::gp()](https://github.com/MangoTheCat/goodpractice) on wlandau-lilly/drake@ee475f103f514905d3ed9c3a5dd7b2cacbc46021:

```r
It is good practice to

  x avoid the attach() and detach() functions, they are
    fragile and code that uses them will probably break sooner than
    later.

    tests/testthat/test-Makefile.R:165:3
    tests/testthat/test-Makefile.R:167:3
    tests/testthat/test-Makefile.R:192:3
    tests/testthat/test-Makefile.R:194:3

  x avoid calling setwd(), it changes the global environment.
    If you need it, consider using on.exit() to restore the working
    directory.

    tests/testthat/test-cache.R:43:3
    tests/testthat/test-cache.R:153:3
    tests/testthat/test-cache.R:216:3
    tests/testthat/test-cache.R:225:3
    tests/testthat/test-cache.R:241:3
```

I can explain these idiosyncrasies.

## Calls to `detach()` `tests/testthat/test-Makefile.R`

`Drake` promises to load the user's packages, which is especially important for distributed computing across multiple nodes on a cluster. To test, I occasionally need to call `detach()` in the test scripts to remove packages from `search()`. Unfortunately, `unloadNamespace()` does not have the desired effect.

## Calls to `setwd()` in `tests/testthat/test-cache.R`

By default, `drake` searches through parent directories to find the current `drake` project's [storr cache](https://github.com/richfitz/storr). To test, I need to change directories. However, every test is wrapped in a call to [test_with_dir()](https://github.com/wlandau-lilly/drake/blob/ee475f103f514905d3ed9c3a5dd7b2cacbc46021/R/test.R#L40), which uses `withr::with_dir()` to ensure that the original working directory is restored.
