### 2017-11-11 d0169b6
Add code of conduct
