### 2017-11-11 ddac914
Rearrange sections in README and intro vignette
