### 2017-11-11 #156 @maelle
### Editor checks:

---

#### Editor comments

Thanks for your submission @wlandau-lilly. Running `goodpractice::gp()` is actually my role but now we have your comments (and I get the same flags) so all is good. :wink:

`devtool::spell_check` identified:

* chioces          make.Rd:193

I'm now looking for reviewers.

---

Reviewers:
Due date:
