### 2017-11-11 4f0e944
Fix spelling mistake
chioces => choices
