### 2017-11-11 #156 @wlandau-lilly
Thanks, @maelle! I appreciate your forgiveness regarding `goodpractice::gp()`, and I just fixed the spelling mistake in wlandau-lilly/drake@5c9388a1a7873277332de26a3f8dc0de5bd94104. I will add the badge soon, and I am excited for the review process!
