### 2017-11-11 a7ca60d
Add rOpenSci onboarding badge (under review)
