### 2017-11-12 205b37a
Add comments to the examples in the help files
to prepare for the rOpenSci review. So excited!
This is going to be some of the most thorough feedback
and scrutiny drake has ever received.
