+++ R/dataframes_graph.R
@@ -115,20 +115,25 @@
 #'
 #' @examples
 #' \dontrun{
-#' load_basic_example()
+#' load_basic_example() # Load drake's canonical example.
+#' # Get a list of data frames representing the nodes, edges,
+#' # and legend nodes of the visNetwork graph from plot_graph().
 #' raw_graph <- dataframes_graph(my_plan)
+#' # Choose a subset of the graph.
 #' smaller_raw_graph <- dataframes_graph(
 #'   my_plan,
 #'   from = c("small", "reg2"),
-#'   to = "summ_regression2_small"
+#'   mode = "in"
 #' )
+#' # Inspect the raw graph.
 #' str(raw_graph)
-#' # Plot your own custom visNetwork graph
+#' # Use the data frames to plot your own custom visNetwork graph.
+#' # For example, you can omit the legend nodes
+#' # and change the direction of the graph.
 #' library(magrittr)
 #' library(visNetwork)
 #' visNetwork(nodes = raw_graph$nodes, edges = raw_graph$edges) %>%
-#'   visLegend(useGroups = FALSE, addNodes = raw_graph$legend_nodes) %>%
-#'   visHierarchicalLayout(direction = 'LR')
+#'   visHierarchicalLayout(direction = 'UD')
 #' }
 dataframes_graph <- function(
   plan = workplan(), targets = drake::possible_targets(plan),
@@ -155,8 +160,9 @@ dataframes_graph <- function(
     return(null_graph())
   }
 
-  config$plan <- sanitize_plan(plan = plan)
-  config$targets <- sanitize_targets(plan = plan, targets = targets)
+  config$plan <- sanitize_plan(plan = config$plan)
+  config$targets <- sanitize_targets(
+    plan = config$plan, targets = config$targets)
   if (from_scratch){
     config$outdated <- config$plan$target
   } else {
