### 2017-11-12 fbee850
Document the return values of all exported functions.
Tedious...
