### 2017-11-12 f73c5d9
Fix some changes to the examples
