+++ man/default_short_hash_algo.Rd
@@ -46,7 +46,7 @@ to avoid collisions.
 }
 \examples{
 default_short_hash_algo()
-#' \dontrun{
+\dontrun{
 load_basic_example() # Load the canonical example for drake.
 # Run the project and return the internal master configuration list.
 config <- make(my_plan)
