### 2017-11-12 2138daa
Fix linting
