### 2017-11-12 4ca9f71
Fix #143
Users can now supply a graph argument to make(),
which can save lots of time.
