### 2017-11-12 c2c1571
Fix #144
Still need to add a "skip_imports" argument to make().
