+++ R/config.R
@@ -105,7 +105,7 @@ config <- function(
   trigger <- match.arg(arg = trigger, choices = triggers())
   if (is.null(graph)){
     graph <- build_graph(plan = plan, targets = targets,
-      envir = envir, verbose = verbose, jobs = jobs, trigger = trigger)
+      envir = envir, verbose = verbose, jobs = jobs)
   }
   config <- list(
     plan = plan, targets = targets, envir = envir, cache = cache,
