### 2017-11-12 841fded
Add a skip_imports option to make()
Only for debugging and testing. Also updated
vignettes for this and the fix to #144.
