### 2017-11-12 a8754b8
Lint outdated.R, regenerate Rd, and edit guidelines...
for contributing.
