+++ .github/CONTRIBUTING.md
@@ -4,7 +4,11 @@ Maintainers and contributors must follow this repository's [code of conduct](htt
 
 # Contributions
 
-Suggestions, bug reports, and code are welcome. Please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a [pull request](https://github.com/wlandau-lilly/drake/pulls).
+Suggestions, bug reports, and code are welcome. Please [fork this repository](https://help.github.com/articles/fork-a-repo/) and submit a [pull request](https://github.com/wlandau-lilly/drake/pulls). For pull requests, please consider some ways that you could make life easier for the maintainers.
+
+- Run `devtools::test()` and `covr::package_coverage()` to ensure that the unit tests are still clean and exhaustive.
+- Run `roxygen2::roxygenize()` to update the `.Rd` help files.
+- Check `lintr::lint_package()` and `goodpractice::gp()` to catch stylistic idiosyncracies and potential mistakes.
 
 # Best ways to help
 
