### 2017-11-12 cc5bab9
Move next_targets() out of missing_input_files()
