### 2017-11-12 6e8ccad
Add prune_graph() + a little more speed
