+++ NAMESPACE
@@ -70,6 +70,7 @@ export(possible_targets)
 export(predict_runtime)
 export(progress)
 export(prune)
+export(prune_graph)
 export(r_recipe_wildcard)
 export(rate_limiting_times)
 export(read_config)
