### 2017-11-12 a1e6156
Break up edge case tests
