### 2017-11-12 8c9bcdb
Reorganize more tests
