### 2017-11-12 de4b76f
Document the inputs and outputs of prune_graph()
