### 2017-11-12 2079e19
Tone down overaggressive test
