### 2017-11-12 wlandau-lilly/drake#147 @wlandau-lilly
rOpenSci submission and review
I am delighted to announce that [`drake` is under consideration for entry](https://github.com/ropensci/onboarding/issues/156) into [rOpenSci](https://ropensci.org/)! The process has started, and you can follow it at ropensci/onboarding#156. The initial round of reviews takes about a month, and as of November 11, the reviewers have not yet been chosen.

If `drake` is accepted, there will be more support for package maintenance and promotion in the community. Also, the URL will change to https://github.com/ropensci/drake, and old links to the `wlandau-lilly` URL will automatically redirect to the new location.

I labeled this issue `help wanted` because **this is an excellent time to contribute to `drake`**. All suggestions are welcome, and no contribution is too small or too late. This is already a time of rapid development, and from browsing the [experiences of other package authors](https://github.com/ropensci/onboarding/issues), development is about to accelerate even more. The reviewers are some of the best experts I could hope to scrutinize `drake`, and I expect them to discover major weaknesses. At the end of the review process, I will have learned a ton, and `drake` will emerge stronger than ever, acceptance or not. Afterwards (hopefully after changing all the links to point to https://github.com/ropensci/drake), I will submit the upcoming version 5.0.0 to CRAN (probably sometime in January).
