### 2017-11-12 a5f51d3
Adjust console messaging
