### 2017-11-12 a267d04
Deprecate several functions
expand => expand_plan
evaluate => evaluate_plan
gather => gather_plan
check => check_plan
