### 2017-11-12 d62f315
Fix some checks
