+++ tests/testthat.R
@@ -3,4 +3,4 @@ Sys.setenv("R_TESTS" = "")
 library(testthat)
 library(drake)
 
-test_check_plan("drake")
+test_plan("drake")
