### 2017-11-12 a7700ff
Deprecate session() in favor of drake_session()
