+++ tests/testthat/test-migrate.R
@@ -24,29 +24,29 @@ test_with_dir("force loading a non-back-compatible cache", {
   expect_true(length(cached()) == 0)
 })
 
-test_with_dir("null cases for migrate()", {
-  expect_true(migrate(path = "not_found"))
+test_with_dir("null cases for migrate_drake_project()", {
+  expect_true(migrate_drake_project(path = "not_found"))
   x <- new_cache(path = "path")
-  expect_true(migrate(path = "path"))
+  expect_true(migrate_drake_project(path = "path"))
   expect_silent(tmp <- null_proc_time())
 })
 
-test_with_dir("migrate() an up to date cache", {
+test_with_dir("migrate_drake_project() an up to date cache", {
   write_v4.3.0_project() # nolint
   file.rename(from = default_cache_path(), to = "old")
   expect_error(this_cache(path = "old"))
   cache <- this_cache(path = "old", force = TRUE)
-  expect_true(migrate(path = "old", jobs = 2))
+  expect_true(migrate_drake_project(path = "old", jobs = 2))
   load_basic_example()
   config <- make(my_plan, cache = cache)
  # expect_equal(justbuilt(config = config), character(0)) # r-lib/covr#289 # nolint
   clean(cache = cache)
   expect_equal(cached(cache = cache), character(0))
   # Already migrated, nothing to do
-  expect_true(migrate(path = "old", jobs = 2))
+  expect_true(migrate_drake_project(path = "old", jobs = 2))
 })
 
-test_with_dir("migrate() a partially outdated cache", {
+test_with_dir("migrate_drake_project() a partially outdated cache", {
   write_v4.3.0_project() # nolint
   file.rename(from = default_cache_path(), to = "old")
   cache <- this_cache(path = "old", force = TRUE)
@@ -56,7 +56,7 @@ test_with_dir("migrate() a partially outdated cache", {
   plan <- cache$get("plan", namespace = "config")
   plan$command[plan$target == "small"] <- "simulate(6)"
   cache$set(key = "plan", value = plan, namespace = "config")
-  expect_true(migrate(path = "old", jobs = 2))
+  expect_true(migrate_drake_project(path = "old", jobs = 2))
   out <- c("'report.md'", plan$target[grep("small", plan$target)])
   load_basic_example()
   out2 <- outdated(plan, cache = cache)
