### 2017-11-12 41c50a1
Deprecate session() in favor of drake_session()
