### 2017-11-12 374728b
Deprecate backend() in favor of future_backend()
