+++ R/future.R
@@ -16,7 +16,7 @@ worker_future_lapply <- function(targets, meta_list, config){
   )
 }
 
-#' @title Function \code{backend}
+#' @title Function \code{future_backend}
 #' @export
 #' @seealso \code{\link{workplan}}, \code{\link{make}},
 #' @description Select the specific \code{future}-style
@@ -24,7 +24,7 @@ worker_future_lapply <- function(targets, meta_list, config){
 #' For more information, please read the documentation,
 #' tutorials, and vignettes of the R packages
 #' \code{future} and \code{future.batchtools}.
-#' @details The \code{backend()} function is exactly
+#' @details The \code{future_backend()} function is exactly
 #' the same as \code{future::plan()}.
 #' We provide it only because \code{workplan()} conflicts
 #' with \code{future::plan()}.
@@ -36,22 +36,23 @@ worker_future_lapply <- function(targets, meta_list, config){
 #' # For future-based backends, you will need the "future" package,
 #' # and for high-performance computing, maybe future.batchtools.
 #' library(future)
-#' backend(multicore) # Same as future::plan(multicore)
+#' future_backend(multicore) # Same as future::plan(multicore)
 #' # Run the project with future-powered multicore parallelism.
 #' # The number of jobs is automatic: however many would benefit.
 #' # Set options(mc.cores = 2) to cap the number of jobs at 2.
 #' make(my_plan, parallelism = "future_lapply")
 #' clean() # Erase the targets to start from scratch.
-#' backend(multisession) # Use separate background R sessions.
+#' future_backend(multisession) # Use separate background R sessions.
 #' # This next make() is similar to "parLapply" parallelism.
 #' make(my_plan, parallelism = "future_lapply")
 #' clean() # Remove targets/imports from the cache, start from scratch.
-#' library(future.batchtools) # More heavy-duty future-style parallel backends
+#' # More heavy-duty future-style parallel backends
+#' library(future.batchtools)
 #' # https://github.com/HenrikBengtsson/future.batchtools#choosing-batchtools-backend # nolint
-#' backend(batchtools_local) # Test out future.batchtools locally.
+#' future_backend(batchtools_local) # Test out future.batchtools locally.
 #' # Should proceed prety much like the multisession backend.
 #' make(my_plan, parallelism = "future_lapply")
 #' }
-backend <- function(...){
+future_backend <- function(...){
   future::plan(...)
 }
