### 2017-11-12 80fe175
Deprecate analyses() in favor of plan_analyses()
