### 2017-11-12 95be99d
Deprecate ..dataset.. and ..analysis.. wildcards
in favor of dataset__ and analysis__. These behave
better under linting.
