+++ R/generate.R
@@ -1,3 +1,28 @@
+#' @title Function analysis_wildcard
+#' @description Show the analysis wildcard
+#' used in \code{\link{plan_summaies}()}.
+#' @export
+#' @seealso \code{\link{plan_summaries}()}
+#' @return The analysis wildcard used in \code{\link{plan_summaries}()}.
+#' @examples 
+#' # See ?plan_analyses for examples
+analysis_wildcard <- function(){
+  "analysis__"
+}
+
+#' @title Function dataset_wildcard
+#' @description Show the dataset wildcard
+#' used in \code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
+#' @export
+#' @seealso \code{\link{plan_analyses}()}
+#' @return The dataset wildcard used in
+#' \code{\link{plan_analyses}()} and \code{\link{plan_summaries}()}.
+#' @examples 
+#' # See ?plan_analyses for examples
+dataset_wildcard <- function(){
+  "dataset__"
+}
+
 #' @title Function \code{evaluate_plan}
 #' @description The commands in workflow plan data frames can have
 #' wildcard symbols that can stand for datasets, parameters, function
@@ -44,16 +69,16 @@
 #'   large = simulate(50))
 #' # Create a template workflow plan for the analyses.
 #' methods <- workplan(
-#'   regression1 = reg1(..dataset..),
-#'   regression2 = reg2(..dataset..))
+#'   regression1 = reg1(dataset__),
+#'   regression2 = reg2(dataset__))
 #' # Evaluate the wildcards in the template
 #' # to produce the actual part of the workflow plan
 #' # that encodes the analyses of the datasets.
 #' # Create one analysis for each combination of dataset and method.
-#' evaluate_plan(methods, wildcard = "..dataset..",
+#' evaluate_plan(methods, wildcard = "dataset__",
 #'   values = datasets$target)
 #' # Only choose some combinations of dataset and analysis method.
-#' ans <- evaluate_plan(methods, wildcard = "..dataset..",
+#' ans <- evaluate_plan(methods, wildcard = "dataset__",
 #'   values = datasets$target, expand = FALSE)
 #' ans
 #' # For the complete workflow plan, row bind the pieces together.
@@ -199,14 +224,14 @@ gather_plan <- function(
 #' @title Function \code{plan_analyses}
 #' @description Generate a workflow plan data frame to
 #' analyze multiple datasets using multiple methods of analysis.
-#' @seealso \code{\link{plan_summaries}},
+#' @seealso \code{\link{plan_summaries(}},
 #'  \code{\link{make}}, \code{\link{workplan}}
 #' @export
 #' @return An evaluated workflow plan data frame of analysis targets.
 #' @param plan workflow plan data frame of analysis methods.
 #' The commands in the \code{command} column must
-#' have the \code{..dataset..} wildcard where the datasets go.
-#' For example, one command could be \code{lm(..dataset..)}. Then,
+#' have the \code{dataset__} wildcard where the datasets go.
+#' For example, one command could be \code{lm(dataset__)}. Then,
 #' the commands in the output will include \code{lm(your_dataset_1)},
 #' \code{lm(your_dataset_2)}, etc.
 #' @param datasets workflow plan data frame with instructions
@@ -218,8 +243,8 @@ gather_plan <- function(
 #'   large = simulate(50))
 #' # Create a template for the analysis methods.
 #' methods <- workplan(
-#'   regression1 = reg1(..dataset..),
-#'   regression2 = reg2(..dataset..))
+#'   regression1 = reg1(dataset__),
+#'   regression2 = reg2(dataset__))
[…]
