### 2017-11-12 04150bc
Clean up checks
