+++ man/gather.Rd
@@ -2,11 +2,8 @@
 % Please edit documentation in R/deprecate.R
 \name{gather}
 \alias{gather}
-\alias{gather}
 \title{Deprecated function \code{gather}}
 \usage{
-gather(plan = NULL, target = "target", gather = "list")
-
 gather(plan = NULL, target = "target", gather = "list")
 }
 \arguments{
@@ -15,34 +12,19 @@ gather(plan = NULL, target = "target", gather = "list")
 \item{target}{Same as for \code{\link{gather_plan}}}
 
 \item{gather}{Same as for \code{\link{gather_plan}}}
-
-\item{plan}{Same as for \code{\link{gather_plan}}}
-
-\item{target}{Same as for \code{\link{gather_plan}}}
-
-\item{gather}{Same as for \code{\link{gather_plan}}}
 }
 \value{
-Same as for \code{\link{gather_plan}}
-
 Same as for \code{\link{gather_plan}}
 }
 \description{
 Use \code{\link{gather_plan}()} instead.
-
-Use \code{\link{gather_plan}()} instead.
 }
 \details{
-Deprecated on 2017-11-12.
-
 Deprecated on 2017-11-12.
 }
 \examples{
 # See ?gather_plan for examples.
-# See ?gather_plan for examples.
 }
 \seealso{
-\code{\link{gather_plan}}
-
 \code{\link{gather_plan}}
 }
