### 2017-11-12 ef33fa0
Skim through vignettes
