+++ vignettes/drake.Rmd
@@ -73,7 +73,7 @@ load_basic_example() # Also (over)writes report.Rmd.
 plot_graph(my_plan)  # Hover, click, drag, zoom, pan. See args 'from' and 'to'.
 outdated(my_plan)    # Which targets need to be (re)built?
 missed(my_plan)      # Are you missing anything from your workspace?
-check_plan(my_plan)       # Are you missing files? Is your workflow plan okay?
+check_plan(my_plan)  # Are you missing files? Is your workflow plan okay?
 make(my_plan)        # Run the workflow.
 diagnose(large)      # View error info if the target "large" failed to build.
 outdated(my_plan)    # Everything is up to date.
