+++ vignettes/caution.Rmd
@@ -46,33 +46,33 @@ A common pitfall is using the `evaluate_plan()` function to expand wildcards aft
 ```{r filethenevaluate}
 library(magrittr) # for the pipe operator %>%
 workplan(
-  data = readRDS("data_..datasize...rds")
+  data = readRDS("data_DATASIZE__rds")
 ) %>%
   rbind(drake::workplan(
     file.csv = write.csv(
-      data_..datasize.., # nolint
-      "file_..datasize...csv"
+      data_DATASIZE__, # nolint
+      "file_DATASIZE__csv"
     ),
     strings_in_dots = "literals",
     file_targets = T
   )) %>%
   evaluate_plan(
-    rules = list(..datasize.. = c("small", "large"))
+    rules = list(DATASIZE__ = c("small", "large"))
   )
 ```
 
 The single quotes in the middle of `'file.csv'_small` and `'file.csv'large` are illegal, and the target names do not even correspond to the files written. Instead, construct your workflow plan in multiple stages and apply the single quotes at the very end.
 
 ```{r correctevaldatasize}
-rules <- list(..datasize.. = c("small", "large"))
-datasets <- workplan(data = readRDS("data_..datasize...rds")) %>%
+rules <- list(DATASIZE__ = c("small", "large"))
+datasets <- workplan(data = readRDS("data_DATASIZE__rds")) %>%
   evaluate_plan(rules = rules)
 ```
 
 Plan the CSV files separately.
 ```{r correctevaldatasize2}
 files <- workplan(
-  file = write.csv(data_..datasize.., "file_..datasize...csv"), # nolint
+  file = write.csv(data_DATASIZE__, "file_DATASIZE__csv"), # nolint
   strings_in_dots = "literals"
 ) %>%
   evaluate_plan(rules = rules)
@@ -141,7 +141,7 @@ error$calls # View the traceback.
 
 As of version 3.0.0, drake's execution environment is the user's workspace by default. As an upshot, the workspace is vulnerable to side-effects of `make()`. To protect your workspace, you may want to create a custom evaluation environment containing all your imported objects and then pass it to the envir argument of `make()`. Here is how.
 
-```{r envir}
+```{r envircaution, eval = FALSE}
 library(drake)
 envir <- new.env(parent = globalenv())
 eval(expression({
