### 2017-11-12 f88d255
Revert back to previous code for prune_graph()
igraph::make_ego_graph() was slow.
