+++ inst/examples/torque/run.R
@@ -6,7 +6,7 @@ library(future.batchtools)
 library(drake)
 
 # Use backend(multicore) instead for a dry run.
-future_backend(batchtools_torque(template = "torque.tmpl"))
+future::plan(batchtools_torque(template = "torque.tmpl"))
 
 load_basic_example()
 make(
