### 2017-11-12 3c73c7e
Fix linting
