### 2017-11-12 ebea619
Fix installation instructions
