### 2017-11-12 bde8150
Remove rescue_cache() from clean()
Was taking too long. Addresses #149.
