+++ vignettes/drake.Rmd
@@ -54,6 +54,8 @@ error$calls              # Call stack / traceback
 
 # Installation
 
+You can choose among different versions of `drake`:
+
 ```{r devinstall, eval = FALSE}
 install.packages("drake")    # Latest CRAN release.
 install.packages("devtools") # For installing from GitHub.
