### 2017-11-12 05df65e
Use '@keywords internal' for all deprecated functions
