+++ man/analyses.Rd
@@ -26,3 +26,4 @@ Deprecated on 2017-11-12.
 \seealso{
 \code{\link{plan_analyses}}
 }
+\keyword{internal}
