+++ man/summaries.Rd
@@ -30,3 +30,4 @@ Deprecated on 2017-11-12.
 \seealso{
 \code{\link{summaries}}
 }
+\keyword{internal}
