### 2017-11-13 042a710
Deprecate build_graph() in favor of build_drake_graph()
