### 2017-11-13 285071c
Change prune_graph() to prune_drake_graph()
