+++ NAMESPACE
@@ -80,7 +80,7 @@ export(plot_graph)
 export(possible_targets)
 export(predict_runtime)
 export(progress)
-export(prune_graph)
+export(prune_drake_graph)
 export(r_recipe_wildcard)
 export(rate_limiting_times)
 export(read_config)
