### 2017-11-13 83f25b3
Deprecate plot_graph() in favor of vis_drake_graph()
