+++ R/color.R
@@ -1,6 +1,6 @@
 # Below, the colors from "target" through
 # "fail" are for the console. The rest
-# are for plot_graph().
+# are for vis_drake_graph().
 colors <- c(
   target = "green3",
   import = "dodgerblue3",
@@ -25,12 +25,12 @@ colors <- c(
 #' @title Function palette
 #' @export
 #' @description show color palette for drake.
-#' Used in both the console and \code{\link{plot_graph}()}
+#' Used in both the console and \code{\link{vis_drake_graph}()}
 #' Your console must have the crayon package enabled.
 #' @details This palette applies to console output
 #' (internal functions \code{console()} and
 #' \code{console_many_targets()}) and the node colors
-#' in \code{\link{plot_graph}()}.
+#' in \code{\link{vis_drake_graph}()}.
 #' So if you want to contribute improvements to the palette,
 #' please both \code{drake_palette()} and
 #' \code{visNetwork::visNetwork(nodes = \link{legend_nodes}())}
@@ -40,7 +40,7 @@ colors <- c(
 #' # Show drake's color palette as text.
 #' drake_palette()
 #' # Show part of the palette as an interactive visNetwork graph.
-#' # These are the nodes in the legend of plot_graph().
+#' # These are the nodes in the legend of vis_drake_graph().
 #' visNetwork::visNetwork(nodes = legend_nodes())
 drake_palette <- function(){
   out <- lapply(
