+++ R/sanitize.R
@@ -43,7 +43,7 @@ sanitize_nodes <- function(nodes, choices){
     stop(
       "All import/target names are invalid ",
       "in argument 'targets', 'from', or 'subset' ",
-      "for make(), plot_graph(), or similar.",
+      "for make(), vis_drake_graph(), or similar.",
       call. = FALSE
     )
   }
