### 2017-11-13 2291df1
Deprecate render_graph() in favor of render_drake_graph()
