+++ R/plot_graph.R
@@ -2,7 +2,7 @@
 #' @description Plot the dependency structure of your workplan.
 #' IMPORTANT: you must be in the root directory of your project.
 #' To save time for repeated plotting, this function is divided into
-#' \code{\link{dataframes_graph}()} and \code{\link{render_graph}()}.
+#' \code{\link{dataframes_graph}()} and \code{\link{render_drake_graph}()}.
 #' @export
 #' @seealso \code{\link{build_drake_graph}}
 #' @return A visNetwork graph.
@@ -198,13 +198,13 @@ vis_drake_graph <- function(
   if (is.null(main)){
     main <- raw_graph$default_title
   }
-  render_graph(raw_graph, file = file, selfcontained = selfcontained,
+  render_drake_graph(raw_graph, file = file, selfcontained = selfcontained,
     layout = layout, direction = direction,
     navigationButtons = navigationButtons, # nolint
     hover = hover, main = main, ncol_legend = ncol_legend, ...)
 }
 
-#' @title Function \code{render_graph}
+#' @title Function \code{render_drake_graph}
 #' @description render a graph from the data frames
 #' generated by \code{\link{dataframes_graph}()}
 #' @export
@@ -256,11 +256,11 @@ vis_drake_graph <- function(
 #' # Instead of jumpting right to vis_drake_graph(), get the data frames
 #' # of nodes, edges, and legend nodes.
 #' graph <- dataframes_graph(my_plan)
-#' # You can pass the data frames right to render_graph()
+#' # You can pass the data frames right to render_drake_graph()
 #' # (as in vis_drake_graph()) or you can create your own custom visNewtork graph.
-#' render_graph(graph, width = '100%') # The width is passed to visNetwork.
+#' render_drake_graph(graph, width = '100%') # The width is passed to visNetwork.
 #' }
-render_graph <- function(
+render_drake_graph <- function(
   graph_dataframes, file = character(0),
   layout = "layout_with_sugiyama", direction = "LR", hover = TRUE,
   main = graph_dataframes$default_title, selfcontained = FALSE,
