+++ man/build_drake_graph.Rd
@@ -1,0 +1,56 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/graph.R
+\name{build_drake_graph}
+\alias{build_drake_graph}
+\title{Function \code{build_drake_graph}}
+\usage{
+build_drake_graph(plan = workplan(),
+  targets = drake::possible_targets(plan), envir = parent.frame(),
+  verbose = TRUE, jobs = 1, skip_imports = FALSE)
+}
+\arguments{
+\item{plan}{workflow plan data frame, same as for function
+\code{\link{make}()}.}
+
+\item{targets}{names of targets to build, same as for function
+\code{\link{make}()}.}
+
+\item{envir}{environment to import from, same as for function
+\code{\link{make}()}.}
+
+\item{verbose}{logical, whether to output messages to the console.}
+
+\item{jobs}{number of jobs to accelerate the construction
+of the dependency graph. A light \code{mclapply}-based
+parallelism is used if your operating system is not Windows.}
+
+\item{skip_imports}{logical, whether to totally neglect to
+process the imports and jump straight to the targets. This can be useful
+if your imports are massive and you just want to test your project,
+but it is bad practice for reproducible data analysis.}
+}
+\value{
+An igraph object representing
+the workflow plan dependency network.
+}
+\description{
+Make a graph of the dependency structure of your workplan.
+}
+\details{
+This function returns an igraph object representing how
+the targets in your workplan depend on each other.
+(\code{help(package = "igraph")}). To plot the graph, call
+to \code{\link{plot.igraph}()} on your graph, or just use
+\code{\link{vis_drake_graph}()} from the start.
+}
+\examples{
+\dontrun{
+load_basic_example() # Load the canonical example for drake.
+# Make the igraph network connecting all the targets and imports.
+g <- build_drake_graph(my_plan)
+class(g) # "igraph"
+}
+}
+\seealso{
+\code{\link{vis_drake_graph}}
+}
