+++ vignettes/drake.Rmd
@@ -123,7 +123,7 @@ outdated()
 missed()
 vis_drake_graph() # Now with subgraphs too.
 dataframes_graph()
-render_graph()
+render_drake_graph()
 read_graph()
 deps()
 knitr_deps
