+++ man/render_drake_graph.Rd
@@ -64,7 +64,8 @@ load_basic_example() # Load the canonical example for drake.
 # of nodes, edges, and legend nodes.
 graph <- dataframes_graph(my_plan)
 # You can pass the data frames right to render_drake_graph()
-# (as in vis_drake_graph()) or you can create your own custom visNewtork graph.
-render_drake_graph(graph, width = '100\%') # The width is passed to visNetwork.
+# (as in vis_drake_graph()) or you can create
+# your own custom visNewtork graph.
+render_drake_graph(graph, width = '100\%') # Width is passed to visNetwork.
 }
 }
