### 2017-11-13 31a56c3
Deprecate config() and read_config() in favor of better names
