+++ tests/testthat/test-deprecate.R
@@ -4,12 +4,14 @@ test_with_dir("deprecation: future", {
   expect_warning(backend())
 })
 
-test_with_dir("deprecation: make()", {
+test_with_dir("deprecation: make() and config()", {
   expect_warning(default_system2_args(jobs = 1, verbose = FALSE))
   expect_warning(make(workplan(x = 1), return_config = TRUE,
     verbose = FALSE))
   expect_warning(make(workplan(x = 1), clear_progress = TRUE,
     verbose = FALSE))
+  expect_warning(read_config())
+  expect_warning(config(workplan(x = 1)))
 })
 
 test_with_dir("deprecation: cache functions", {
