+++ R/graph.R
@@ -128,7 +128,7 @@ build_drake_graph <- function(
 #' \dontrun{
 #' load_basic_example() # Load the canonical example.
 #' # Build the igraph object representing the workflow dependency network.
-#' # You could also use config(my_plan)$graph
+#' # You could also use drake_config(my_plan)$graph
 #' graph <- build_drake_graph(my_plan)
 #' # The default plotting is not the greatest,
 #' # but you will get the idea.

[…]
