### 2017-11-13 92d4c31
Add new man pages
