+++ man/read_drake_config.Rd
@@ -1,0 +1,45 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/read.R
+\name{read_drake_config}
+\alias{read_drake_config}
+\title{Function \code{read_drake_config}}
+\usage{
+read_drake_config(path = getwd(), search = TRUE,
+  cache = drake::get_cache(path = path, search = search, verbose = verbose),
+  verbose = TRUE)
+}
+\arguments{
+\item{path}{Root directory of the drake project,
+or if \code{search} is \code{TRUE}, either the
+project root or a subdirectory of the project.}
+
+\item{search}{logical. If \code{TRUE}, search parent directories
+to find the nearest drake cache. Otherwise, look in the
+current working directory only.}
+
+\item{cache}{optional drake cache. See code{\link{new_cache}()}.
+If \code{cache} is supplied,
+the \code{path} and \code{search} arguments are ignored.}
+
+\item{verbose}{whether to print console messages}
+}
+\value{
+The cached master internal configuration list
+of the last \code{\link{make}()}.
+}
+\description{
+Read all the configuration parameters
+from your last attempted call to \code{\link{make}()}.
+These include the workflow plan
+}
+\examples{
+\dontrun{
+load_basic_example() # Load drake's canonical example.
+make(my_plan) # Run the project, build the targets.
+# Retrieve the master internal configuration list from the cache.
+read_drake_config()
+}
+}
+\seealso{
+\code{\link{make}}
+}
