### 2017-11-13 0916890
Deprecate example(s)_drake() in favor of drake_example(s)()
