+++ R/make.R
@@ -61,7 +61,7 @@
 #' To list the options, call \code{\link{parallelism_choices}()}.
 #' For detailed explanations, see \code{?\link{parallelism_choices}},
 #' the tutorial vignettes, or the tutorial files generated by
-#' \code{\link{example_drake}("basic")}
+#' \code{\link{drake_example}("basic")}
 #'
 #' @param jobs number of parallel processes or jobs to run.
 #' For \code{"future_lapply"} parallelism, \code{jobs}
@@ -316,7 +316,8 @@ make <- function(
 #' @examples
 #' \dontrun{
 #' load_basic_example() # Load the canonical example.
-#' con <- drake_config(my_plan) # Generate the master internal configuration list.
+#' # Generate the master internal configuration list.
+#' con <- drake_config(my_plan)
 #' # Just cache the imports, do not build any targets.
 #' make_imports(config = con)
 #' }
