+++ tests/testthat/test-arbitrary-cache.R
@@ -130,8 +130,9 @@ test_with_dir("arbitrary storr in-memory cache", {
   expect_true(is.list(read_drake_config(cache = cache, verbose = FALSE)))
   expect_false(file.exists(default_cache_path()))
 
-  expect_error(read_graph(verbose = FALSE))
-  expect_equal(class(read_graph(cache = cache, verbose = FALSE)), "igraph")
+  expect_error(read_drake_graph(verbose = FALSE))
+  expect_equal(class(
+    read_drake_graph(cache = cache, verbose = FALSE)), "igraph")
   expect_false(file.exists(default_cache_path()))
 
   expect_error(read_plan(verbose = FALSE))
