+++ vignettes/drake.Rmd
@@ -124,7 +124,7 @@ missed()
 vis_drake_graph() # Now with subgraphs too.
 dataframes_graph()
 render_drake_graph()
-read_graph()
+read_drake_graph()
 deps()
 knitr_deps
 tracked()
