+++ man/read_graph.Rd
@@ -1,49 +1,37 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/read.R
+% Please edit documentation in R/deprecate.R
 \name{read_graph}
 \alias{read_graph}
-\title{Function \code{read_graph}}
+\title{Deprecated function \code{read_graph}}
 \usage{
 read_graph(path = getwd(), search = TRUE, cache = drake::get_cache(path =
   path, search = search, verbose = verbose), verbose = TRUE, ...)
 }
 \arguments{
-\item{path}{Root directory of the drake project,
-or if \code{search} is \code{TRUE}, either the
-project root or a subdirectory of the project.}
+\item{path}{Same as for \code{\link{read_drake_graph}()}.}
 
-\item{search}{logical. If \code{TRUE}, search parent directories
-to find the nearest drake cache. Otherwise, look in the
-current working directory only.}
+\item{search}{Same as for \code{\link{read_drake_graph}()}.}
 
-\item{cache}{optional drake cache. See code{\link{new_cache}()}.
-If \code{cache} is supplied,
-the \code{path} and \code{search} arguments are ignored.}
+\item{cache}{Same as for \code{\link{read_drake_graph}()}.}
 
-\item{verbose}{logical, whether to print console messages}
+\item{verbose}{Same as for \code{\link{read_drake_graph}()}.}
 
-\item{...}{arguments to \code{visNetwork()} via
-\code{\link{vis_drake_graph}()}}
+\item{...}{Same as for \code{\link{read_drake_graph}()}.}
 }
 \value{
 An \code{igraph} object representing the dependency
 network of the workflow.
 }
 \description{
-Read the igraph-style dependency graph of your targets
-from your last attempted call to \code{\link{make}()}.
-For better graphing utilities, see \code{\link{vis_drake_graph}()}
-and related functions.
+Use \code{\link{read_drake_graph}()} instead.
+}
+\details{
+Deprecated on 2017-11-12.
 }
 \examples{
-\dontrun{
-load_basic_example() # Load the canonical example.
-make(my_plan) # Run the project, build the targets.
-# Retrieve the igraph network from the cache.
-g <- read_graph()
-class(g) # "igraph"
-}
+# See ?read_drake_graph for the examples.
 }
 \seealso{
-\code{\link{vis_drake_graph}}, \code{\link{read_config}}
+\code{\link{read_drake_graph}}.
 }
+\keyword{internal}
