+++ NAMESPACE
@@ -88,6 +88,7 @@ export(r_recipe_wildcard)
 export(rate_limiting_times)
 export(read_config)
 export(read_drake_config)
+export(read_drake_graph)
 export(read_graph)
 export(read_plan)
 export(readd)
