### 2017-11-13 be9913b
Improve the solution to #144
