+++ man/build_drake_graph.Rd
@@ -6,7 +6,7 @@
 \usage{
 build_drake_graph(plan = workplan(),
   targets = drake::possible_targets(plan), envir = parent.frame(),
-  verbose = TRUE, jobs = 1, skip_imports = FALSE)
+  verbose = TRUE, jobs = 1)
 }
 \arguments{
 \item{plan}{workflow plan data frame, same as for function
@@ -23,11 +23,6 @@ build_drake_graph(plan = workplan(),
 \item{jobs}{number of jobs to accelerate the construction
 of the dependency graph. A light \code{mclapply}-based
 parallelism is used if your operating system is not Windows.}
-
-\item{skip_imports}{logical, whether to totally neglect to
-process the imports and jump straight to the targets. This can be useful
-if your imports are massive and you just want to test your project,
-but it is bad practice for reproducible data analysis.}
 }
 \value{
 An igraph object representing
