### 2017-11-13 29a299f
Fix #118: clean(garbage_collection = TRUE)
