+++ R/clean.R
@@ -1,6 +1,11 @@
 #' @title Function \code{clean}
 #' @description Cleans up all work done by \code{\link{make}()}.
-#' @details You must be in your project's working directory
+#' @details
+#' By default, \code{clean()} removes references to cached data.
+#' To deep-clean the data to free up storage/memory, use
+#' \code{clean(garbage_collection = TRUE)}. Garbage collection is slower,
+#' but it purges data with no remaining references.
+#' Also, for \code{clean()}, you must be in your project's working directory
 #' or a subdirectory of it.
 #' \code{clean(search = TRUE)} searches upwards in your folder structure
 #' for the drake cache and acts on the first one it sees. Use
@@ -52,6 +57,14 @@
 #' even though the project may not be back compatible with the
 #' current version of drake.
 #'
+#' @param garbage_collection logical, whether to call
+#' \code{cache$gc()} to do garbage collection.
+#' If \code{TRUE}, cached data with no remaining references
+#' will be removed.
+#' This will slow down \code{clean()}, but the cache
+#' could take up far less space afterwards.
+#' See the \code{gc()} method for \code{storr} caches.
+#'
 #' @examples
 #' \dontrun{
 #' load_basic_example() # Load drake's canonical example.
@@ -68,6 +81,12 @@
 #' # Remove all the targets and imports.
 #' # On non-Windows machines, parallelize over at most 2 jobs.
 #' clean(jobs = 2)
+#' # Make the targets again.
+#' make(my_plan)
+#' # Garbage collection removes data whose references are no longer present.
+#' # It is slow, but you should enable it if you want to reduce the
+#' # size of the cache.
+#' clean(garbage_collection = TRUE)
 #' # All the targets and imports are gone.
 #' cached()
 #' # Completely remove the entire cache (default: '.drake/' folder).
@@ -82,7 +101,8 @@ clean <- function(
   cache = NULL,
   verbose = TRUE,
   jobs = 1,
-  force = FALSE
+  force = FALSE,
+  garbage_collection = FALSE
 ){
   dots <- match.call(expand.dots = FALSE)$...
   targets <- targets_from_dots(dots, list)
@@ -101,6 +121,9 @@ clean <- function(
     )
   } else {
     uncache(targets = targets, cache = cache, jobs = jobs)
+  }
+  if (garbage_collection){
+    cache$gc()
   }
   invisible()
 }
