### 2017-11-13 0d5a48f
Add drake_gc() for garbage collection
