+++ NAMESPACE
@@ -41,6 +41,7 @@ export(drake_build)
 export(drake_config)
 export(drake_example)
 export(drake_examples)
+export(drake_gc)
 export(drake_palette)
 export(drake_session)
 export(drake_tip)
