### 2017-11-13 07d871e
Implement drake_gc(): garbage collection without clean()
