### 2017-11-13 6494ac7
Test that drake_gc() does not uncache targets/imports
