### 2017-11-13 34d3655
Add comment to drake_gc() test
