### 2017-11-13 a03c32a
Address #155
- rescue_del() invisibly returns NULL (hopefully addresses the memory blow-up).
- rescue_cache() can limit itself to specific targets
