+++ man/rescue_cache.Rd
@@ -4,11 +4,14 @@
 \alias{rescue_cache}
 \title{Function rescue_cache}
 \usage{
-rescue_cache(path = getwd(), search = TRUE, verbose = TRUE,
-  force = FALSE, cache = drake::get_cache(path = path, search = search,
-  verbose = verbose, force = force), jobs = 1)
+rescue_cache(targets = NULL, path = getwd(), search = TRUE,
+  verbose = TRUE, force = FALSE, cache = drake::get_cache(path = path,
+  search = search, verbose = verbose, force = force), jobs = 1)
 }
 \arguments{
+\item{targets}{Character vector, names of the targets to rescue.
+Defaults to everything so that the whole cache is rescued.}
+
 \item{path}{same as for \code{\link{get_cache}()}}
 
 \item{search}{same as for \code{\link{get_cache}()}}
@@ -37,6 +40,9 @@ load_basic_example() # Load the canonical example.
 make(my_plan) # Run the project, build targets. This creates the cache.
 # Remove dangling cache files that could cause errors.
 rescue_cache(jobs = 2)
+# Alternatively, just rescue targets 'small' and 'large'.
+# Rescuing specific targets is usually faster.
+rescue_cache(targets = c("small", "large"))
 }
 }
 \seealso{
