### 2017-11-13 ab89944
Change backend() to future::plan() in comments
