### 2017-11-13 7e8600d
Remove more mentions of backend()
