+++ R/make.R
@@ -1,7 +1,7 @@
 #' @title Function \code{make}
 #' @description Run your project (build the targets).
 #' @seealso \code{\link{workplan}}, \code{\link{workplan}},
-#' \code{\link{backend}}, \code{\link{vis_drake_graph}},
+#' \code{\link{vis_drake_graph}},
 #' \code{\link{max_useful_jobs}}, \code{\link{shell_file}},
 #' \code{\link{default_hook}}, \code{\link{silencer_hook}},
 #' \code{\link{triggers}}
