### 2017-11-13 9f121be
Add to documentation of rescue_cache()
