### 2017-11-14 93783be
Remove dependency eply
Plus talk about the 'workers' argument
to future::multisession, future.batchtools::batchtools_local, etc.
