### 2017-11-14 7c643c8
Suppress spurious goodpractice::gp() messages
In the test scripts, there are legitimate needs
for setwd() and detach(), and the usual risks are moot.
