### 2017-11-14 fa981e6
Remove spurious goodpractice messages
