### 2017-11-14 968ee88
Align comments in docs
