+++ vignettes/drake.Rmd
@@ -75,15 +75,15 @@ install_github("wlandau-lilly/drake", build = TRUE) # dev version
 
 ```{r quickstartdrakermd, eval = FALSE}
 library(drake)
-load_basic_example() # Also (over)writes report.Rmd.
-vis_drake_graph(my_plan)  # Hover, click, drag, zoom, pan. See args 'from' and 'to'.
-outdated(my_plan)    # Which targets need to be (re)built?
-missed(my_plan)      # Are you missing anything from your workspace?
-check_plan(my_plan)  # Are you missing files? Is your workflow plan okay?
-make(my_plan)        # Run the workflow.
-diagnose(large)      # View error info if the target "large" failed to build.
-outdated(my_plan)    # Everything is up to date.
-vis_drake_graph(my_plan)  # The graph also shows what is up to date.
+load_basic_example()     # Also (over)writes report.Rmd.
+vis_drake_graph(my_plan) # Hover, click, drag, zoom, pan. See args 'from' and 'to'.
+outdated(my_plan)        # Which targets need to be (re)built?
+missed(my_plan)          # Are you missing anything from your workspace?
+check_plan(my_plan)      # Are you missing files? Is your workflow plan okay?
+make(my_plan)            # Run the workflow.
+diagnose(large)          # View error info if the target "large" failed to build.
+outdated(my_plan)        # Everything is up to date.
+vis_drake_graph(my_plan) # The graph also shows what is up to date.
 ```
 
 Dive deeper into the built-in examples.
@@ -91,7 +91,7 @@ Dive deeper into the built-in examples.
 ```{r examplesdrakermd, eval = FALSE}
 drake_example("basic") # Write the code files of the canonical tutorial.
 drake_examples()       # List the other examples.
-vignette("quickstart") # See https://cran.r-project.org/package=drake/vignettes
+vignette("quickstart") # https://cran.r-project.org/package=drake/vignettes
 ```
 
 # Useful functions
@@ -198,18 +198,17 @@ There is room to improve the conversation and the landscape of reproducibility i
 ```{r reproducibilitydrakermd, eval = FALSE}
 library(drake)
 load_basic_example()
-outdated(my_plan) # Which targets need to be (re)built?
-make(my_plan)     # Build what needs to be built.
-outdated(my_plan) # Everything is up to date.
-# Change one of your functions.
-reg2 <- function(d) {
+outdated(my_plan)        # Which targets need to be (re)built?
+make(my_plan)            # Build what needs to be built.
+outdated(my_plan)        # Everything is up to date.
+reg2 <- function(d){     # Change one of your functions.
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
-outdated(my_plan)   # Some targets depend on reg2().
+outdated(my_plan)        # Some targets depend on reg2().
 vis_drake_graph(my_plan) # Set targets_only to TRUE for smaller graphs.
-make(my_plan)       # Rebuild just the outdated targets.
-outdated(my_plan)   # Everything is up to date again.
+make(my_plan)            # Rebuild just the outdated targets.
+outdated(my_plan)        # Everything is up to date again.
 vis_drake_graph(my_plan) # The colors changed in the graph.
 ```
 
