### 2017-11-14 b356072
Streamline parallel_stage() a little.
