+++ R/parallel.R
@@ -15,11 +15,9 @@ run_parallel <- function(config, worker) {
 }
 
 parallel_stage <- function(worker, config) {
-  config <- quick_inventory(config)
-  remaining_targets <- V(config$graph_remaining_targets) %>%
-    names %>% intersect(config$targets)
   candidates <- next_targets(
     config$graph_remaining_targets, jobs = config$jobs)
+  config <- quick_inventory(config)
   meta_list <- meta_list(targets = candidates, config = config)
   build_these <- Filter(candidates,
     f = function(target)
