### 2017-11-15 b36bcc0
No more inventories!
Using cache$exists() instead. Should be a whole lot faster.
