+++ R/dependencies.R
@@ -85,7 +85,6 @@ deps <- function(x){
 #'   namespace = "kernels")
 #' }
 dependency_profile <- function(target, config){
-  config <- thorough_inventory(config)
   config$plan$trigger <- NULL
   cached_command <- safe_get(key = target, namespace = "commands",
     config = config)
