+++ R/meta.R
@@ -11,11 +11,11 @@ meta_list <- function(targets, config) {
   out
 }
 
-meta <- function(target,  config) {
+meta <- function(target, config) {
   meta <- list(
     target = target,
     imported = !(target %in% config$plan$target),
-    missing = target %in% config$inventory$kernels
+    missing = target_exists(target = target, config = config)
   )
   trigger <- get_trigger(target = target, config = config)
   # Need to make sure meta includes all these
@@ -52,7 +52,7 @@ dependency_hash <- function(target, config) {
 }
 
 self_hash <- Vectorize(function(target, config) {
-  if (target %in% config$inventory$kernels) {
+  if (target_exists(target = target, config = config)) {
     config$cache$get_hash(target, namespace = "kernels")
   } else {
     as.character(NA)
@@ -86,9 +86,11 @@ file_hash <- function(target, config, size_cutoff = 1e5) {
   }
   if (!file.exists(filename))
     return(as.character(NA))
-  old_mtime <- ifelse(target %in% config$inventory$mtimes,
+  old_mtime <- ifelse(
+    file_target_exists(target = target, config = config),
     config$cache$get(key = target, namespace = "mtimes"),
-    -Inf)
+    -Inf
+  )
   new_mtime <- file.mtime(filename)
   do_rehash <- should_rehash_file(
     filename = filename,
