### 2017-11-15 7586876
Add a skip_safety_checks option to make() and drake_config()
