### 2017-11-15 a28cfeb
Update NEWS.md to reflect recent speed improvements.
