### 2017-11-15 5169bcd
Add make_from_config()
