+++ NEWS.md
@@ -53,7 +53,7 @@
 - Tell the user the location of the cache using a console message. Happens on every call to `get_cache(..., verbose = TRUE)`.
 - Increase efficiency of internal preprocessing via `lightly_parallelize()` and `lightly_parallelize_atomic()`. Now, processing happens faster, and only over the unique values of a vector.
 - Add a new `storr` namespace called `imports` to be used in `is_imported()`. That way, the whole object need not be read to `clean()` is. `clean()` is much faster and safer.
-
+- Add a new `make_from_config()` function to do the work of `make()` on an existing internal configuration list from `drake_config()`.
 
 # 2017-10-17
 
