+++ NAMESPACE
@@ -70,6 +70,7 @@ export(load_basic_example)
 export(loadd)
 export(long_hash)
 export(make)
+export(make_from_config)
 export(make_imports)
 export(max_useful_jobs)
 export(message_sink_hook)
