+++ man/make_from_config.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/make.R
+\name{make_from_config}
+\alias{make_from_config}
+\title{Function make_from_config}
+\usage{
+make_from_config(config)
+}
+\arguments{
+\item{config}{An input internal configuration list}
+}
+\value{
+An output internal configuration list
+}
+\description{
+Run \code{\link{make}()},
+on an existing internal configuration list
+that you can get from \code{\link{drake_config}()}.
+}
+\examples{
+\dontrun{
+load_basic_example() # Load the canonical example
+# The following lines are the same as make(my_plan)
+config <- drake_config(my_plan) # Create the internal config list.
+make_from_config(config = config) # Run the project, build the targets.
+}
+}
+\seealso{
+\code{\link{make}}, \code{\link{drake_config}}
+}
