### 2017-11-15 46e8b73
Change make_from_config() to make_with_config()
