+++ man/make_with_config.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/make.R
-\name{make_from_config}
-\alias{make_from_config}
-\title{Function make_from_config}
+\name{make_with_config}
+\alias{make_with_config}
+\title{Function make_with_config}
 \usage{
-make_from_config(config)
+make_with_config(config)
 }
 \arguments{
 \item{config}{An input internal configuration list}
@@ -22,7 +22,7 @@ that you can get from \code{\link{drake_config}()}.
 load_basic_example() # Load the canonical example
 # The following lines are the same as make(my_plan)
 config <- drake_config(my_plan) # Create the internal config list.
-make_from_config(config = config) # Run the project, build the targets.
+make_with_config(config = config) # Run the project, build the targets.
 }
 }
 \seealso{
