+++ NAMESPACE
@@ -70,8 +70,8 @@ export(load_basic_example)
 export(loadd)
 export(long_hash)
 export(make)
-export(make_from_config)
 export(make_imports)
+export(make_with_config)
 export(max_useful_jobs)
 export(message_sink_hook)
 export(migrate_drake_project)
