### 2017-11-15 6ced811
Add a 'cache' argument to drake_gc()
