### 2017-11-15 4e62fe4
Try to fix memory issue in rescue_cache()
