+++ R/clean.R
@@ -129,58 +129,6 @@ clean <- function(
   invisible()
 }
 
-
-#' @title Function drake_gc()
-#' @description Do garbage collection on the cache
-#' @seealso \code{\link{clean}}
-#' @export
-#' @return\code{NULL}
-#' @param path file path to the folder containing the cache.
-#' Yes, this is the parent directory containing the cache,
-#' not the cache itself, and it assumes the cache is in the
-#' `.drake` folder. If you are looking for a different cache
-#' with a known folder different from `.drake`, use
-#' the \code{\link{this_cache}()} function.
-#' @param search logical, whether to search back in the file system
-#' for the cache.
-#' @param verbose logical, whether to print the location of the cache
-#' @param cache the \code{drake}/\code{storr} cache object itself,
-#' if available.
-#' @param force logical, whether to load the cache
-#' despite any back compatibility issues with the
-#' running version of drake.
-#' @examples
-#' \dontrun{
-#' load_basic_example() # Load drake's canonical example.
-#' make(my_plan) # Run the project, build the targets.
-#' # At this point, check the size of the '.drake/' cache folder.
-#' # Clean without garbage collection.
-#' clean(garbage_collection = FALSE)
-#' # The '.drake/' cache folder is still about the same size.
-#' drake_gc() # Do garbage collection on the cache.
-#' # The '.drake/' cache folder should have gotten much smaller.
-#' }
-drake_gc <- function(
-  path = getwd(),
-  search = TRUE,
-  verbose = TRUE,
-  cache = NULL,
-  force = FALSE
-){
-  if (is.null(cache)){
-    cache <- get_cache(
-      path = path,
-      search = search,
-      verbose = verbose,
-      force = force
-    )
-  }
-  if (!is.null(cache)){
-    cache$gc()
-  }
-  invisible()
-}
-
 clean_everything <- function(
   destroy,
   cache,
@@ -239,4 +187,129 @@ uncache_single <- function(target, cache, namespace, list){
   if (target %in% list){
     cache$del(target, namespace = namespace)
   }
+}
+
+#' @title Function drake_gc()
+#' @description Do garbage collection on the cache
+#' @seealso \code{\link{clean}}
+#' @export
+#' @return\code{NULL}
+#' @param path file path to the folder containing the cache.
+#' Yes, this is the parent directory containing the cache,
+#' not the cache itself, and it assumes the cache is in the
+#' `.drake` folder. If you are looking for a different cache
+#' with a known folder different from `.drake`, use
+#' the \code{\link{this_cache}()} function.
+#' @param search logical, whether to search back in the file system
+#' for the cache.
+#' @param verbose logical, whether to print the location of the cache
+#' @param cache the \code{drake}/\code{storr} cache object itself,
+#' if available.
+#' @param force logical, whether to load the cache
+#' despite any back compatibility issues with the
+#' running version of drake.
+#' @examples
+#' \dontrun{
+#' load_basic_example() # Load drake's canonical example.
+#' make(my_plan) # Run the project, build the targets.
+#' # At this point, check the size of the '.drake/' cache folder.
+#' # Clean without garbage collection.
+#' clean(garbage_collection = FALSE)
+#' # The '.drake/' cache folder is still about the same size.
+#' drake_gc() # Do garbage collection on the cache.
+#' # The '.drake/' cache folder should have gotten much smaller.
+#' }
+drake_gc <- function(
+  path = getwd(),
+  search = TRUE,
+  verbose = TRUE,
+  cache = NULL,
+  force = FALSE
+){
+  if (is.null(cache)){
+    cache <- get_cache(
[…]
