+++ R/cache_interaction.R
@@ -207,73 +207,6 @@ is_built_or_imported_file <- Vectorize(function(target, cache) {
 },
 "target", SIMPLIFY = TRUE)
 
-#' @title Function rescue_cache
-#' @description Sometimes, \code{storr} caches may have
-#' dangling orphaned files that prevent you from loading or cleaning.
-#' This function tries to remove those files so you can use the
-#' cache normally again.
-#' @return The rescued drake/storr cache.
-#' @export
-#' @seealso \code{\link{get_cache}}, \code{\link{cached}}
-#' @param targets Character vector, names of the targets to rescue.
-#' As with many other drake utility functions, the word \code{target}
-#' is defined generally in this case, encompassing imports
-#' as well as true targets.
-#' If \code{targets} is \code{NULL}, everything in the
-#' cache is rescued.
-#' @param path same as for \code{\link{get_cache}()}
-#' @param search same as for \code{\link{get_cache}()}
-#' @param verbose same as for \code{\link{get_cache}()}
-#' @param force same as for \code{\link{get_cache}()}
-#' @param cache a `storr` cache object
-#' @param jobs number of jobs for light parallelism
-#' (disabled on Windows)
-#' @examples
-#' \dontrun{
-#' load_basic_example() # Load the canonical example.
-#' make(my_plan) # Run the project, build targets. This creates the cache.
-#' # Remove dangling cache files that could cause errors.
-#' rescue_cache(jobs = 2)
-#' # Alternatively, just rescue targets 'small' and 'large'.
-#' # Rescuing specific targets is usually faster.
-#' rescue_cache(targets = c("small", "large"))
-#' }
-rescue_cache <- function(
-  targets = NULL,
-  path = getwd(), search = TRUE, verbose = TRUE, force = FALSE,
-  cache = drake::get_cache(
-    path = path, search = search, verbose = verbose, force = force),
-  jobs = 1
-){
-  if (is.null(cache)){
-    return(invisible())
-  }
-  for (namespace in cache$list_namespaces()){
-    X <- cache$list(namespace = namespace)
-    if (!is.null(targets)){
-      X <- intersect(X, targets)
-    }
-    lightly_parallelize(
-      X = X,
-      FUN = rescue_del,
-      jobs = jobs,
-      cache = cache,
-      namespace = namespace
-    )
-  }
-  invisible(cache)
-}
-
-rescue_del <- function(key, cache, namespace){
-  tryCatch(
-    cache$get(key = key, namespace = namespace),
-    error = function(e){
-      cache$del(key = key, namespace = namespace)
-    }
-  )
-  invisible()
-}
-
 file_target_exists <- function(target, config){
   config$cache$exists(key = target, namespace = "mtimes")
 }
