+++ man/rescue_cache.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cache_interaction.R
+% Please edit documentation in R/clean.R
 \name{rescue_cache}
 \alias{rescue_cache}
 \title{Function rescue_cache}
