+++ R/clean.R
@@ -309,7 +309,9 @@ rescue_del <- function(key, cache, namespace){
 }
 
 touch_storr_object <- function(key, cache, namespace){
+  envir <- environment()
   hash <- cache$get_hash(key = key, namespace = namespace)
   value <- cache$driver$get_object(hash = hash)
+  remove(value, envir = envir)
   invisible(NULL)
 }
