### 2017-11-15 88eb841
Make garbage collection optional in rescue_cache()
