+++ man/rescue_cache.Rd
@@ -6,7 +6,8 @@
 \usage{
 rescue_cache(targets = NULL, path = getwd(), search = TRUE,
   verbose = TRUE, force = FALSE, cache = drake::get_cache(path = path,
-  search = search, verbose = verbose, force = force), jobs = 1)
+  search = search, verbose = verbose, force = force), jobs = 1,
+  garbage_collection = TRUE)
 }
 \arguments{
 \item{targets}{Character vector, names of the targets to rescue.
@@ -28,6 +29,9 @@ cache is rescued.}
 
 \item{jobs}{number of jobs for light parallelism
 (disabled on Windows)}
+
+\item{garbage_collection}{logical, whether to do garbage collection
+as a final step. See \code{\link{drake_gc}}}
 }
 \value{
 The rescued drake/storr cache.
@@ -50,5 +54,6 @@ rescue_cache(targets = c("small", "large"))
 }
 }
 \seealso{
-\code{\link{get_cache}}, \code{\link{cached}}
+\code{\link{get_cache}}, \code{\link{cached}},
+\code{\link{drake_gc}}
 }
