### 2017-11-15 4b67bbb
Turn of garbage collection in rescue_cache() by default
