### 2017-11-15 31c5dfa
Fix code coverage
