### 2017-11-16 651088f
Remove superfluous namespaces
