+++ R/parallel.R
@@ -21,8 +21,8 @@ parallel_stage <- function(worker, config) {
   build_these <- Filter(candidates,
     f = function(target)
       should_build(target = target, meta_list = meta_list, config = config))
-  attempts <- intersect(build_these, config$plan$target)
-  log_attempts(targets = attempts, config = config)
+  intersect(build_these, config$plan$target) %>%
+    increment_attempt_flag(config = config)
   meta_list <- meta_list[build_these]
   if (length(build_these)){
     worker(targets = build_these, meta_list = meta_list,
