+++ R/Makefile.R
@@ -3,7 +3,7 @@ run_Makefile <- function( #nolint: we want Makefile capitalized.
   run = TRUE,
   debug = FALSE
 ){
-  prepare_distributed(config = config)
+  build_these <- prepare_distributed(config = config)
   with_output_sink(
     new = "Makefile",
     code = {
@@ -11,7 +11,7 @@ run_Makefile <- function( #nolint: we want Makefile capitalized.
       makefile_rules(config)
     }
   )
-  time_stamps(config = config)
+  time_stamps(build_these = build_these, config = config)
   error_code <- ifelse(
     run,
     system2(command = config$command, args = config$args),
@@ -102,10 +102,10 @@ build_recipe <- function(target, recipe_command,
 #' con <- drake_config(my_plan) # Construct the internal configuration list.
 #' # Prepare to use a distributed computing parallel backend
 #' # such as "Makefile" or "future_lapply".
-#' drake:::prepare_distributed(config = con)
+#' build_these <- drake:::prepare_distributed(config = con)
 #' # Write the dummy timestamp files usually written at the beginning
 #' # of make(..., parallelism = "Makefile").
-#' time_stamps(config = con)
+#' time_stamps(build_these = build_these, config = con)
 #' # Use mk() to build a target. Usually called inside a Makefile recipe.
 #' mk(target = "small", cache_path = default_cache_path())
 #' }
