+++ R/distributed.R
@@ -8,9 +8,9 @@ prepare_distributed <- function(config){
     )
   }
   config$cache$set(key = "envir", value = config$envir, namespace = "config")
-  attempts <- outdated(config = config, make_imports = !config$skip_imports)
-  log_attempts(targets = attempts, config = config)
-  invisible()
+  build_these <- outdated(config = config, make_imports = !config$skip_imports)
+  increment_attempt_flag(targets = build_these, config = config)
+  invisible(build_these)
 }
 
 finish_distributed <- function(config){
