### 2017-11-16 92dbadd
Eliminate more superfluous namespaces
