+++ tests/testthat/test-full-build.R
@@ -11,12 +11,12 @@ test_with_dir("scratch build with custom filesystem cache.", {
   )
   expect_error(drake_session(cache = cache))
   expect_true(length(progress(cache = cache)) == 0)
-  expect_equal(config$cache$list(namespace = "readd"), character(0))
+  expect_equal(config$cache$list(), character(0))
 
   testrun(config)
 
   expect_true(is.numeric(readd(final, cache = cache)))
-  expect_true(length(config$cache$list(namespace = "readd")) > 2)
+  expect_true(length(config$cache$list()) > 2)
   expect_false(any(c("f", "final") %in% ls()))
   expect_true(is.list(drake_session(cache = cache)))
   expect_true(all(drake_session(cache = cache)$target %in% config$plan$target))
@@ -36,7 +36,7 @@ test_with_dir("scratch build with custom filesystem cache.", {
     "'intermediatefile.rds'", "a",
     "b", "c", "combined", "f", "final", "g", "h", "i", "j",
     "myinput", "nextone", "readRDS", "saveRDS", "yourinput"))
-  expect_equal(config$cache$list(namespace = "readd"), all)
+  expect_equal(config$cache$list(), all)
   expect_true(file.exists("intermediatefile.rds"))
   expect_true(file.exists("input.rds"))
   expect_false(file.exists(default_cache_path()))
@@ -47,31 +47,31 @@ test_with_dir("scratch build with custom filesystem cache.", {
     cache = cache)
   expect_false(file.exists("intermediatefile.rds"))
   expect_true(file.exists("input.rds"))
-  expect_equal(config$cache$list(namespace = "readd"),
+  expect_equal(config$cache$list(),
     setdiff(all, c("b", "c", "'intermediatefile.rds'", "nextone")))
 
   # clean does not remove imported files
   expect_true(file.exists("input.rds"))
   expect_true("'input.rds'" %in%
-    config$cache$list(namespace = "readd"))
+    config$cache$list())
   clean("'input.rds'", cache = cache)
   expect_true(file.exists("input.rds"))
   expect_false("'input.rds'" %in%
-    config$cache$list(namespace = "readd"))
+    config$cache$list())
 
   # clean removes imported functions and cleans up 'functions'
   # namespace
   expect_true(cached(f, cache = cache))
-  for (n in c("depends", "readd", "kernels")) {
+  for (n in c(cache$default_namespace, "depends", "kernels")) {
     expect_true("f" %in% config$cache$list(namespace = n))
   }
   clean(f, cache = cache)
-  for (n in c("depends", "readd", "kernels")) {
+  for (n in c(cache$default_namespace, "depends", "kernels")) {
     expect_false("f" %in% config$cache$list(namespace = n))
   }
 
   clean(destroy = FALSE, cache = cache)
-  expect_equal(config$cache$list("readd"), character(0))
+  expect_equal(config$cache$list(), character(0))
   expect_equal(config$cache$list("depends"), character(0))
   expect_equal(config$cache$list("reporducibly_tracked"), character(0))
   expect_false(file.exists("intermediatefile.rds"))
@@ -87,11 +87,11 @@ test_with_dir("scratch build with custom filesystem cache.", {
 test_with_dir("clean in full build.", {
   config <- dbug()
   make(config$plan, envir = config$envir, verbose = FALSE)
-  expect_true("final" %in% config$cache$list(namespace = "readd"))
+  expect_true("final" %in% config$cache$list())
   clean(final, search = TRUE)
-  expect_false("final" %in% config$cache$list(namespace = "readd"))
+  expect_false("final" %in% config$cache$list())
   clean(search = TRUE)
[…]
