### 2017-11-16 8250f55
Fix #69
Thanks to [Jesse Maegan](https://community.rstudio.com/u/jessemaegan)'s advice on an [RStudio Community forum](https://community.rstudio.com/t/designing-hex-logos/2756/19), I completely redid the logo! There will probably be some touching up to do, but probably nothing major. Once #147 is solved, I will definitely need to change the GitHub link at the bottom.

Also, I cleaned up the vignettes a little: turned down verbosity so cache() did not spout out long file paths so often.
