+++ vignettes/caution.Rmd
@@ -357,7 +357,7 @@ Unfortunately, there is no one-size-fits-all `*.tmpl` configuration file for any
 The [Makefile](http://kbroman.org/minimal_make/) generated by `make(myplan, parallelism = "Makefile")` is not standalone. Do not run it outside of `drake::make()`. Drake uses dummy timestamp files to tell the [Makefile](http://kbroman.org/minimal_make/) what to do, and running `make` in the [terminal](https://www.howtogeek.com/140679/beginner-geek-how-to-start-using-the-linux-terminal/) will most likely give incorrect results.
 
 ```{r clean, echo = FALSE}
-clean(destroy = TRUE)
+clean(destroy = TRUE, verbose = FALSE)
 unlink(c("report.Rmd", "Thumbs.db"))
 ```
 
