+++ vignettes/drake.Rmd
@@ -17,7 +17,7 @@ vignette: >
 
 ```{r, echo = F}
 suppressMessages(suppressWarnings(library(drake)))
-clean(destroy = TRUE)
+clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
 ```
 
@@ -76,12 +76,12 @@ install_github("wlandau-lilly/drake", build = TRUE) # dev version
 ```{r quickstartdrakermd, eval = FALSE}
 library(drake)
 load_basic_example()     # Also (over)writes report.Rmd.
-vis_drake_graph(my_plan) # Hover, click, drag, zoom, pan. See args 'from' and 'to'.
+vis_drake_graph(my_plan) # Hover, click, drag, zoom, pan.
 outdated(my_plan)        # Which targets need to be (re)built?
 missed(my_plan)          # Are you missing anything from your workspace?
-check_plan(my_plan)      # Are you missing files? Is your workflow plan okay?
+check_plan(my_plan)      # Is your workflow plan okay? Missing files?
 make(my_plan)            # Run the workflow.
-diagnose(large)          # View error info if the target "large" failed to build.
+diagnose(large)          # Error info if the target "large" failed to build.
 outdated(my_plan)        # Everything is up to date.
 vis_drake_graph(my_plan) # The graph also shows what is up to date.
 ```
@@ -202,7 +202,8 @@ load_basic_example()
 outdated(my_plan)        # Which targets need to be (re)built?
 make(my_plan)            # Build what needs to be built.
 outdated(my_plan)        # Everything is up to date.
-reg2 <- function(d){     # Change one of your functions.
+# Change one of your functions.
+reg2 <- function(d){
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
@@ -310,6 +311,6 @@ Please refer to [TROUBLESHOOTING.md](https://github.com/wlandau-lilly/drake/blob
 Bug reports, suggestions, and code are welcome. Please see [.github/CONTRIBUTING.md](https://github.com/wlandau-lilly/drake/blob/master/.github/CONTRIBUTING.md). Maintainers and contributors must follow this repository's [code of conduct](https://github.com/wlandau-lilly/drake/blob/master/CONDUCT.md).
 
 ```{r rmfiles_main, echo = FALSE}
-clean(destroy = TRUE)
+clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
 ```
