+++ vignettes/quickstart.Rmd
@@ -17,7 +17,7 @@ vignette: >
 
 ```{r suppression, echo = F}
 suppressMessages(suppressWarnings(library(drake)))
-clean(destroy = TRUE)
+clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
 ```
 
@@ -28,7 +28,7 @@ Inspect and run your project.
 ```{r quickstartquickstart, eval = FALSE}
 library(drake)
 load_basic_example() # Also (over)writes report.Rmd.
-vis_drake_graph(my_plan) # Hover, click, drag, zoom, pan. See args 'from' and 'to'.
+vis_drake_graph(my_plan) # Hover, click, drag, zoom, pan.
 make(my_plan) # Run the workflow.
 outdated(my_plan) # Check that everything is already up to date.
 ```
@@ -377,10 +377,11 @@ make(my_plan)
 If you ever need to erase your work, use `clean()`. Any targets removed from the cache will have to be rebuilt on the next call to `make()`, so be careful. You may notice that by default, the size of the cache does not go down very much. To purge old data, you could use `clean(garbage_collection = TRUE)`. To do garbage collection without removing any important targets, use `drake_gc()`.
 
 ```{r cleanup}
-clean(small, reg1) # uncaches individual targets and imported objects
-clean() # cleans all targets out of the cache
-drake_gc() # Just garbage collection.
-clean(destroy = TRUE) # removes the cache entirely
+# Uncaches individual targets and imported objects.
+clean(small, reg1, verbose = FALSE)
+clean(verbose = FALSE) # Cleans all targets out of the cache.
+drake_gc(verbose = FALSE) # Just garbage collection.
+clean(destroy = TRUE, verbose = FALSE) # removes the cache entirely
 ```
 
 # Automatic watching for changed dependencies
