### 2017-11-17 a7bf954
Add batchtools_drake_tmpl_file()
