+++ tests/testthat/test-examples.R
@@ -10,6 +10,9 @@ test_with_dir("examples are listed and written", {
     expect_true(file.info(i)$isdir)
     unlink(i, recursive = TRUE, force = TRUE)
   }
+  expect_warning(drake_example(destination = getwd()))
+  expect_silent(batchtools_drake_tmpl_file("slurm"))
+  expect_error(batchtools_drake_tmpl_file("basic"))
 })
 
 test_with_dir("overwrites of report.Rmd handled correctly", {
