+++ NAMESPACE
@@ -7,6 +7,7 @@ export(as_drake_filename)
 export(as_file)
 export(available_hash_algos)
 export(backend)
+export(batchtools_drake_tmpl_file)
 export(build_drake_graph)
 export(build_graph)
 export(build_times)
