### 2017-11-17 3c2902a
Document batchtools_drake_tmpl_file()
