+++ vignettes/drake.Rmd
@@ -121,7 +121,7 @@ explore the dependency network,
 ```{r draakedepsdrakermd, eval = FALSE}
 outdated()
 missed()
-vis_drake_graph() # Now with subgraphs too.
+vis_drake_graph() # Same as drake_graph()
 dataframes_graph()
 render_drake_graph()
 read_drake_graph()
