+++ vignettes/quickstart.Rmd
@@ -134,7 +134,7 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-See also `dataframes_graph()`, `render_drake_graph()`, and `drake_config()` for faster and more customized regraphing.
+Also see the alias `drake_graph()`. Functions `dataframes_graph()`, `render_drake_graph()`, and `drake_config()` allow for faster and more customized regraphing.
 
 You can also check the dependencies of individual targets.
 
@@ -333,6 +333,7 @@ config$cache$get_hash(key = "reg2") # different
 
 ```{r graph3quick, eval = FALSE}
 # Hover, click, drag, zoom, and pan.
+# Same as drake_graph():
 vis_drake_graph(my_plan, width = "100%", height = "500px")
 ```
 
@@ -395,7 +396,7 @@ As you have seen with `reg2()`, `drake` reacts to changes. In other words, `make
 1. Upstream targets.
 1. For [dynamic knitr reports](https://yihui.name/knitr/) (with `knit('your_report.Rmd')` as a command in your workflow plan data frame), targets and imports mentioned in calls to `readd()` and `loadd()` in the code chunks to be evaluated. `Drake` treats these targets and imports as dependencies of the compiled output target (say, 'report.md').
 
-To explore the dependencies, please refer to the `deps()` and `tracked()` functions. Better yet, generate interactive graphs with `vis_drake_graph()` as shown above. Hover over the nodes in the graph to see the content that `drake` watches.
+To explore the dependencies, please refer to the `deps()` and `tracked()` functions. Better yet, generate interactive graphs with `drake_graph()` or `vis_drake_graph()` as shown above. Hover over the nodes in the graph to see the content that `drake` watches.
 
 There is more to reproducibility than just using `drake` to watch for dependencies. [Packrat](https://rstudio.github.io/packrat) creates a tightly-controlled local library of packages to extend the shelf life of your project. And with [Docker](https://www.docker.com/), you can execute your project on a [virtual machine](https://en.wikipedia.org/wiki/Virtual_machine) to ensure platform independence. Together, [packrat](https://rstudio.github.io/packrat) and [Docker](https://www.docker.com/) can help others reproduce your work even if they have different software and hardware.
 
