+++ vignettes/parallelism.Rmd
@@ -190,7 +190,12 @@ make(my_plan, parallelism = "future_lapply")
 You can even nest parallelism strategies together. In the following example, targets are submitted as jobs on the Univa Grid engine, and then `future`-style multicore parallelism is applied to each target's command individually.
 
 ```{r hybridparallelism, eval = FALSE}
-future::plan(list(batchjobs_sge, multiprocess))
+future::plan(
+  list(
+    batchjobs_sge(template = "sge-simple.tmpl"), 
+    multiprocess
+  )
+)
 make(my_plan, parallelism = "future_lapply")
 ```
 
@@ -201,6 +206,12 @@ drake_example("sge")   # Sun/Univa Grid Engine workflow and supporting files
 drake_example("slurm") # SLURM workflow and supporting files
 ```
 
+To just write the [batchtools](https://github.com/mllg/batchtools) `*.tmpl` for an example, use
+
+```{r writexampletmpl, eval = FALSE}
+batchtools_drake_tmpl_file("sge")   # Writes sge-simple.tmpl
+batchtools_drake_tmpl_file("slurm") # Writes batchtools.slurm.tmpl
+```
 
 Be sure to heed the previously-mentioned cautionary note about deploying too many jobs. In `"future_lapply"` parallelism, the `jobs` argument is totally ignored. In at least some cases, you can limit the maximum number of jobs to 2 by calling `options(mc.cores = 2)` before `make()`. Depending on the `future` backend you select with `future::plan()` or `future::plan()`, you might make use of one of the other environment variables listed in `?future::future.options`.
 
