### 2017-11-17 254c004
In sanitize_plan(), remove rows with blank targets
