### 2017-11-17 b014218
Fix batchtools_drake_tmpl_file()
Was writing the incorrect files before.
