### 2017-11-19 014d7f4
Fix new problem in clean() with namespaces
Until this commit, drake was cleaning too many namespaces.
Now, clean() uses cleaned_namespaces() to just clean the right ones.
