+++ NEWS.md
@@ -56,6 +56,7 @@
 - Add a new `storr` namespace called `imports` to be used in `is_imported()`. That way, the whole object need not be read to `clean()` is. `clean()` is much faster and safer.
 - Add a new `make_with_config()` function to do the work of `make()` on an existing internal configuration list from `drake_config()`.
 - Add a new function `batchtools_drake_tmpl_file()` to write a `batchtools` template file from one of the examples (`drake_example()`), if one exists.
+- Add a new `cleaned_namespaces()` function to show which `storr` namespaces are cleaned in `clean()`.
 
 # 2017-10-17
 
