+++ NAMESPACE
@@ -19,6 +19,7 @@ export(cached)
 export(check)
 export(check_plan)
 export(clean)
+export(cleaned_namespaces)
 export(config)
 export(configure_cache)
 export(dataframes_graph)
