### 2017-11-19 981f045
Try to fix appveyor build.
The previous commmit failed because stringi was just updated.

Temporary fix: https://github.com/krlmlr/r-appveyor/issues/69#issuecomment-229519640
