+++ appveyor.yml
@@ -13,6 +13,7 @@ install:
 # Adapt as necessary starting from here
 
 build_script:
+  - travis-tool.sh install_r stringi
   - travis-tool.sh install_deps
 
 test_script:
