### 2017-11-19 dc794c9
Use Rtools in appveyor build
https://github.com/krlmlr/r-appveyor/issues/69#issuecomment-235432008
