### 2017-11-19 ec275bb
Rewrite drake.Rmd vignette
