+++ vignettes/drake.Rmd
@@ -13,242 +13,94 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-![](logo-vignettes.png)
-
 ```{r, echo = F}
 suppressMessages(suppressWarnings(library(drake)))
+unlink(".drake", recursive = TRUE)
 clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
 ```
 
-# Data frames in R for [Make](http://kbroman.org/minimal_make/)
+Reproducibility carries an implicit promise that alleged results of a computation are up to date with the dependencies. In other words, the final tables, figures, and reports should match the data and code that generated them. This form of internal consistency is nontrivial for medium to large projects, where dependencies change frequently but the results take a long time to refresh. In fact, dependency awareness is arguably just as important as [scientific replicability](http://www.nature.com/news/1-500-scientists-lift-the-lid-on-reproducibility-1.19970), [literate programming](https://www.r-bloggers.com/reproducible-research-training-wheels-and-knitr/), and [version control](https://nicercode.github.io/git/why.html). However, few tools in R properly address the need, and it is a glaring blind spot in most conversations around reproducibility in the R and Statistics communities.
 
-Drake is a workflow manager and build system for
+# Reproducibility with drake
 
-1. [Reproducibility](https://CRAN.R-project.org/view=ReproducibleResearch).
-2. [High-performance computing](https://CRAN.R-project.org/view=HighPerformanceComputing).
+![](logo-vignettes.png)
 
-Organize your work in a data frame.
+Enter the [`drake` package](https://github.com/wlandau-lilly/drake), one of the only [pipeline toolkits](https://github.com/pditommaso/awesome-pipeline) available in R. `Drake` detects the dependencies of intermediate steps, links everything together in a [network graph](https://cran.r-project.org/web/packages/drake/vignettes/graph.html#dependency-reactivity), fingerprints the available data, and rebuilds only the pieces that are out of date or missing. For a demonstration, we turn to the built-in basic example.
 
-```{r firstintroindrake}
-library(drake)
-load_basic_example() # Also (over)writes report.Rmd.
-my_plan              # Each target is a file (single-quoted) or object.
-```
-
-Then `make()` it to build all your targets.
-
-```{r makedrakenoevalrmd, eval = FALSE}
-make(my_plan) # Run the commands to build the targets.
-```
-
-If a target fails, diagnose it.
-
-```{r diagnose, eval = FALSE}
-failed()                 # Targets that failed in the most recent `make()`
-diagnose()               # Targets that failed in any previous `make()`
-error <- diagnose(large) # Most recent verbose error log of `large`
-str(error)               # Object of class "error"
-error$calls              # Call stack / traceback
-```
-
-# Installation
-
-You can choose among different versions of `drake`:
-
-```{r devinstall, eval = FALSE}
-install.packages("drake")    # Latest CRAN release.
-install.packages("devtools") # For installing from GitHub.
-# Choose a GitHub release.
-devtools::install_github(
-  "wlandau-lilly/drake@v4.4.0",
-  build = TRUE
-)
-# Install the development version.
-install_github("wlandau-lilly/drake", build = TRUE) # dev version
-```
-
-- You must properly install `drake` using `install.packages()`, `devtools::install_github()`, or similar. It is not enough to use `devtools::load_all()`, particularly for the parallel computing functionality, in which multiple R sessions initialize and then try to `require(drake)`.
[…]
