### 2017-11-20 fba9830
Move check_drake_config(). Polish vignettes.
