### 2017-11-20 85275f2
Fix code coverage
