### 2017-11-20 e58afd1
Use formatR::tidy_source(). Was having problems: command "\"A\"" was interpreted as the symbol A, not the character string "A".
