+++ NAMESPACE
@@ -121,6 +121,7 @@ importFrom(codetools,findGlobals)
 importFrom(crayon,make_style)
 importFrom(digest,digest)
 importFrom(evaluate,try_capture_stack)
+importFrom(formatR,tidy_source)
 importFrom(future,future_lapply)
 importFrom(future,plan)
 importFrom(grDevices,col2rgb)
