### 2017-11-20 20bdde3
Update NEWS.md
Comment on change to `tidy_command()`, previously `tidy()`.
Now using `formatR::tidy_source()`.
