+++ NEWS.md
@@ -4,6 +4,7 @@
 - Let the user define a trigger for each target to customize when `make()` decides to build targets.
 - Document triggers and other debugging/testing tools in the new [debug vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/debug.Rmd).
 - Restructure the internals of the `storr` cache in a way that is not back-compatible with projects from versions 4.4.0 and earlier. The main change is to make more intelligent use of `storr` namespaces, improving speed and opening up possibilities for new features. If you attempt to run drake >= 5.0.0 on a project from drake <= 4.0.0, drake will stop you before any damage to the cache is done, and you will be instructed how to migrate your project to the new drake.
+- Use `formatR::tidy_source()` instead of `parse()` in `tidy_command()` (in `R/dependencies.R`). Previously, `drake` was having problems with an edge case: as a command, the literal string `"A"` was interpreted as the symbol `A` after tidying. With `tidy_source()`, literal quoted strings should stay literal quoted strings in commands. This may put some targets out of date in old projects, but that is unlikely. The change should be mostly back-compatible.
 - Speed up clean() by refactoring the cache inventory and using light parallelism.
 - Implement `rescue_cache()`, exposed to the user and used in `clean()`. This function removes dangling orphaned files in the cache so that a broken cache can be cleaned and used in the usual ways once more.
 - Change the default `cpu` and `elapsed` arguments of `make()` to `NULL`. This solves an elusive bug in how drake imposes timeouts.
