### 2017-11-20 f28bf8b
Set all formal arguments of formatR::tidy_source()
