### 2017-11-20 8bf40fb
Use Windows stringi build in Appveyor
Assumes stringi build is available. Will not merge
until Appveyor passes.
