### 2017-11-20 e4c6aab
Remove Rtools lines in appveyor
With the stringi build fixed, we're good for now.
