### 2017-11-20 2540a65
Use legacy_tidy() in migrate.R. Also hedge NEWS.md.
