+++ vignettes/debug.Rmd
@@ -140,10 +140,16 @@ outdated(my_plan, verbose = FALSE)
 
 dependency_profile(target = "regression2_small", config = config)
 
+config$cache$get_hash(key = "small", namespace = "kernels") # same
+
 config$cache$get_hash(key = "small") # same
+
+config$cache$get_hash(key = "reg2", namespace = "kernels") # same
 
 config$cache$get_hash(key = "reg2") # different
 ```
+
+In `drake`, the "kernel" of a target or import is the piece of the output that is reproducibly tracked. For ordinary R objects, the kernel is just the object itself. For custom external files, it is a separate hash. But for functions, the kernel is the deparsed body of the function, together with the dependency hash if the function is imported (see `drake:::store_function()`).
 
 # Test with triggers.
 
