### 2017-11-21 2e04a0b
Return config from prepare_distributed() rather than build_these
It's better form. Yes, it's still bad form to have both
a return value and a side effect, but in this case it makes sense.
It would be silly to put build_these or outdated_targets
in the cache just to avoid having a return value.
