+++ R/timestamp.R
@@ -1,10 +1,10 @@
-time_stamps <- function(build_these, config){
+time_stamps <- function(config){
   cache_path <- cache_path(config$cache)
   stamp_dir <- time_stamp_dir(cache_path)
   dir_empty(stamp_dir)
   write_time_stamp_template(cache_path)
   targets <- intersect(V(config$graph)$name, config$plan$target)
-  stamp_these <- setdiff(targets, build_these)
+  stamp_these <- setdiff(targets, config$outdated_targets)
   lightly_parallelize(
     stamp_these, write_time_stamp, jobs = config$jobs, config = config)
   return(invisible())
