### 2017-11-21 b63bfd4
Solve #20 properly
Add a 'deps' flag so the user can load
the dependencies of the targets rather than the targets themselves.
Requires that a 'drake_config()' list is already in the cache,
so you must need to have called 'make()' first.
