+++ man/loadd.Rd
@@ -6,7 +6,8 @@
 \usage{
 loadd(..., list = character(0), imported_only = FALSE, path = getwd(),
   search = TRUE, cache = drake::get_cache(path = path, search = search,
-  verbose = verbose), envir = parent.frame(), jobs = 1, verbose = TRUE)
+  verbose = verbose), envir = parent.frame(), jobs = 1, verbose = TRUE,
+  deps = FALSE)
 }
 \arguments{
 \item{...}{targets to load from the cache, as names (unquoted)
@@ -40,7 +41,18 @@ can be lightly parallelized via \code{parallel::mclapply()}.
 just set jobs to be an integer greater than 1. On Windows,
 \code{jobs} is automatically demoted to 1.}
 
-\item{verbose}{whether to print console messages}
+\item{verbose}{logical, whether to print console messages}
+
+\item{deps}{logical, whether to load the dependencies of the targets
+instead of the targets themselves. This is useful if you know your
+target failed and you want to debug the command in an interactive
+session with the dependencies in your workspace.
+One caveat: to find the dependencies,
+\code{\link{loadd}()} uses information that was stored
+in a \code{\link{drake_config}()} list and cached
+during the last \code{\link{make}()}.
+That means you need to have already called \code{\link{make}()}
+if you set \code{deps} to \code{TRUE}.}
 }
 \value{
 \code{NULL}
@@ -62,9 +74,12 @@ small
 # For many targets, you can parallelize loadd()
 # using the 'jobs' argument.
 loadd(list = c("small", "large"), jobs = 2)
-loadd(imported_only = TRUE) # Load all the imported objects/functions.
-# Load everything, including built targets. Be sure your computer
-# has enough memory.
+# Load the dependencies of the target, coef_regression2_small
+loadd(coef_regression2_small, deps = TRUE)
+# Load all the imported objects/functions.
+loadd(imported_only = TRUE)
+# Load everything, including built targets.
+# Be sure your computer has enough memory.
 loadd()
 }
 }
