### 2017-11-21 cc312ff
Document #20 in debug.Rmd
