### 2017-11-21 f24ffec
Add read_drake_meta()
and clean up some of the internals
