+++ R/progress.R
@@ -201,7 +201,7 @@ progress <- function(
 list_progress <- function(no_imported_objects, cache){
   all_marked <- cache$list(namespace = "progress")
   all_progress <- get_progress(target = all_marked, cache = cache)
-  plan <- read_plan(cache = cache)
+  plan <- read_drake_plan(cache = cache)
   abridged_marked <- Filter(
     all_marked,
     f = function(target){
