+++ R/mclapply.R
@@ -7,8 +7,12 @@ worker_mclapply <- function(targets, meta_list, config){
   prune_envir(targets = targets, config = config)
   jobs <- safe_jobs(config$jobs)
   values <- mclapply(
-    X = targets, FUN = drake_build, meta_list = meta_list,
-    config = config, mc.cores = jobs)
+    X = targets,
+    FUN = drake_build_worker,
+    meta_list = meta_list,
+    config = config,
+    mc.cores = jobs
+  )
   assign_to_envir(target = targets, value = values, config = config)
 }
 

[…]
