### 2017-11-21 b53ffce
Add a purge argument to clean()
