+++ R/cache.R
@@ -1,54 +1,3 @@
-#' @title Function cache_namespaces
-#' @export
-#' @seealso \code{\link{make}}
-#' @return A character vector of \code{storr} namespaces used for drake.
-#' @description List the important \code{storr} cache namespaces
-#' that need to be inventoried periodically in a call to \code{\link{make}()}.
-#' Ordinary users do not need to worry about this function.
-#' It is just another window into \code{drake}'s internals.
-#' @param default name of the default \code{storr} namespace
-#' @examples
-#' cache_namespaces()
-cache_namespaces <- function(
-  default = storr::storr_environment()$default_namespace
-){
-  c(
-    cleaned_namespaces(default = default),
-    "build_times",
-    "config",
-    "errors",
-    "progress",
-    "session"
-  ) %>%
-    sort
-}
-
-#' @title Function cleaned_namespaces
-#' @description List the \code{storr} namespaces that are cleaned
-#' during a call to \code{\link{clean}()}. All these
-#' namespaces store target-level data, but not all
-#' target-level namespaces are cleaned during
-#' \code{\link{clean}()}.
-#' @export
-#' @seealso \code{\link{cache_namespaces}}, \code{\link{clean}}
-#' @return A character vector of \code{storr} namespaces
-#' that are cleaned during \code{\link{clean}()}.
-#' @param default Name of the default \code{storr} namespace.
-#' @examples
-#' cleaned_namespaces()
-cleaned_namespaces <- function(
-  default = storr::storr_environment()$default_namespace
-){
-  c(
-    default,
-    "commands",
-    "depends",
-    "kernels",
-    "mtimes"
-  ) %>%
-    sort
-}
-
 #' @title Function cache_path
 #' @export
 #' @description Returns the file path
