+++ tests/testthat/test-full-build.R
@@ -80,6 +80,12 @@ test_with_dir("scratch build with custom filesystem cache.", {
   expect_true(file.exists(path))
   expect_equal(config$cache$list("file_modification_times"), character(0))
 
+  meta_list <- read_drake_meta(cache = cache)
+  expect_true(length(meta_list) > 0)
+  clean(purge = TRUE, cache = cache)
+  meta_list <- read_drake_meta(cache = cache)
+  expect_true(length(meta_list) == 0)
+
   clean(destroy = TRUE, cache = cache)
   expect_false(file.exists(path))
 })
