### 2017-11-21 3f461a7
Add a 'namespace' argument to cached()
so users can inspect individual storr namespaces
