### 2017-11-21 6c3aaa6
Update man pages
