+++ man/cached.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/cache_interaction.R
+% Please edit documentation in R/cache_ui.R
 \name{cached}
 \alias{cached}
 \title{Function \code{cached}}
