### 2017-11-21 4cb5bb8
Fix non-canonical CRAN URL
