### 2017-11-22 e4306d6
In loadd(..., deps = TRUE), only load deps that exist (cached)
