+++ R/read.R
@@ -96,8 +96,10 @@ readd <- function(
 #'
 #' @param verbose logical, whether to print console messages
 #'
-#' @param deps logical, whether to load the dependencies of the targets
-#' instead of the targets themselves. This is useful if you know your
+#' @param deps logical, whether to load any cached
+#' dependencies of the targets
+#' instead of the targets themselves.
+#' This is useful if you know your
 #' target failed and you want to debug the command in an interactive
 #' session with the dependencies in your workspace.
 #' One caveat: to find the dependencies,
@@ -154,7 +156,8 @@ loadd <- function(
   }
   if (deps){
     config <- read_drake_config(cache = cache)
-    targets <- dependencies(targets = targets, config = config)
+    targets <- dependencies(targets = targets, config = config) %>%
+      Filter(f = cache$exists)
   }
   lightly_parallelize(
     X = targets, FUN = load_target, cache = cache,
