### 2017-11-22 6f19985
Lightly parallelize checks that deps exist
In loadd(..., deps = TRUE)
