### 2017-11-22 d13eaec
Try to make loadd more memory safe
Counteract mclapply's memory inefficiency.
