+++ R/parallel_ui.R
@@ -189,7 +189,10 @@ default_parallelism <- function() {
 #' \code{\link{make}()}. \code{config$envir} is ignored in favor of
 #' \code{envir}.
 #'
-#' @param verbose logical, whether to output messages to the console.
+#' @param verbose logical or numeric, control printing to the console.
+#' Set to 0 (or \code{FALSE}) to print nothing,
+#' 1 (or \code{TRUE}) to print everything,
+#' or 2 to print everything except progress on individual imports.
 #'
 #' @param hook same as for \code{\link{make}}
 #'
@@ -293,7 +296,7 @@ default_parallelism <- function() {
 max_useful_jobs <- function(
   plan = workplan(), from_scratch = FALSE,
   targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = TRUE,
+  envir = parent.frame(), verbose = 2,
   hook = default_hook,
   cache = drake::get_cache(verbose = verbose),
   jobs = 1, parallelism = drake::default_parallelism(),
