+++ R/make.R
@@ -35,8 +35,10 @@
 #' from \code{envir} and the global environment and
 #' then reproducibly tracked as dependencies.
 #'
-#' @param verbose logical, whether to print progress to the console.
-#' Skipped objects are not printed.
+#' @param verbose logical or numeric, control printing to the console.
+#' Set to 0 (or \code{FALSE}) to print nothing,
+#' 1 (or \code{TRUE}) to print everything,
+#' or 2 to print everything except progress on individual imports.
 #'
 #' @param hook function with at least one argument.
 #' The hook is as a wrapper around the code that drake uses
@@ -238,7 +240,7 @@ make <- function(
   plan = workplan(),
   targets = drake::possible_targets(plan),
   envir = parent.frame(),
-  verbose = TRUE,
+  verbose = 2,
   hook = default_hook,
   cache = drake::get_cache(verbose = verbose, force = force),
   parallelism = drake::default_parallelism(),
