+++ R/predict_runtime.R
@@ -76,7 +76,7 @@ predict_runtime <- function(
   targets_only = FALSE,
   targets = drake::possible_targets(plan),
   envir = parent.frame(),
-  verbose = TRUE,
+  verbose = 2,
   hook = default_hook,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
   parallelism = drake::default_parallelism(),
@@ -233,7 +233,7 @@ rate_limiting_times <- function(
   targets_only = FALSE,
   targets = drake::possible_targets(plan),
   envir = parent.frame(),
-  verbose = TRUE,
+  verbose = 2,
   hook = default_hook,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
   parallelism = drake::default_parallelism(),
