### 2017-11-22 ac4374c
Change verbose = 2 to print everything, 1 to silence imports
