+++ R/parallel_ui.R
@@ -191,8 +191,8 @@ default_parallelism <- function() {
 #'
 #' @param verbose logical or numeric, control printing to the console.
 #' Set to 0 (or \code{FALSE}) to print nothing,
-#' 1 (or \code{TRUE}) to print everything,
-#' or 2 to print everything except progress on individual imports.
+#' 1 (or \code{TRUE}) to print everything except progress on individual imports,
+#' or 2 to print everything,
 #'
 #' @param hook same as for \code{\link{make}}
 #'
@@ -296,7 +296,7 @@ default_parallelism <- function() {
 max_useful_jobs <- function(
   plan = workplan(), from_scratch = FALSE,
   targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 2,
+  envir = parent.frame(), verbose = 1,
   hook = default_hook,
   cache = drake::get_cache(verbose = verbose),
   jobs = 1, parallelism = drake::default_parallelism(),
