+++ R/outdated.R
@@ -43,7 +43,7 @@ outdated <-  function(
   plan = workplan(),
   targets = drake::possible_targets(plan),
   envir = parent.frame(),
-  verbose = 2,
+  verbose = 1,
   hook = default_hook,
   cache = drake::get_cache(verbose = verbose),
   parallelism = drake::default_parallelism(),
@@ -158,7 +158,7 @@ missed <- function(
   plan = workplan(),
   targets = drake::possible_targets(plan),
   envir = parent.frame(),
-  verbose = 2,
+  verbose = 1,
   jobs = 1,
   parallelism = drake::default_parallelism(),
   packages = rev(.packages()),
