+++ man/drake_config.Rd
@@ -5,7 +5,7 @@
 \title{Function drake_config}
 \usage{
 drake_config(plan = workplan(), targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 2, hook = default_hook,
+  envir = parent.frame(), verbose = 1, hook = default_hook,
   cache = drake::get_cache(verbose = verbose, force = force),
   parallelism = drake::default_parallelism(), jobs = 1,
   packages = rev(.packages()), prework = character(0),
