+++ man/dataframes_graph.Rd
@@ -5,7 +5,7 @@
 \title{Function \code{dataframes_graph}}
 \usage{
 dataframes_graph(plan = workplan(), targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 2, hook = default_hook,
+  envir = parent.frame(), verbose = 1, hook = default_hook,
   cache = drake::get_cache(verbose = verbose), jobs = 1,
   parallelism = drake::default_parallelism(), packages = rev(.packages()),
   prework = character(0), build_times = TRUE, digits = 3,
@@ -27,8 +27,8 @@ of \code{envir}.}
 
 \item{verbose}{logical or numeric, control printing to the console.
 Set to 0 (or \code{FALSE}) to print nothing,
-1 (or \code{TRUE}) to print everything,
-or 2 to print everything except progress on individual imports.}
+1 (or \code{TRUE}) to print everything except progress on individual imports,
+or 2 to print everything,}
 
 \item{hook}{same as for \code{\link{make}}}
 
