+++ man/build_graph.Rd
@@ -5,7 +5,7 @@
 \title{Deprecated function \code{build_graph}}
 \usage{
 build_graph(plan = workplan(), targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 2, jobs = 1)
+  envir = parent.frame(), verbose = 1, jobs = 1)
 }
 \arguments{
 \item{plan}{Same as for \code{\link{build_drake_graph}()}.}
