### 2017-11-22 b23b61d
Fix console test
