+++ tests/testthat/test-console.R
@@ -33,6 +33,9 @@ test_with_dir("multiline message cap", {
 test_with_dir("console", {
   config <- dbug()
   config$verbose <- TRUE
+  expect_silent(console(imported = TRUE, target = "myinput",
+    config = config))
+  config$verbose = 2
   expect_message(console(imported = FALSE, target = "myinput",
     config = config))
   expect_message(console(imported = TRUE, target = "myinput",
