### 2017-11-22 54cc5d1
Remove rounding in verbose
