### 2017-11-22 50f9063
Lint test-console.R
