### 2017-11-23 d540eaa
Use call. in stop()
