### 2017-11-23 f3f9658
Improve trigger tests
