### 2017-11-23 298e4bd
Expand verbosity options:
Related: #166
0: nothing
1: targets and checks
2: +missing
3: +imports
