+++ R/parallel_ui.R
@@ -58,65 +58,66 @@ batchtools_drake_tmpl_file <- function(
 #' the following values of \code{x} to distribute targets over parallel
 #' units of execution.
 #' \describe{
-#'  \item{'parLapply'}{launches multiple processes in a single R session
-#'  using \code{parallel::\link{parLapply}()}.
-#'  This is single-node, (potentially) multicore computing.
-#'  It requires more overhead than the \code{'mclapply'} option,
-#'  but it works on Windows. If \code{jobs} is \code{1} in
-#'  \code{\link{make}()}, then no 'cluster' is created and
-#'  no parallelism is used.}
+#'   \item{'parLapply'}{launches multiple processes in a single R session
+#'   using \code{parallel::\link{parLapply}()}.
+#'   This is single-node, (potentially) multicore computing.
+#'   It requires more overhead than the \code{'mclapply'} option,
+#'   but it works on Windows. If \code{jobs} is \code{1} in
+#'   \code{\link{make}()}, then no 'cluster' is created and
+#'   no parallelism is used.}
 #'
-#'  \item{'mclapply'}{uses multiple processes in a single R session.
-#'  This is single-node, (potentially) multicore computing.
-#'  Does not work on Windows for \code{jobs > 1}
-#'  because \code{\link{mclapply}()} is based on forking.}
+#'   \item{'mclapply'}{uses multiple processes in a single R session.
+#'   This is single-node, (potentially) multicore computing.
+#'   Does not work on Windows for \code{jobs > 1}
+#'   because \code{\link{mclapply}()} is based on forking.}
 #'
-#'  \item{'future_lapply'}{
-#'  opens up a whole trove of parallel backends
-#'  powered by the \code{future} and \code{future.batchtools}
-#'  packages. First, set the parallel backend globally using
-#'  \code{future::plan()}.
-#'  Then, apply the backend to your workplan
-#'  using \code{make(..., parallelism = "future_lapply", jobs = ...)}.
-#'  But be warned: the environment for each target needs to be set up
-#'  from scratch, so this backend type is higher overhead than either
-#'  \code{mclapply} or \code{parLapply}.
-#'  Also, the \code{jobs} argument only applies to the imports.
-#'  To set the max number of jobs, set the \code{workers}
-#'  argument where it exists. For example, call
-#'  \code{future::plan(multisession(workers = 4))},
-#'  then call \code{\link{make}(your_plan, parallelism = "future_lapply")}.
-#'  You might also try options(mc.cores = jobs),
-#'  or see \code{?future::future::.options}
-#'  for environment variables that set the max number of jobs.
-#'  }
+#'   \item{'future_lapply'}{
+#'   opens up a whole trove of parallel backends
+#'   powered by the \code{future} and \code{future.batchtools}
+#'   packages. First, set the parallel backend globally using
+#'   \code{future::plan()}.
+#'   Then, apply the backend to your workplan
+#'   using \code{make(..., parallelism = "future_lapply", jobs = ...)}.
+#'   But be warned: the environment for each target needs to be set up
+#'   from scratch, so this backend type is higher overhead than either
+#'   \code{mclapply} or \code{parLapply}.
+#'   Also, the \code{jobs} argument only applies to the imports.
+#'   To set the max number of jobs, set the \code{workers}
+#'   argument where it exists. For example, call
+#'   \code{future::plan(multisession(workers = 4))},
+#'   then call \code{\link{make}(your_plan, parallelism = "future_lapply")}.
+#'   You might also try options(mc.cores = jobs),
[…]
