+++ man/check_plan.Rd
@@ -18,10 +18,7 @@ check_plan(plan = workplan(), targets = drake::possible_targets(plan),
 
 \item{cache}{optional drake cache. See \code{\link{new_cache}()}}
 
-\item{verbose}{logical or numeric, control printing to the console.
-Set to 0 (or \code{FALSE}) to print nothing,
-1 (or \code{TRUE}) to print everything except progress on individual imports,
-or 2 to print everything,}
+\item{verbose}{same as for \code{\link{make}()}}
 }
 \value{
 Invisibly return \code{plan}.
