+++ man/make.Rd
@@ -42,9 +42,14 @@ from \code{envir} and the global environment and
 then reproducibly tracked as dependencies.}
 
 \item{verbose}{logical or numeric, control printing to the console.
-Set to 0 (or \code{FALSE}) to print nothing,
-1 (or \code{TRUE}) to print everything except progress on individual imports,
-or 2 to print everything,}
+\describe{
+  \item{0 or \code{FALSE}:}{print nothing.}
+  \item{1 or \code{TRUE}:}{print checks and targets to build.}
+  \item{2:}{print checks, targets, to build,
+    and any potentially missing items.}
+  \item{3:}{full verbosity: print checks, targets to build, potentially
+    missing items, and imports.}
+}}
 
 \item{hook}{function with at least one argument.
 The hook is as a wrapper around the code that drake uses
