+++ man/dataframes_graph.Rd
@@ -25,10 +25,7 @@ dataframes_graph(plan = workplan(), targets = drake::possible_targets(plan),
 \code{\link{make}()}. \code{config$envir} is ignored in favor
 of \code{envir}.}
 
-\item{verbose}{logical or numeric, control printing to the console.
-Set to 0 (or \code{FALSE}) to print nothing,
-1 (or \code{TRUE}) to print everything except progress on individual imports,
-or 2 to print everything,}
+\item{verbose}{same as for \code{\link{make}()}}
 
 \item{hook}{same as for \code{\link{make}}}
 

[…]
