### 2017-11-24 2badae8
Speed up part of parallel_stage()
