### 2017-11-24 867f643
Shorten leaf node function (previously next_targets())
