+++ R/package.R
@@ -37,7 +37,7 @@
 #' @importFrom formatR tidy_source
 #' @importFrom future future_lapply plan
 #' @importFrom grDevices col2rgb rgb
-#' @importFrom igraph adjacent_vertices delete_vertices edge
+#' @importFrom igraph adjacent_vertices delete_vertices degree edge
 #' is_dag make_empty_graph plot.igraph subcomponent V vertex
 #' @importFrom knitr knit
 #' @importFrom lubridate dseconds duration
