+++ NAMESPACE
@@ -132,6 +132,7 @@ importFrom(grDevices,col2rgb)
 importFrom(grDevices,rgb)
 importFrom(igraph,V)
 importFrom(igraph,adjacent_vertices)
+importFrom(igraph,degree)
 importFrom(igraph,delete_vertices)
 importFrom(igraph,edge)
 importFrom(igraph,is_dag)
