### 2017-11-24 6a50067
Simple slow solution to #168
