### 2017-11-24 af4578b
Work more efficiently, but with the full graph
