### 2017-11-25 e44f761
Always process all the imports before building any targets.
This is part of the solution to #168: if imports and targets are processed together, the full power of parallelism is taken away from the targets. Also, the way parallelism happens is now consistent for all parallel backends.
