### 2017-11-25 4730139
Add #nocov lines where appropriate
