+++ NAMESPACE
@@ -45,6 +45,7 @@ export(drake_config)
 export(drake_example)
 export(drake_examples)
 export(drake_gc)
+export(drake_graph)
 export(drake_palette)
 export(drake_quotes)
 export(drake_session)
