### 2017-11-25 28a3f4c
Adjust vignettes and images
