### 2017-11-25 4a30984
Lint and adjust graph.png
