### 2017-11-25 240ba9d
Add a new store_meta arg to make()
