+++ R/read.R
@@ -329,7 +329,8 @@ read_drake_graph <- function(
 
 #' @title Function \code{read_drake_meta}
 #' @description For each target, read the information from
-#' \code{drake:::meta()} stored from the last \code{\link{make}()}.
+#' \code{drake:::meta()} stored from the last
+#' \code{\link{make}(..., store_meta = TRUE)}.
 #' This metadata was computed right before the target was built,
 #' and it was used in \code{drake:::should_build_target})
 #' to decide whether to build or skip the target.
