### 2017-11-25 38260c6
Get ready to sketch a new max_useful_jobs()
