### 2017-11-25 c9c85ac
More sketches: use imports_graph() and targets_graph()
