### 2017-11-25 275dca0
Sketch real_stages.R
to be used in max_useful_jobs() and rate_limiting_times()
