+++ man/max_useful_jobs.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/parallel_ui.R
+% Please edit documentation in R/max_useful_jobs.R
 \name{max_useful_jobs}
 \alias{max_useful_jobs}
 \title{Function \code{max_useful_jobs}}
