### 2017-11-25 4397306
Progress on real_stages() in max_useful_jobs()
