### 2017-11-25 48cdbaa
real_stages() is still buggy
Changing reg2() should make half the targets out of date
in the basic example. Instead, I get this:

> real_stages(config = config)

                item imported  file stage
1       'report.Rmd'     TRUE  TRUE     1
2       coefficients     TRUE FALSE     1
3         data.frame     TRUE FALSE     1
4               knit     TRUE FALSE     1
5                 lm     TRUE FALSE     1
6              rpois     TRUE FALSE     1
7       stats::rnorm     TRUE FALSE     1
8            summary     TRUE FALSE     1
9   suppressWarnings     TRUE FALSE     1
10              re
[…]
