### 2017-11-25 89b5409
Make strides in fixing real_stages()
