+++ R/envir.R
@@ -9,15 +9,8 @@ assign_to_envir <- Vectorize(
   )
 
 prune_envir <- function(targets, config){
-  downstream <- lightly_parallelize(
-    targets,
-    function(vertex){
-      subcomponent(config$graph, v = vertex, mode = "out")$name
-    },
-    jobs = config$jobs
-  ) %>%
-    unlist() %>%
-    unique()
+  downstream <- downstream_nodes(
+    from = targets, graph = config$graph, jobs = config$jobs)
   already_loaded <- ls(envir = config$envir, all.names = TRUE) %>%
     intersect(y = config$plan$target)
   load_these <- nonfile_target_dependencies(
