+++ R/real_stages.R
@@ -2,18 +2,23 @@ real_stages <- function(config){
   config$cache$clear(namespace = "real_stages")
   config$execution_graph <- imports_graph(config = config)
   resolve_real_stages(config = config)
-  config$execution_graph <- targets_graph(config = config)
+  targets_graph <- targets_graph(config = config)
+  config$execution_graph <- targets_graph
   resolve_real_stages(config = config)
-  keys <- config$cache$list(namespace = "real_stages") %>%
-    setdiff(y = "stage")
-  out <- lightly_parallelize(
-    X = keys,
-    FUN = function(key){
-      config$cache$get(key = key, namespace = "real_stages")
-    },
+  out <- read_real_stages(config = config)
+  also_outdated <- downstream_nodes(
+    from = out$item[!out$imported],
+    graph = config$graph,
     jobs = config$jobs
-  ) %>%
-    do.call(what = "rbind")
+  )
+  delete_these <- setdiff(V(targets_graph)$name, also_outdated)
+  config$execution_graph <- delete_vertices(
+    graph = targets_graph,
+    v = delete_these
+  )
+  config$trigger = "always"
+  resolve_real_stages(config = config)
+  out <- read_real_stages(config = config)
   config$cache$clear(namespace = "real_stages")
   out[order(out$stage, decreasing = FALSE), ]
 }
@@ -25,6 +30,14 @@ resolve_real_stages <- function(config){
 }
 
 worker_real_stages <- function(targets, meta_list, config){
+  imports <- setdiff(targets, config$plan$target)
+  if (any_imports){
+    worker_mclapply(
+      targets = targets,
+      meta_list = meta_list,
+      config = config
+    )
+  }
   if (!config$cache$exists(key = "stage", namespace = "real_stages")){
     config$cache$set(key = "stage", value = 1, namespace = "real_stages")
   }
@@ -33,7 +46,8 @@ worker_real_stages <- function(targets, meta_list, config){
     item = targets,
     imported = !targets %in% config$plan$target,
     file = is_file(targets),
-    stage = stage
+    stage = stage,
+    stringsAsFactors = FALSE
   )
   config$cache$set(key = "stage", value = stage + 1, namespace = "real_stages")
   config$cache$set(
@@ -42,4 +56,17 @@ worker_real_stages <- function(targets, meta_list, config){
     namespace = "real_stages"
   )
   invisible()
+}
+
+read_real_stages <- function(config){
+  keys <- config$cache$list(namespace = "real_stages") %>%
+    setdiff(y = "stage")
+  out <- lightly_parallelize(
+    X = keys,
+    FUN = function(key){
+      config$cache$get(key = key, namespace = "real_stages")
+    },
+    jobs = config$jobs
+  ) %>%
+    do.call(what = "rbind")
 }
