### 2017-11-25 4589a06
Switch targets to imports in a place in worker_real_stages
