### 2017-11-25 82e7316
Fix undefined symbol
