### 2017-11-25 ce73fa0
I think I fixed real_stages() and also max_useful_jobs()
