### 2017-11-25 0362394
Use real_stages() in rate_limiting_times()
Now it's just a matter of cleaning up and finishing #170.
