### 2017-11-25 a58a96c
Document parallel_stages() (previously real_stages())
