+++ NAMESPACE
@@ -85,6 +85,7 @@ export(mk)
 export(new_cache)
 export(outdated)
 export(output_sink_hook)
+export(parallel_stages)
 export(parallelism_choices)
 export(plan)
 export(plan_analyses)

[…]
