### 2017-11-25 cfd0f72
More reliance on the config list for some functions
