+++ R/dataframes_graph.R
@@ -7,11 +7,8 @@
 #' the legend nodes. The list also contains the
 #' default title of the graph.
 #' @seealso \code{\link{vis_drake_graph}}, \code{\link{build_drake_graph}}
-#' @param plan workflow plan data frame, same as for function
-#' \code{\link{make}()}.
-#'
-#' @param targets names of targets to build, same as for function
-#' \code{\link{make}()}.
+#' @param config a \code{\link{drake_config}()} configuration list.
+#' You can get one as a return value from \code{\link{make}()} as well.
 #'
 #' @param from Optional collection of target/import names.
 #' If \code{from} is nonempty,
@@ -37,45 +34,7 @@
 #' Applied after \code{from}, \code{mode}, and \code{order}.
 #' Be advised: edges are only kept for adjacent nodes in \code{subset}.
 #' If you do not select all the intermediate nodes,
-#' edges will drop from the graph.
-#'
-#' @param envir environment to import from, same as for function
-#' \code{\link{make}()}. \code{config$envir} is ignored in favor
-#' of \code{envir}.
-#'
-#' @param verbose same as for \code{\link{make}()}
-#'
-#' @param hook same as for \code{\link{make}}
-#'
-#' @param cache optional drake cache. Only used if the \code{config}
-#' argument is \code{NULL} (default). See code{\link{new_cache}()}.
-#'
-#' @param jobs The \code{outdated()} function is called internally,
-#' and it needs to import objects and examine your
-#' input files to see what has been updated. This could take some time,
-#' and parallel computing may be needed
-#' to speed up the process. The \code{jobs} argument is number of parallel jobs
-#' to use for faster computation.
-#'
-#' @param parallelism Choice of parallel backend to speed up the computation.
-#' Execution order in \code{\link{make}()} is slightly different when
-#' \code{parallelism} equals \code{'Makefile'}
-#' because in that case, all the imports are imported
-#' before any target is built.
-#' Thus, the arrangement in the graph is different for Makefile parallelism.
-#' See \code{?parallelism_choices} for details.
-#'
-#' @param font_size numeric, font size of the node labels in the graph
-#'
-#' @param packages same as for \code{\link{make}}
-#'
-#' @param prework same as for \code{\link{make}}
-#'
-#' @param build_times logical, whether to show the \code{\link{build_times}()}
-#' of the targets and imports, if available.
-#' These are just elapsed times from \code{system.time()}.
-#'
-#' @param digits number of digits for rounding the build times
+#' edges will drop from the graph.=
 #'
 #' @param targets_only logical,
 #' whether to skip the imports and only include the
@@ -90,23 +49,13 @@
 #' there may be more conditional independence than the graph
 #' indicates.)
 #'
-#' @param config option internal runtime parameter list of
-#' \code{\link{make}(...)},
-#' produced with \code{\link{config}()}.
-#' \code{config$envir} is ignored.
-#' Otherwise, computing this
-#' in advance could save time if you plan multiple calls to
-#' \code{dataframes_graph()}.
-#' If not \code{NULL},
-#' \code{config} overrides all arguments except
-#' \code{build_times}, \code{digits}, \code{targets_only},
-#' \code{split_columns}, and \code{font_size}.
+#' @param font_size numeric, font size of the node labels in the graph
 #'
-#' @param make_imports logical, whether to import external files
-#' and objects from the user's workspace to determine
-#' which targets are up to date. If \code{FALSE}, the computation
-#' is faster, but all the relevant information is drawn from the cache
[…]
