### 2017-11-25 688295d
Work on #171 through plot_graph.R
