+++ R/outdated.R
@@ -11,77 +11,28 @@
 #' @seealso \code{\link{missed}}, \code{\link{workplan}},
 #' \code{\link{make}}, \code{\link{vis_drake_graph}}
 #' @return Character vector of the names of outdated targets.
-#' @param plan same as for \code{\link{make}}
-#' @param targets same as for \code{\link{make}}
-#' @param envir same as for \code{\link{make}}. Overrides
-#' \code{config$envir}.
-#' @param verbose same as for \code{\link{make}}
-#' @param hook same as for \code{\link{make}}
-#' @param cache optional drake cache. See code{\link{new_cache}()}.
-#' The \code{cache} argument is ignored if a
-#' non-null \code{config} argument is supplied.
-#' @param parallelism same as for \code{\link{make}}
-#' @param jobs same as for \code{\link{make}}
-#' @param packages same as for \code{\link{make}}
-#' @param prework same as for \code{\link{make}}
-#' @param graph same as for \code{\link{make}}
 #' @param config option internal runtime parameter list of
 #' \code{\link{make}(...)},
 #' produced with \code{\link{config}()}.
+#' @param make_imports logical, whether to make the imports first.
+#' Set to \code{FALSE} to save some time and risk obselete output.
 #' \code{config$envir} is ignored.
 #' Overrides all the other arguments if not \code{NULL}.
 #' For example,
 #' \code{plan} is replaced with \code{config$plan}.
-#' @param make_imports logical, whether to import external files
-#' and objects from the user's workspace to determine
-#' which targets are up to date. If \code{FALSE}, the computation
-#' is faster, but all the relevant information is drawn from the cache
-#' and may be out of date.
-#' @param trigger same as for \code{\link{make}}
 #' @examples
 #' \dontrun{
-#' load_basic_example() # Load the canonical example of drake.
-#' outdated(my_plan) # Which targets are out of date?
-#' make(my_plan) # Run the projects, build the targets.
+#' config <- load_basic_example() # Load the canonical example of drake.
+#' outdated(config = config) # Which targets are out of date?
+#' config <- make(my_plan) # Run the projects, build the targets.
 #' # Now, everything should be up to date (no targets listed).
-#' outdated(my_plan)
+#' outdated(config = config)
 #' # outdated() is sensitive to triggers.
 #' # See the "debug" vignette for more on triggers.
-#' outdated(my_plan, trigger = "always")
+#' config$trigger <- "always"
+#' outdated(config = config)
 #' }
-outdated <-  function(
-  plan = workplan(),
-  targets = drake::possible_targets(plan),
-  envir = parent.frame(),
-  verbose = 1,
-  hook = default_hook,
-  cache = drake::get_cache(verbose = verbose),
-  parallelism = drake::default_parallelism(),
-  jobs = 1,
-  packages = rev(.packages()),
-  prework = character(0),
-  graph = NULL,
-  config = NULL,
-  make_imports = TRUE,
-  trigger = drake::default_trigger()
-){
-  force(envir)
-  if (is.null(config)){
-    config <- drake_config(
-      plan = plan,
-      targets = targets,
-      envir = envir,
-      verbose = verbose,
-      hook = hook,
-      cache = cache,
-      parallelism = parallelism,
-      jobs = jobs,
-      packages = packages,
-      prework = prework,
-      graph = graph,
-      trigger = trigger
-    )
-  }
+outdated <-  function(config, make_imports = TRUE){
   if (make_imports){
     make_imports(config = config)
   }
@@ -113,77 +64,18 @@ outdated <-  function(
 #' @seealso \code{\link{outdated}}
 #' @return Character vector of names of missing objects and files.
 #'
-#' @param plan workflow plan data frame, same as for function
-#' \code{\link{make}()}.
-#'
[…]
