+++ R/predict_runtime.R
@@ -13,24 +13,18 @@
 #' @examples
 #' \dontrun{
 #' load_basic_example() # Load drake's canonical example.
-#' make(my_plan) # Run the project, build the targets.
-#' predict_runtime(my_plan, digits = 4) # Everything is up to date.
-#' predict_runtime(my_plan, digits = 4, from_scratch = TRUE) # 1 job
+#' config <- make(my_plan) # Run the project, build the targets.
+#' predict_runtime(config, digits = 4) # Everything is up to date.
+#' predict_runtime(config, digits = 4, from_scratch = TRUE) # 1 job
 #' # Assumes you clean() out your project and start from scratch with 2 jobs.
-#' predict_runtime(my_plan, future_jobs = 2, digits = 4, from_scratch = TRUE)
-#' # Just predict how long it will take to build
-#' # the targets 'small' and 'large'.
-#' predict_runtime(
-#'   my_plan,
-#'   targets = c("small", "large"),
-#'   future_jobs = 2,
-#'   digits = 4,
-#'   from_scratch = TRUE
-#' )
+#' predict_runtime(config, future_jobs = 2, digits = 4, from_scratch = TRUE)
 #' }
 #' @return A \code{lubridate} \code{Duration} object
 #' with the predicted runtime of the next \code{\link{make}()}.
-#' @param plan same as for \code{\link{make}}
+#' @param config option internal runtime parameter list of
+#' \code{\link{make}(...)},
+#' produced by both \code{\link{make}()} and
+#' \code{\link{drake_config}()}.
 #' @param from_scratch logical, whether to predict a
 #' \code{\link{make}()} build from scratch or to
 #' take into account the fact that some targets may be
@@ -38,77 +32,23 @@
 #' @param targets_only logical, whether to factor in
 #' just the targets into the calculations or use the
 #' build times for everything, including the imports
-#' @param targets Targets to build in the workplan.
-#' Timing information is
-#' limited to these targets and their dependencies.
-#' @param envir same as for \code{\link{make}}
-#' @param verbose same as for \code{\link{make}}
-#' @param hook same as for \code{\link{make}}
-#' @param cache optional drake cache. See code{\link{new_cache}()}.
-#' The \code{cache} argument is ignored if a
-#' non-null \code{config} argument is supplied.
-#' @param parallelism same as for \code{\link{make}}. Used
-#' to parallelize import objects.
-#' @param jobs same as for \code{\link{make}}, just used to
-#' process imports
 #' @param future_jobs hypothetical number of jobs
 #' assumed for the predicted runtime.
 #' assuming this number of jobs.
-#' @param packages same as for \code{\link{make}}
-#' @param prework same as for \code{\link{make}}
-#' @param config option internal runtime parameter list of
-#' \code{\link{make}(...)},
-#' produced with \code{\link{config}()}.
-#' Overrides all arguments except \code{from_scratch},
-#' \code{targets_only}, and \code{digits}.
-#' Computing \code{config}
-#' in advance could save time if you plan multiple calls to
-#' this function.
-#' @param digits number of digits for rounding the times.
-#' @param path file path to the folder containing the cache.
-#' Yes, this is the parent directory containing the cache,
-#' not the cache itself.
-#' @param search logical, whether to search back in the file system
-#' for the cache.
 predict_runtime <- function(
-  plan = workplan(),
+  config,
+  future_jobs = 1,
   from_scratch = FALSE,
   targets_only = FALSE,
-  targets = drake::possible_targets(plan),
-  envir = parent.frame(),
-  verbose = 1,
[…]
