### 2017-11-25 6731a9e
Update all functions and examples for #170.
Need to update tests and vignettes.
