+++ man/rate_limiting_times.Rd
@@ -4,16 +4,14 @@
 \alias{rate_limiting_times}
 \title{Function rate_limiting_times}
 \usage{
-rate_limiting_times(plan = workplan(), from_scratch = FALSE,
-  targets_only = FALSE, targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 1, hook = default_hook,
-  cache = drake::get_cache(path = path, search = search, verbose = verbose),
-  parallelism = drake::default_parallelism(), jobs = 1,
-  future_jobs = jobs, packages = rev(.packages()), prework = character(0),
-  config = NULL, digits = 3, path = getwd(), search = TRUE)
+rate_limiting_times(config, from_scratch = FALSE, targets_only = FALSE,
+  future_jobs = 1, digits = 3)
 }
 \arguments{
-\item{plan}{same as for \code{\link{make}}}
+\item{config}{option internal runtime parameter list of
+\code{\link{make}(...)},
+produced by both \code{\link{make}()} and
+\code{\link{drake_config}()}.}
 
 \item{from_scratch}{logical, whether to assume
 next hypothetical call to \code{\link{make}()}
@@ -22,52 +20,11 @@ is a build from scratch (after \code{\link{clean}()}).}
 \item{targets_only}{logical, whether to factor in just the
 targets or use times from everything, including the imports.}
 
-\item{targets}{Targets to build in the workplan.
-Timing information is
-limited to these targets and their dependencies.}
-
-\item{envir}{same as for \code{\link{make}}. Supersedes
-\code{config$envir}.}
-
-\item{verbose}{same as for \code{\link{make}}}
-
-\item{hook}{same as for \code{\link{make}}}
-
-\item{cache}{optional drake cache. See code{\link{new_cache}()}.
-The \code{cache} argument is ignored if a
-non-null \code{config} argument is supplied.}
-
-\item{parallelism}{same as for \code{\link{make}}, and
-also used to process imported objects/files.}
-
-\item{jobs}{same as for \code{\link{make}}, just used to
-process imports.}
-
 \item{future_jobs}{hypothetical number of jobs
 assumed for the predicted runtime.
 assuming this number of jobs.}
 
-\item{packages}{same as for \code{\link{make}}}
-
-\item{prework}{same as for \code{\link{make}}}
-
-\item{config}{option internal runtime parameter list of
-\code{\link{make}(...)},
-produced with \code{\link{config}()}.
-Overrides all arguments except \code{from_scratch},
-\code{targets_only}, and \code{digits}.
-Computing \code{config}
-in advance could save time if you plan multiple calls to
-this function.}
-
 \item{digits}{number of digits for rounding the times.}
-
-\item{path}{file path to the folder containing the cache.
-Yes, this is the parent directory containing the cache,
-not the cache itself.}
-
-\item{search}{logical, whether to search back in the file system
-for the cache.}
 }
 \value{
 A data frame of times of the worst-case scenario
@@ -101,22 +58,13 @@ will warn you and tell you which targets have missing times.
 \examples{
 \dontrun{
 load_basic_example() # Load drake's canonical example.
-make(my_plan) # Run the project, build the targets.
-rate_limiting_times(my_plan) # Everything is up to date.
+config <- make(my_plan) # Run the project, build the targets.
+rate_limiting_times(config) # Everything is up to date.
 # Assume everything runs from scratch with 1 job.
-rate_limiting_times(my_plan, from_scratch  = TRUE, digits = 4)
+rate_limiting_times(config, from_scratch  = TRUE, digits = 4)
 # With 2 jobs, some of the targets are not rate-limiting.
 rate_limiting_times(
[…]
