+++ man/predict_runtime.Rd
@@ -4,16 +4,18 @@
 \alias{predict_runtime}
 \title{Function predict_runtime}
 \usage{
-predict_runtime(plan = workplan(), from_scratch = FALSE,
-  targets_only = FALSE, targets = drake::possible_targets(plan),
-  envir = parent.frame(), verbose = 1, hook = default_hook,
-  cache = drake::get_cache(path = path, search = search, verbose = verbose),
-  parallelism = drake::default_parallelism(), jobs = 1,
-  future_jobs = jobs, packages = rev(.packages()), prework = character(0),
-  config = NULL, digits = 3, path = getwd(), search = TRUE)
+predict_runtime(config, future_jobs = 1, from_scratch = FALSE,
+  targets_only = FALSE, digits = 3)
 }
 \arguments{
-\item{plan}{same as for \code{\link{make}}}
+\item{config}{option internal runtime parameter list of
+\code{\link{make}(...)},
+produced by both \code{\link{make}()} and
+\code{\link{drake_config}()}.}
+
+\item{future_jobs}{hypothetical number of jobs
+assumed for the predicted runtime.
+assuming this number of jobs.}
 
 \item{from_scratch}{logical, whether to predict a
 \code{\link{make}()} build from scratch or to
@@ -23,52 +25,6 @@ already up to date and therefore skipped.}
 \item{targets_only}{logical, whether to factor in
 just the targets into the calculations or use the
 build times for everything, including the imports}
-
-\item{targets}{Targets to build in the workplan.
-Timing information is
-limited to these targets and their dependencies.}
-
-\item{envir}{same as for \code{\link{make}}}
-
-\item{verbose}{same as for \code{\link{make}}}
-
-\item{hook}{same as for \code{\link{make}}}
-
-\item{cache}{optional drake cache. See code{\link{new_cache}()}.
-The \code{cache} argument is ignored if a
-non-null \code{config} argument is supplied.}
-
-\item{parallelism}{same as for \code{\link{make}}. Used
-to parallelize import objects.}
-
-\item{jobs}{same as for \code{\link{make}}, just used to
-process imports}
-
-\item{future_jobs}{hypothetical number of jobs
-assumed for the predicted runtime.
-assuming this number of jobs.}
-
-\item{packages}{same as for \code{\link{make}}}
-
-\item{prework}{same as for \code{\link{make}}}
-
-\item{config}{option internal runtime parameter list of
-\code{\link{make}(...)},
-produced with \code{\link{config}()}.
-Overrides all arguments except \code{from_scratch},
-\code{targets_only}, and \code{digits}.
-Computing \code{config}
-in advance could save time if you plan multiple calls to
-this function.}
-
-\item{digits}{number of digits for rounding the times.}
-
-\item{path}{file path to the folder containing the cache.
-Yes, this is the parent directory containing the cache,
-not the cache itself.}
-
-\item{search}{logical, whether to search back in the file system
-for the cache.}
 }
 \value{
 A \code{lubridate} \code{Duration} object
@@ -88,20 +44,11 @@ will warn you and tell you which targets have missing times.
 \examples{
 \dontrun{
 load_basic_example() # Load drake's canonical example.
-make(my_plan) # Run the project, build the targets.
-predict_runtime(my_plan, digits = 4) # Everything is up to date.
-predict_runtime(my_plan, digits = 4, from_scratch = TRUE) # 1 job
+config <- make(my_plan) # Run the project, build the targets.
+predict_runtime(config, digits = 4) # Everything is up to date.
+predict_runtime(config, digits = 4, from_scratch = TRUE) # 1 job
[…]
