### 2017-11-25 781792a
Fix test-arbitrary.cache.R
