### 2017-11-25 e63a290
Add envir to read_drake_config()
