### 2017-11-25 acab4ac
Re-fix test-arbitrary-cache.R
