### 2017-11-25 3a802ba
Work on fixing test-basic.R
