+++ R/max_useful_jobs.R
@@ -32,6 +32,10 @@
 #'  \item{'none'}{: Ignore all the imports and just focus on the max number
 #'    of useful jobs for parallelizing targets.}
 #' }
+#' 
+#' @param from_scratch logical, whether to assume
+#' the next \code{\link{make}()} will run from scratch
+#' so that all targets are attempted.
 #'
 #' @examples
 #' \dontrun{
@@ -69,17 +73,16 @@
 #' # By default, max_useful_jobs() takes into account which
 #' # targets are out of date. To assume you are building from scratch,
 #' # consider using the "always" trigger.
-#' config_from_scratch <- config
-#' config_from_scratch$trigger <- "always"
-#' max_useful_jobs(config = config_from_scratch, imports = 'files') # 8
-#' max_useful_jobs(config = config_from_scratch, imports = 'all') # 9
-#' max_useful_jobs(config = config_from_scratch, imports = 'none') # 8
+#' max_useful_jobs(config, from_scratch = TRUE, imports = 'files') # 8
+#' max_useful_jobs(config, from_scratch = TRUE, imports = 'all') # 9
+#' max_useful_jobs(config, from_scratch = TRUE, imports = 'none') # 8
 #' }
 max_useful_jobs <- function(
   config,
-  imports = c("files", "all", "none")
+  imports = c("files", "all", "none"),
+  from_scratch = FALSE
 ){
-  nodes <- parallel_stages(config = config)
+  nodes <- parallel_stages(config = config, from_scratch = from_scratch)
   imports <- match.arg(imports)
   if (imports == "none"){
     nodes <- nodes[!nodes$imported, ]
