+++ R/dataframes_graph_utils.R
@@ -144,7 +144,7 @@ resolve_levels <- function(config) { # nolint
   config$trigger = "always"
   stages <- parallel_stages(config)
   rownames(stages) <- stages$item
-  config$nodes$levels <- stages[rownames(config$nodes), "stage"]
+  config$nodes$level <- stages[rownames(config$nodes), "stage"]
   config$nodes
 }
 
