### 2017-11-25 42208a0
Fix test-basic.R
