### 2017-11-25 d67d722
Fix test-command-changes.R
