### 2017-11-25 1fdba08
Fix test-custom-caches.R
