### 2017-11-25 fd60ded
Fix test-dependencies.R
