### 2017-11-25 3e9e73b
Fix test-deprecate.R
