### 2017-11-25 2639ffd
Fix test-examples.R
