### 2017-11-25 c479695
Fix test-future.R
