### 2017-11-25 cce7272
Fix test-graph.R
