### 2017-11-25 39054b0
Fix up through test-namespaced.R
