### 2017-11-25 adad192
Fix up through test-other-features.R
