### 2017-11-25 757545f
Fix test-time.R
