+++ tests/testthat/test-time.R
@@ -42,25 +42,26 @@ test_with_dir("empty time predictions", {
   }
 
   my_plan <- workplan(y = 1)
+  config <- drake_config(my_plan)
   expect_warning(
-    x <- rate_limiting_times(plan = my_plan, verbose = FALSE) %>%
+    x <- rate_limiting_times(config) %>%
       min_df
   )
   expect_equal(nrow(x), 0)
-  make(my_plan, verbose = FALSE)
-  x <- rate_limiting_times(plan = my_plan, verbose = FALSE) %>%
+  config <- make(my_plan, verbose = FALSE)
+  x <- rate_limiting_times(config) %>%
     min_df
   expect_equal(nrow(x), 0)
-  x <- rate_limiting_times(plan = my_plan, verbose = FALSE,
+  x <- rate_limiting_times(config,
     targets_only = TRUE) %>%
     min_df
   expect_equal(nrow(x), 0)
-  x <- rate_limiting_times(plan = my_plan, verbose = FALSE,
+  x <- rate_limiting_times(config,
     from_scratch = TRUE) %>%
     min_df
   expect_equal(nrow(x), 1)
   expect_true(all(complete.cases(x)))
-  x <- rate_limiting_times(plan = my_plan, verbose = FALSE,
+  x <- rate_limiting_times(config,
     targets_only = TRUE, from_scratch = TRUE) %>%
     min_df
   expect_equal(nrow(x), 1)
@@ -83,55 +84,43 @@ test_with_dir("time predictions: incomplete targets", {
   my_plan <- e$my_plan
   config <- drake_config(my_plan, envir = e,
     jobs = 1, verbose = FALSE)
-  make_imports(config)
+  config <- make_imports(config)
 
   dats <- c("small", "large")
   expect_warning(
     x <- rate_limiting_times(
-      plan = config$plan,
-      targets = dats,
-      envir = config$envir,
-      verbose = FALSE
+      config,
+      targets = dats
     ) %>%
     min_df
   )
   expect_equal(nrow(x), 4)
   expect_warning(
     x <- rate_limiting_times(
-      plan = config$plan,
+      config,
       targets = dats,
-      envir = config$envir,
-      verbose = FALSE,
       targets_only = TRUE
     ) %>%
     min_df
   )
   expect_equal(nrow(x), 0)
   expect_warning(
-    x <- rate_limiting_times(
-      plan = config$plan,
-      envir = config$envir,
-      verbose = FALSE
-    ) %>%
+    x <- rate_limiting_times(config) %>%
     min_df
   )
   expect_equal(nrow(x), 12)
   expect_warning(
     y <- predict_runtime(
-      plan = config$plan,
+      config,
       targets = dats,
-      envir = config$envir,
-      verbose = FALSE,
       digits = Inf
     )
   )
   expect_equal(length(y), 1)
   expect_warning(
     y <- predict_runtime(
-      plan = config$plan,
+      config,
       targets = dats,
-      envir = config$envir,
-      verbose = FALSE,
       digits = Inf,
       targets_only = TRUE
     )
@@ -143,40 +132,36 @@ test_with_dir("time predictions: incomplete targets", {
 
   expect_silent(
     x <- rate_limiting_times(
-      plan = config$plan,
+      con,
       targets = dats,
-      envir = config$envir,
-      verbose = FALSE,
       from_scratch = TRUE
     ) %>%
     min_df
   )
   expect_equal(nrow(x), 6)
-  expect_warning(
+  
+  config <- drake_config(plan = con$plan, envir = con$envir, verbose = FALSE)
+  testrun(config)
+  expect_silent(
     x <- rate_limiting_times(
[…]
