+++ R/predict_runtime.R
@@ -189,9 +189,10 @@ rate_limiting_times <- function(
   nodes <- parallel_stages(config = config)
   times <- times[nodes$item, ]
   times$stage <- nodes$stage
-  ddply(times, "stage", rate_limiting_at_stage, future_jobs = future_jobs) %>%
+  out <- ddply(times, "stage", rate_limiting_at_stage, future_jobs = future_jobs) %>%
     round_times(digits = digits) %>%
     unname_rows
+  out[!is.na(out$item), ]
 }
 
 rate_limiting_at_stage <- function(stage, future_jobs){
