### 2017-11-25 346ac27
Fix test-triggers.R
