### 2017-11-25 0171638
Do some linting
