### 2017-11-25 07388c3
Add an optional config arg to make()
