+++ R/make.R
@@ -220,6 +220,10 @@
 #' so you can read it later with \code{\link{read_drake_meta}()}.
 #' Useful for debugging.
 #'
+#' @param config optional master configuration list created by
+#' \code{\link{drake_config}()}. Using one could cut out some
+#' superfluous overhead.
+#'
 #' @examples
 #' \dontrun{
 #' load_basic_example() # Load drake's canonical example.
@@ -275,7 +279,8 @@ make <- function(
   trigger = drake::default_trigger(),
   skip_imports = FALSE,
   skip_safety_checks = FALSE,
-  store_meta = TRUE
+  store_meta = TRUE,
+  config = NULL
 ){
   force(envir)
   if (!is.null(return_config)){
@@ -292,34 +297,36 @@ make <- function(
       call. = FALSE
     )
   }
-  config <- drake_config(
-    plan = plan,
-    targets = targets,
-    envir = envir,
-    verbose = verbose,
-    hook = hook,
-    parallelism = parallelism,
-    jobs = jobs,
-    packages = packages,
-    prework = prework,
-    prepend = prepend,
-    command = command,
-    args = args,
-    recipe_command = recipe_command,
-    clear_progress = TRUE,
-    cache = cache,
-    timeout = timeout,
-    cpu = cpu,
-    elapsed = elapsed,
-    retries = retries,
-    force = force,
-    graph = graph,
-    trigger = trigger,
-    imports_only = imports_only,
-    skip_imports = skip_imports,
-    skip_safety_checks = skip_safety_checks,
-    store_meta = store_meta
-  )
+  if (is.null(config)){
+    config <- drake_config(
+      plan = plan,
+      targets = targets,
+      envir = envir,
+      verbose = verbose,
+      hook = hook,
+      parallelism = parallelism,
+      jobs = jobs,
+      packages = packages,
+      prework = prework,
+      prepend = prepend,
+      command = command,
+      args = args,
+      recipe_command = recipe_command,
+      clear_progress = TRUE,
+      cache = cache,
+      timeout = timeout,
+      cpu = cpu,
+      elapsed = elapsed,
+      retries = retries,
+      force = force,
+      graph = graph,
+      trigger = trigger,
+      imports_only = imports_only,
+      skip_imports = skip_imports,
+      skip_safety_checks = skip_safety_checks,
+      store_meta = store_meta
+    )
+  }
   make_with_config(config = config)
 }
 
