+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: drake
 Title: Data Frames in R for Make
-Version: 4.4.1.9000
+Version: 4.4.1.9001
 Authors@R: c(
   person(
     family = "Landau",
