### 2017-11-25 b6a469e
Tweak arbitrary cache test
