### 2017-11-25 d63c948
Fix bug in clean
File targets were not being removed.
