+++ images/images.R
@@ -3,14 +3,14 @@
 # Requires pandoc.
 devtools::load_all() # load current drake
 clean(destroy = TRUE)
-load_basic_example()
+config <- load_basic_example()
 
-vis_drake_graph(my_plan, file = "outdated.html", selfcontained = TRUE,
+vis_drake_graph(config, file = "outdated.html", selfcontained = TRUE,
   width = "100%", height = "500px")
 
-make(my_plan)
+config <- make(my_plan)
 
-vis_drake_graph(my_plan, file = "built.html", selfcontained = TRUE,
+vis_drake_graph(config, file = "built.html", selfcontained = TRUE,
   width = "100%", height = "500px")
 
 reg2 <- function(d){
@@ -18,51 +18,40 @@ reg2 <- function(d){
   lm(y ~ x3, data = d)
 }
 
-vis_drake_graph(my_plan, file = "reg2.html", selfcontained = TRUE,
+vis_drake_graph(config, file = "reg2.html", selfcontained = TRUE,
   width = "100%", height = "500px")
 
 
 vis_drake_graph(
-  my_plan, file = "targetsonly.html", selfcontained = TRUE,
+  config, file = "targetsonly.html", selfcontained = TRUE,
   targets_only = TRUE,
   width = "100%", height = "500px",
   from = c("large", "small")
 )
 
 vis_drake_graph(
-  my_plan, file = "fromout.html", selfcontained = TRUE,
+  config, file = "fromout.html", selfcontained = TRUE,
   width = "100%", height = "500px",
   from = c("regression2_small", "regression2_large")
 )
 
 vis_drake_graph(
-  my_plan, file = "fromin.html", selfcontained = TRUE,
+  config, file = "fromin.html", selfcontained = TRUE,
   width = "100%", height = "500px",
   from = "small", mode = "in"
 )
 
 vis_drake_graph(
-  my_plan, file = "fromall.html", selfcontained = TRUE,
+  config, file = "fromall.html", selfcontained = TRUE,
   width = "100%", height = "500px",
   from = "small", mode = "all", order = 1
 )
 
 vis_drake_graph(
-  my_plan, file = "subset.html", selfcontained = TRUE,
+  config, file = "subset.html", selfcontained = TRUE,
   width = "100%", height = "500px",
   subset = c("regression2_small", "'report.md'")
 )
-
-small_plan <- workplan(a = 1, b = f(2))
-f <- function(x){
-  x
-}
-
-vis_drake_graph(small_plan, file = "small_local.html", selfcontained = TRUE,
-  width = "100%", height = "500px")
-
-vis_drake_graph(small_plan, file = "small_distributed.html", selfcontained = TRUE,
-  width = "100%", height = "500px", parallelism = "future_lapply")
 
 clean(destroy = TRUE)
 unlink("report.Rmd")
