### 2017-11-25 13ddee3
Tweak basic example
