### 2017-11-25 26cff9c
Update vignettes through debug.Rmd
