### 2017-11-25 39d541b
Udpate drake.Rmd vignette
