### 2017-11-25 90e6c37
Update all vignettes
