+++ vignettes/graph.Rmd
@@ -36,7 +36,8 @@ Initially, your entire project is out of date.
 ```{r graphoutdated}
 library(drake)
 load_basic_example()
-vis_drake_graph(my_plan) # Same as drake_graph()
+config <- drake_config(my_plan)
+vis_drake_graph(config) # Same as drake_graph()
 ```
 
 <iframe
@@ -49,8 +50,8 @@ style="border: none; box-shadow: none">
 In the previous graph, all the targets were out of date. But after a `make()`, we will be all caught up, and the graph will show you.
 
 ```{r graphmake}
-make(my_plan, jobs = 4, verbose = FALSE)
-vis_drake_graph(my_plan)
+config <- make(my_plan, jobs = 4, verbose = FALSE)
+vis_drake_graph(config)
 ```
 
 <iframe
@@ -67,7 +68,7 @@ reg2 <- function(d){
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
-vis_drake_graph(my_plan)
+vis_drake_graph(config)
 ```
 
 <iframe
@@ -81,7 +82,7 @@ style="border: none; box-shadow: none">
 Graphs can grow enormous for serious projects, so there are multiple ways to focus on a manageable subgraph. The most brute-force way is to just pick a manual `subset` of nodes to show. However, with the `subset` argument, `vis_drake_graph()` is prone to losing intermediate nodes and thus dropping edges.
 
 ```{r subsetgraph}
-vis_drake_graph(my_plan, subset = c("regression2_small", "'report.md'"))
+vis_drake_graph(config, subset = c("regression2_small", "'report.md'"))
 ```
 
 <iframe
@@ -93,7 +94,7 @@ style="border: none; box-shadow: none">
 The other subgraph functionality is much better at preserving connectedness. Use `targets_only` to ignore the imports.
 
 ```{r targetsonly}
-vis_drake_graph(my_plan, targets_only = TRUE)
+vis_drake_graph(config, targets_only = TRUE)
 ```
 
 <iframe
@@ -106,7 +107,7 @@ style="border: none; box-shadow: none">
 Similarly, you can just show downstream nodes.
 
 ```{r fromout}
-vis_drake_graph(my_plan, from = c("regression2_small", "regression2_large"))
+vis_drake_graph(config, from = c("regression2_small", "regression2_large"))
 ```
 
 <iframe
@@ -119,7 +120,7 @@ style="border: none; box-shadow: none">
 Or upstream ones.
 
 ```{r fromin}
-vis_drake_graph(my_plan, from = "small", mode = "in")
+vis_drake_graph(config, from = "small", mode = "in")
 ```
 
 <iframe
@@ -132,7 +133,7 @@ In fact, let's just take a small neighborhood around a target in both directions
 
 
 ```{r fromall}
-vis_drake_graph(my_plan, from = "small", mode = "all", order = 1)
+vis_drake_graph(config, from = "small", mode = "all", order = 1)
 ```
 
 <iframe
