+++ vignettes/parallelism.Rmd
@@ -35,14 +35,14 @@ knitr::opts_chunk$set(
 ```{r hpcplotgraph, eval = FALSE}
 clean()
 load_basic_example()
-make(my_plan, jobs = 2, verbose = FALSE) # Parallelize over 2 jobs.
+config <- make(my_plan, jobs = 2, verbose = FALSE) # Parallelize over 2 jobs.
 # Change a dependency.
 reg2 <- function(d) {
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
 # Hover, click, drag, zoom, and pan.
-vis_drake_graph(my_plan, width = "100%", height = "500px")
+vis_drake_graph(config, width = "100%", height = "500px")
 ```
 
 <iframe
@@ -68,29 +68,30 @@ For `drake`, the max useful jobs is the maximum number of targets in any paralle
 ```{r hpcquick, eval = FALSE}
 library(drake)
 load_basic_example()
-vis_drake_graph(my_plan) # Set targets_only to TRUE for smaller graphs.
-max_useful_jobs(my_plan) # 8
-max_useful_jobs(my_plan, imports = "files") # 8
-max_useful_jobs(my_plan, imports = "all") # 8
-max_useful_jobs(my_plan, imports = "none") # 8
-make(my_plan, jobs = 4)
-vis_drake_graph(my_plan)
+config <- drake_config(my_plan)
+vis_drake_graph(config) # Set targets_only to TRUE for smaller graphs.
+max_useful_jobs(config) # 8
+max_useful_jobs(config, imports = "files") # 8
+max_useful_jobs(config, imports = "all") # 8
+max_useful_jobs(config, imports = "none") # 8
+config <- make(my_plan, jobs = 4)
+vis_drake_graph(config)
 # Ignore the targets already built.
-max_useful_jobs(my_plan) # 1
-max_useful_jobs(my_plan, imports = "files") # 1
-max_useful_jobs(my_plan, imports = "all") # 8
-max_useful_jobs(my_plan, imports = "none") # 0
+max_useful_jobs(config) # 1
+max_useful_jobs(config, imports = "files") # 1
+max_useful_jobs(config, imports = "all") # 8
+max_useful_jobs(config, imports = "none") # 0
 # Change a function so some targets are now out of date.
 reg2 <- function(d){
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
-vis_drake_graph(my_plan)
-max_useful_jobs(my_plan) # 4
-max_useful_jobs(my_plan, from_scratch = TRUE) # 8
-max_useful_jobs(my_plan, imports = "files") # 4
-max_useful_jobs(my_plan, imports = "all") # 8
-max_useful_jobs(my_plan, imports = "none") # 4
+vis_drake_graph(config)
+max_useful_jobs(config) # 4
+max_useful_jobs(config, from_scratch = TRUE) # 8
+max_useful_jobs(config, imports = "files") # 4
+max_useful_jobs(config, imports = "all") # 8
+max_useful_jobs(config, imports = "none") # 4
 ```
 
 # Parallel backends
