### 2017-11-25 9c9aa18
Explain real parallel stages a bit better
