+++ vignettes/drake.Rmd
@@ -147,7 +147,7 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-The columns in the graph represent parallelizable stages. All the objects in a column are conditionally independent given their dependencies. In `make()`, the columns are processed sequentially from left to right, and everything within an individual column is parallelized over the available `jobs`. Assuming the targets are rate-limiting (as opposed to the imports), the next `make(..., jobs = 4)` should be faster than `make(..., jobs = 1)`, and it would be superfluous to use more than 4 `jobs`.
+The columns in the graph represent parallelizable stages. All the objects in a column are conditionally independent given their dependencies. In `make()`, the columns are processed sequentially from left to right, and everything within an individual column is parallelized over the available `jobs`. (This is not a hard and fast rule: `drake` includes targets further ahead if some targets in the current column are already up to date.) Assuming the targets are rate-limiting (as opposed to the imports), the next `make(..., jobs = 4)` should be faster than `make(..., jobs = 1)`, and it would be superfluous to use more than 4 `jobs`.
 
 `Drake` supports a [vast arsenal](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd#parallel-backends) of parallel computing backends. You can leverage the spare cores on your laptop's processor or scale up to serious distributed computing on a cluster. You can list the supported backends with `parallellism_choices()`.
 
