+++ vignettes/timing.Rmd
@@ -115,7 +115,7 @@ To predict the next runtime with multiple parallel jobs, drake makes some assump
 1. The outdated targets are spread out evenly over the available jobs. (All targets are used if `from_scratch` is `TRUE`)
 1. One job gets all the slowest targets (pessimistic scenario).
 
-Then, `drake` simply takes the targets from the slowest job in each parallelizable stage and sums the corresponding elapsed build times. A parallelizable stage is a usually a column in the workflow graph.
+Then, `drake` simply takes the targets from the slowest job in each parallelizable stage and sums the corresponding elapsed build times. A parallelizable stage is a usually a column in the workflow graph, but if there are up-to-date targets in a column, `drake` skips ahead to try to fit as many conditionally independent targets as possible in a stage.
 
 ```{r faketiminggraph, eval = FALSE}
 # Hover, click, drag, zoom, and pan.
