### 2017-11-25 1fa2725
Change vignette image hashes
