+++ vignettes/drake.Rmd
@@ -77,7 +77,7 @@ vis_drake_graph(my_plan)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/outdated.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/outdated.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -142,7 +142,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/reg2.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/reg2.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
