+++ vignettes/quickstart.Rmd
@@ -136,7 +136,7 @@ vis_drake_graph(config, width = "100%", height = "500px")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/outdated.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/outdated.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -294,7 +294,7 @@ vis_drake_graph(config, width = "100%", height = "500px")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/built.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/built.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -370,7 +370,7 @@ vis_drake_graph(config, width = "100%", height = "500px")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/reg2.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/reg2.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
