+++ vignettes/graph.Rmd
@@ -41,7 +41,7 @@ vis_drake_graph(config) # Same as drake_graph()
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/outdated.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/outdated.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -55,7 +55,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/built.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/built.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -72,7 +72,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/reg2.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/reg2.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -86,7 +86,7 @@ vis_drake_graph(config, subset = c("regression2_small", "'report.md'"))
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/subset.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/subset.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -98,7 +98,7 @@ vis_drake_graph(config, targets_only = TRUE)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/targetsonly.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/targetsonly.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -111,7 +111,7 @@ vis_drake_graph(config, from = c("regression2_small", "regression2_large"))
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/fromout.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/fromout.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -124,7 +124,7 @@ vis_drake_graph(config, from = "small", mode = "in")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/fromin.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/fromin.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -137,7 +137,7 @@ vis_drake_graph(config, from = "small", mode = "all", order = 1)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/fromall.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/fromall.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
[…]
