+++ vignettes/debug.Rmd
@@ -84,7 +84,7 @@ vis_drake_graph(config, width = "100%", height = "500px")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/outdated.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/outdated.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
