+++ vignettes/timing.Rmd
@@ -123,7 +123,7 @@ vis_drake_graph(my_plan, width = "100%", height = "500px")
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/2211b300/images/reg2.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/reg2.html"
 width = "100%" height = "600px" allowtransparency="true" 
 style="border: none; box-shadow: none">
 </iframe>
