### 2017-11-26 56a9d25
Clean up rate_limiting_times() internals
