+++ R/config.R
@@ -47,6 +47,7 @@
 #' @param skip_safety_checks logical, whether to skip the safety checks
 #' on your workflow to save time. Use at your own peril.
 #' @param store_meta same as for \code{\link{make}}
+#' @param lazy_load same as for \code{\link{make}}
 #' @examples
 #' \dontrun{
 #' load_basic_example() # Load drake's canonical example.
@@ -92,7 +93,8 @@ drake_config <- function(
   imports_only = FALSE,
   skip_imports = FALSE,
   skip_safety_checks = FALSE,
-  store_meta = TRUE
+  store_meta = TRUE,
+  lazy_load = FALSE
 ){
   force(envir)
   seed <- get_valid_seed()
@@ -135,7 +137,8 @@ drake_config <- function(
     seed = seed, trigger = trigger,
     timeout = timeout, cpu = cpu, elapsed = elapsed, retries = retries,
     imports_only = imports_only, skip_imports = skip_imports,
-    skip_safety_checks = skip_safety_checks, store_meta = store_meta
+    skip_safety_checks = skip_safety_checks, store_meta = store_meta,
+    lazy_load = lazy_load
   )
 }
 
