### 2017-11-26 dacb0c6
Turn off assign_to_envir for lazy loading
