### 2017-11-26 f0810b6
Contemplate lazy_oad in migrate()
