### 2017-11-26 a371da8
Add complete lazy load test
