### 2017-11-26 76cd803
Add lazy load test file
