### 2017-11-26 18279a4
Continue sketch of lazy loading
