+++ tests/testthat/test-lazy-load.R
@@ -34,6 +34,8 @@ test_with_dir("lazy loading is actually lazy", {
   config <- make(
     lazy_load = TRUE,
     plan = config$plan,
+    parallelism = config$parallelism,
+    jobs = config$jobs,
     targets = "combined",
     envir = config$envir,
     verbose = TRUE
@@ -46,6 +48,8 @@ test_with_dir("lazy loading is actually lazy", {
   config <- make(
     lazy_load = FALSE,
     plan = config$plan,
+    parallelism = config$parallelism,
+    jobs = config$jobs,
     targets = "combined",
     envir = config$envir,
     verbose = TRUE
