### 2017-11-26 412ccd7
Silence verbosity in lazy load test
