### 2017-11-26 802c7c7
Use default backend for part of lazy loading test
