### 2017-11-26 2ab8b9a
Try to get ready for globalenv()
