### 2017-11-26 4e52244
Dodge parLapply with multiple jobs: doesnt work there
