+++ tests/testthat/test-lazy-load.R
@@ -4,6 +4,9 @@ test_with_dir("no overt errors lazy load for the debug example", {
   config <- dbug()
   config$verbose <- FALSE
   config$lazy_load <- TRUE
+  if (config$parallelism == "parLapply"){
+    config$jobs <- 1
+  }
   expect_equal(sort(outdated(config)), sort(config$plan$target))
 
   config <- testrun(config)
