### 2017-11-26 53fa40a
Fix help files
