+++ man/make.Rd
@@ -15,7 +15,8 @@ make(plan = workplan(), targets = drake::possible_targets(plan),
   imports_only = FALSE, timeout = Inf, cpu = NULL, elapsed = NULL,
   retries = 0, force = FALSE, return_config = NULL, graph = NULL,
   trigger = drake::default_trigger(), skip_imports = FALSE,
-  skip_safety_checks = FALSE, store_meta = TRUE, config = NULL)
+  skip_safety_checks = FALSE, store_meta = TRUE, config = NULL,
+  lazy_load = FALSE)
 }
 \arguments{
 \item{plan}{workflow plan data frame.
@@ -228,7 +229,20 @@ Useful for debugging.}
 
 \item{config}{optional master configuration list created by
 \code{\link{drake_config}()}. Using one could cut out some
-superfluous overhead.}
+superfluous overhead. Overrides all other arguments
+if supplied}
+
+\item{lazy_load}{logical. If \code{FALSE},
+drake prunes the execution environment before every
+parallelizable stages, removing all superfluous targets
+and then loading any dependencies it will need
+for the targets in the current parallelizable stage.
+In other words, drake prepares the environment in advance
+for all the whole collection of targets in the stage.
+If \code{lazy_load} is \code{TRUE}, drake assigns
+promises to load any dependencies at the last minute.
+Lazy loading may be more memory efficient in some use cases, but
+it may duplicate the loading of dependencies, costing time.}
 }
 \value{
 The master internal configuration list, mostly
