### 2017-11-26 6b90b50
Fix help files
