+++ R/make.R
@@ -225,7 +225,10 @@
 #' superfluous overhead. Overrides all other arguments
 #' if supplied
 #' 
-#' @param lazy_load logical. If \code{FALSE},
+#' @param lazy_load logical.
+#' Should always be set to \code{FALSE} for \code{"parLapply"}
+#' parallelism and \code{jobs} greater than 1.
+#' If \code{lazy_load} is \code{FALSE},
 #' drake prunes the execution environment before every
 #' parallelizable stages, removing all superfluous targets
 #' and then loading any dependencies it will need
