### 2017-11-26 3fd17c2
Warn about lazy_load in the caution vignette
