+++ R/outdated.R
@@ -15,7 +15,7 @@
 #' \code{\link{make}(...)},
 #' produced with \code{\link{config}()}.
 #' @param make_imports logical, whether to make the imports first.
-#' Set to \code{FALSE} to save some time and risk obselete output.
+#' Set to \code{FALSE} to save some time and risk obsolete output.
 #' \code{config$envir} is ignored.
 #' Overrides all the other arguments if not \code{NULL}.
 #' For example,
