### 2017-11-26 b06f82e
Tweak mk()
