### 2017-11-26 4fbc9e2
Dont repeat prework in parallel_stages()
