+++ R/parallel_stages.R
@@ -40,16 +40,18 @@
 #' parallel_stages(config = config)
 #' }
 parallel_stages <- function(config, from_scratch = FALSE){
+  do_prework(config = config, verbose_packages = TRUE)
+  config$store_meta <- FALSE
   if (from_scratch){
     config$trigger <- "always"
   }
   config$stages_cache <- storr::storr_environment()
   config$stages_cache$clear()
   config$execution_graph <- imports_graph(config = config)
-  resolve_parallel_stages(config = config)
+  run_parallel(config = config, worker = worker_parallel_stages)
   targets_graph <- targets_graph(config = config)
   config$execution_graph <- targets_graph
-  resolve_parallel_stages(config = config)
+  run_parallel(config = config, worker = worker_parallel_stages)
   out <- read_parallel_stages(config = config)
   if (!length(out)){
     return(data.frame(
@@ -71,16 +73,10 @@ parallel_stages <- function(config, from_scratch = FALSE){
     v = delete_these
   )
   config$trigger <- "always"
-  resolve_parallel_stages(config = config)
+  run_parallel(config = config, worker = worker_parallel_stages)
   out <- read_parallel_stages(config = config)
   config$stages_cache$clear()
   out[order(out$stage, decreasing = FALSE), ]
-}
-
-resolve_parallel_stages <- function(config){
-  do_prework(config = config, verbose_packages = TRUE)
-  config$store_meta <- FALSE
-  run_parallel(config = config, worker = worker_parallel_stages)
 }
 
 worker_parallel_stages <- function(targets, meta_list, config){
