### 2017-11-26 50b5c92
Reshuffle functions
