### 2017-11-26 6bff576
Reshuffle functions
