### 2017-11-26 bc0de77
Add next_stage()
