+++ NAMESPACE
@@ -83,6 +83,7 @@ export(migrate_drake_project)
 export(missed)
 export(mk)
 export(new_cache)
+export(next_stage)
 export(outdated)
 export(output_sink_hook)
 export(parallel_stages)
