### 2017-11-26 7666f38
Modify time stamping
