+++ R/lapply.R
@@ -1,5 +1,5 @@
 run_lapply <- function(config){
-  do_prework(config = config, verbose_packages = TRUE)
+  do_prework(config = config, verbose_packages = config$verbose)
   run_parallel(config = config, worker = worker_lapply)
 }
 
