### 2017-11-26 0f39eee
Lint files and fix #173
