### 2017-11-26 ab4180a
Need to make_imports() for Makefile parallelism
