### 2017-11-26 8723d95
Work next_stage() into one of the tests
