### 2017-11-27 799b0b5
Remove redundant prework statements
