### 2017-11-28 f0c78d8
Touch up docs and streamline parallel stage internals
