### 2017-11-28 a5e81b3
Remove superfluous namespace in next_stage() functions
