### 2017-11-29 wlandau-lilly/drake#175 @wlandau-lilly
Too many deprecated functions
`Drake` has undergone several changes in its interface recently. For example, I am about to deprecate `workplan()` in favor of `plan_drake()`. Decisions like these are the right ones, but it is a lot to keep up with, and I apologize. @jules32 and @benmarwick, I am sorry if all this makes your reviews for ropensci/onboarding#156 more difficult.
