### 2017-11-29 d68b8de
Deprecate workplan() in favor of plan_drake()
See #175
