+++ tests/testthat/test-hash.R
@@ -7,7 +7,7 @@ test_with_dir("available hash algos", {
 })
 
 test_with_dir("illegal hashes", {
-  x <- workplan(a = 1)
+  x <- plan_drake(a = 1)
   expect_error(make(x, short_hash_algo = "no_such_algo_aslkdjfoiewlk"))
   expect_error(make(x, long_hash_algo = "no_such_algo_aslkdjfoiewlk"))
 })
