### 2017-11-29 4ba29a3
Fix deprecation message
