### 2017-11-30 a0f74f1
Rewrite the drake.Rmd vignette
Was poorly written before. Needed overhaul.
