+++ vignettes/drake.Rmd
@@ -26,25 +26,20 @@ knitr::opts_chunk$set(
 )
 ```
 
-Reproducibility carries an implicit promise that alleged results of a computation are up to date with the dependencies. In other words, the final tables, figures, and reports should match the data and code that generated them. This form of internal consistency is nontrivial for medium to large projects, where dependencies change frequently but the results take a long time to refresh. In fact, dependency awareness is arguably just as important as [scientific replicability](http://www.nature.com/news/1-500-scientists-lift-the-lid-on-reproducibility-1.19970), [literate programming](https://www.r-bloggers.com/reproducible-research-training-wheels-and-knitr/), and [version control](https://nicercode.github.io/git/why.html). However, few tools in R properly address the need, and it is a conspicuous blind spot in most discussions about reproducibility in the R and Statistics communities.
-
-# Reproducibility with drake
-
 ![](logo-vignettes.png)
 
-Enter the [`drake` package](https://github.com/wlandau-lilly/drake), one of the only [pipeline toolkits](https://github.com/pditommaso/awesome-pipeline) available in R. `Drake` detects the dependencies of intermediate steps, links everything together in a [network graph](https://cran.r-project.org/package=drake/vignettes/graph.html#dependency-reactivity), fingerprints the available data, and rebuilds only the pieces that are out of date or missing. For a demonstration, we turn to the built-in basic example.
+Reproducibility carries the promise that your output matches your input. Your tables, figures, reports, and intermediate objects should stay up to date with all the underlying data and code. To ensure internal consistency, `drake` detects and refreshes the outdated components of your project. The same solution unlocks powerful high-performance computing functionality.
 
-```{r reproducibilitydemo}
-load_basic_example()
-```
+# Reproducibility with `drake`
 
-The intermediate components of the project, or "targets", are declared in an ordinary data frame.
+Outline your project as a data frame of targets.
 
 ```{r myplandrakevig}
+load_basic_example(verbose = FALSE)
 my_plan
 ```
 
-There are [several functions](https://github.com/wlandau-lilly/drake/blob/master/README.md#useful-functions) to help you generate the workflow plan. No need to write every row by hand.
+Wildcard templating generates these data frames at scale.
 
 ```{r plan_drakegeneration}
 library(magrittr)
@@ -70,7 +65,7 @@ whole_plan <- rbind(dataset_plan, analysis_plan)
 whole_plan
 ```
 
-The [visNetwork graph](http://datastorm-open.github.io/visNetwork/) below is interactive, so you can hover, click, drag, zoom, and pan.
+Using static code analysis, `drake` detects the dependencies of all your targets. The result is an interactive network diagram.
 
 ```{r drakevisgraph, eval = FALSE}
 vis_drake_graph(my_plan)
@@ -82,50 +77,76 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-All the targets are out of date because nothing has been built yet.
+At this point, all your targets are out of date because the project is new.
 
 ```{r outdateddrake}
 config <- drake_config(my_plan, verbose = FALSE) # Master configuration list
 outdated(config)
 ```
 
-To build the project, `make()` executes the commands from the workflow plan in the [correct order](https://en.wikipedia.org/wiki/Topological_sorting).
+The `make()` function traverses the network and builds the targets that require updates.
 
 ```{r firstmakedrake}
 make(my_plan)
 ```
 
-Now, everything is up to date. In the next `make()`, there is no work to do.
[…]
