### 2017-11-30 0a4478f
Update the README
based on changes to drake.Rmd
