### 2017-12-01 d190171
Play nicer with tibbles
