### 2017-12-01 baeffa6
Change function name
batchtools_drake_tmpl_file() => drake_batchtools_tmpl_file()
Makes more sense.
