+++ man/drake_batchtools_tmpl_file.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/parallel_ui.R
-\name{batchtools_drake_tmpl_file}
-\alias{batchtools_drake_tmpl_file}
-\title{Function \code{batchtools_drake_tmpl_file}}
+\name{drake_batchtools_tmpl_file}
+\alias{drake_batchtools_tmpl_file}
+\title{Function \code{drake_batchtools_tmpl_file}}
 \usage{
-batchtools_drake_tmpl_file(example = drake::drake_examples(), to = getwd(),
+drake_batchtools_tmpl_file(example = drake::drake_examples(), to = getwd(),
   overwrite = FALSE)
 }
 \arguments{
@@ -32,7 +32,7 @@ load_basic_example() # Load the canonical example.
 # List the drake examples. Only some have template files.
 drake_examples()
 # Write the batchtools template file from the SLURM example.
-batchtools_drake_tmpl_file("slurm") # Writes batchtools.slurm.tmpl.
+drake_batchtools_tmpl_file("slurm") # Writes batchtools.slurm.tmpl.
 # Find batchtools.slurm.tmpl with the rest of the example's files.
 drake_example("slurm") # Writes a new 'slurm' folder with more files.
 # Run the basic example with a
