+++ NAMESPACE
@@ -7,7 +7,6 @@ export(as_drake_filename)
 export(as_file)
 export(available_hash_algos)
 export(backend)
-export(batchtools_drake_tmpl_file)
 export(build_drake_graph)
 export(build_graph)
 export(build_times)
@@ -40,6 +39,7 @@ export(dependency_profile)
 export(deps)
 export(diagnose)
 export(do_prework)
+export(drake_batchtools_tmpl_file)
 export(drake_build)
 export(drake_config)
 export(drake_example)
