### 2017-12-01 20ce533
Improve the writing in debug.Rmd
