### 2017-12-02 8937612
Proof read part of the parallelism vignette
