+++ vignettes/parallelism.Rmd
@@ -26,11 +26,11 @@ knitr::opts_chunk$set(
 )
 ```
 
-`Drake` has extensive high-performance computing support, from local multicore computing on your laptop to serious supercomputing across multiple nodes of a large cluster. In `make()`, just set the `jobs` argument to something greater than 1. That unlocks local multicore parallelism. For large-scale distributed parallelism, set `parallelism` to `"Makefile"` and stay tuned for an explanation.
+`Drake` has extensive high-performance computing support, from local multicore parallelism to serious distributed computing across multiple nodes of a cluster. Control it with the `parallelism` and `jobs` arguments to `make()`, and use `future::plan()` if `parallelism` is `"future_lapply"`.
 
-# The approach
+# The concept
 
-`Drake`'s approach to parallelism relies on the network graph of the targets and imports.
+`Drake`'s approach to parallelism relies on the network graph representation of a project.
 
 ```{r hpcplotgraph, eval = FALSE}
 clean()
@@ -51,21 +51,17 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-When you call `make(my_plan, jobs = 4)`, the work proceeds in chronological order from left to right. The items are built or imported column by column in sequence, and up-to-date targets are skipped. (This is not a hard and fast rule: `drake` includes targets further ahead if some targets in the current column are already up to date.) Within each column, the targets/objects are all independent of each other conditional on the previous steps, so they are distributed over the 4 available parallel jobs/workers. Assuming the targets are rate-limiting (as opposed to imported objects), the next `make(..., jobs = 4)` should be faster than `make(..., jobs = 1)`, but it would be superfluous to use more than 4 jobs.
-
-
+The nodes in each column above are conditionally independent given the dependencies on the left. So in general, the targets and imports are processed column by column from left to right, and everything within a column is executed in parallel. When some targets are already up to date, `drake` searches ahead in the graph to maximize the number of outdated targets in each parallelizable stage.
 
 # How many parallel jobs should you use?
 
 ## Not too many!
 
-Be mindful of the maximum number of simultaneous parallel jobs you deploy. At best, too many jobs is poor etiquette on a system with many users and limited resources. At worst, too many jobs will crash a system. The `jobs` argument to `make()` sets the maximum number of simultaneous jobs in most cases, but not all.
+Be mindful of the maximum number of simultaneous parallel jobs you deploy. Consequences of greed and carelessness range from poor etiquette to system crashes. In most cases, the `jobs` argument to `make()` sets the maximum number of simultaneous jobs, but it does not apply to the parallel execution of targets when `parallelism` is `"future_lapply"`. If you use `"future_lapply"` parallelism, please see the  `workers` argument to most supporting functions passed to `future::plan()` (for example, `future::plan(multisession(workers = 2))`). Depending on the [future](https://github.com/HenrikBengtsson/future) backend you select with `future::plan()` or `future::plan()`, you might also make use of one of the other environment variables listed in `?future::future.options`.
 
[…]
