### 2017-12-02 c39fb2a
Proofread the rest of the parallelism vignette
