+++ vignettes/drake.Rmd
@@ -50,7 +50,7 @@ dataset_plan <- plan_drake(
 dataset_plan
 
 analysis_methods <- plan_drake(
-  regression = regNUMBER(dataset__)
+  regression = regNUMBER(dataset__) # nolint
 ) %>%
   evaluate_plan(wildcard = "NUMBER", values = 1:2)
 analysis_methods
@@ -168,7 +168,7 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-Within each column above, the nodes are conditionally independent given their dependencies. Each `make()` walks through the columns from left to right and applies parallel processing within each column. If any nodes are already up to date, `drake` looks downstream to maximize the number of outdated targets in a parallelizable stage.
+Within each column above, the nodes are conditionally independent given their dependencies. Each `make()` walks through the columns from left to right and applies parallel processing within each column. If any nodes are already up to date, `drake` looks downstream to maximize the number of outdated targets in a parallelizable stage. To show the parallelizable stages of the next `make()` programmatically, use the `parallel_stages()` function.
 
 # Acknowledgements and related work
 
