+++ vignettes/parallelism.Rmd
@@ -32,6 +32,17 @@ knitr::opts_chunk$set(
 
 `Drake`'s approach to parallelism relies on the network graph representation of a project.
 
+```{r hiddenhpcpreplotgraph, eval = TRUE, echo = FALSE, message = FALSE}
+clean()
+load_basic_example()
+config <- make(my_plan, jobs = 2, verbose = FALSE) # Parallelize over 2 jobs.
+# Change a dependency.
+reg2 <- function(d) {
+  d$x3 <- d$x ^ 3
+  lm(y ~ x3, data = d)
+}
+```
+
 ```{r hpcplotgraph, eval = FALSE}
 clean()
 load_basic_example()
@@ -52,6 +63,12 @@ style="border: none; box-shadow: none">
 </iframe>
 
 The nodes in each column above are conditionally independent given the dependencies on the left. So in general, the targets and imports are processed column by column from left to right, and everything within a column is executed in parallel. When some targets are already up to date, `drake` searches ahead in the graph to maximize the number of outdated targets in each parallelizable stage.
+
+To show the parallelizable stages of the next `make()` programmatically, use the `parallel_stages()` function. All the targets/imports in a stage are processed in parallel before moving on to the next stage.
+
+```{r parallelstages}
+parallel_stages(config)
+```
 
 # How many parallel jobs should you use?
 
@@ -109,7 +126,7 @@ default_parallelism() # "parLapply" on Windows, "mclapply" everywhere else
 
 ## mclapply
 
-The `mclapply` backend is powered by the `mclapply()` function from the `parallel` package. It is a way to fork multiple processes on your local machine to take advantage of multicore computing. It spins up quickly, but it lacks scalability, and it does not work on Windows. If you try to call `make(.., parallelism = "mclapply", jobs = 2)` on a Windows machine, `drake` will warn you and then demote the number of jobs to 1.
+The `mclapply` backend is powered by the `mclapply()` function from the `parallel` package, and it forks multiple processes on your local machine. It spins up quickly, but it lacks scalability, and it does not work on Windows. If you try to call `make(.., parallelism = "mclapply", jobs = 2)` on a Windows machine, `drake` will warn you and then demote `jobs` to 1.
 
 ## parLapply
 
@@ -117,7 +134,7 @@ The `mclapply` backend is powered by the `mclapply()` function from the `paralle
 make(.., parallelism = "mclapply", jobs = 2)
 ```
 
-The `parLapply` backend is powered by the `parLapply()` function from the `parallel` package. Like the `mclapply` backend, `parLapply` only scales up to a handful of jobs on your local machine. However, it works on all platforms. The tradeoff is overhead. `parLapply` is fast once it gets going, but it takes a long time to set up because each call to `make()` creates a new parallel socket cluster and transfers all you data and session info to each parallel thread individually. So if `jobs` is less than 2, `make()` does not bother setting up a cluster, and it uses `lapply()` instead. More importantly, the default parallel backend is `parLapply` on Windows machines and `mclapply` everywhere else. 
+The `parLapply` backend is powered by the `parLapply()` function from the `parallel` package. Like the `mclapply` backend, `parLapply` only scales up to a handful of jobs on your local machine. `parLapply` parallelism works on all platforms, but it takes a few seconds to initialize during each `make()`. If `jobs` is less than 2, `make()` does not bother setting up a parallel socket cluster, opting instead for `lapply()` to reduce overhead. The default parallel backend is `parLapply` on Windows machines and `mclapply` everywhere else. 
 
 ```{r hpcparLapply, eval = FALSE}
[…]
