### 2017-12-03 3abbb1c
Proofread the caution vignette
