+++ vignettes/caution.Rmd
@@ -27,18 +27,18 @@ knitr::opts_chunk$set(
 )
 ```
 
-This vignette addresses `drake`'s known edge cases, pitfalls, and weaknesses that may or may not be fixed in future releases. For the most up-to-date information on unhandled edge cases, please visit the [issue tracker](https://github.com/wlandau-lilly/drake/issues), where you can submit your own bug reports as well. Be sure to search the closed issues too, especially if you are not using the most up-to-date development version. For a guide to debugging and testing `drake` projects, please refer to the separate ["debug" vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/debug.Rmd).
+This vignette addresses `drake`'s known edge cases, pitfalls, and weaknesses that might not be fixed in future releases. For the most up-to-date information on unhandled edge cases, please visit the [issue tracker](https://github.com/wlandau-lilly/drake/issues), where you can submit your own bug reports as well. Be sure to search the closed issues too, especially if you are not using the most up-to-date development version of `drake`. For a guide to debugging and testing `drake` projects, please refer to the separate ["debug" vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/debug.Rmd).
 
 # Projects built with drake <= 4.4.0 are not back compatible with drake > 4.4.0.
 
-Versions `drake` after 4.4.0 have different caching internals that increase speed and open up possibilities for future development. However, if you have a project built with version 4.4.0 or earlier, a later version of `drake` will think all your targets are out of date. So by design, a simple `make()` will quit in error to stop you from doing any damage. To migrate your project to a later `drake`, you have two options.
+Versions of `drake` after 4.4.0 have different caching internals. That means if you built your project with `drake` 4.4.0 or earlier, later versions of `drake` will not let you call `make()`, which could mangle your work. To migrate your project to a later `drake`, you have three options.
 
 1. Revert to a back-compatible version of drake with `devtools::install_version("drake", "4.4.0")`. Here, the [devtools package](https://github.com/hadley/devtools) must be installed (`install.packages("devtools")`).
-1. Run your project from scratch with `make(..., force = TRUE)`
+1. Run your project from scratch with `make(..., force = TRUE)`.
 1. Use `migrate_drake_project()` to convert your project to the new format. The `migrate_drake_project()` function
     1. copies your old project to a backup folder.
-    1. converts your project's cache to a format compatible with 5.0.0.
-    1. informs you if the migration succeeded: that is, if outdated targets remained outdated and up to date targets remained up to date in the conversion.
+    1. converts your project's cache to a format compatible with `drake` 5.0.0 and later.
+    1. informs you whether the migration succeeded: that is, whether outdated targets remained outdated and up-to-date targets remain up to date after migration.
 
 # Workflow plans
 
@@ -98,7 +98,7 @@ Put the workflow plan together.
 rbind(datasets, files)
 ```
 
-For more control over target names in cases like this, you may want to use the [wildcard package](https://github.com/wlandau/wildcard).
+For finer control over target names in cases like this, you may want to use the [wildcard package](https://github.com/wlandau/wildcard).
 
 ## Commands are NOT perfectly flexible.
 
@@ -111,22 +111,22 @@ plan_drake(
 )
 ```
 
[…]
