### 2017-12-03 0ebca35
Proofread the graph.Rmd vignette
