### 2017-12-03 d82d0ff
Proofread the quickstart vignette
