+++ vignettes/quickstart.Rmd
@@ -53,11 +53,11 @@ Dive deeper into the built-in examples.
 
 ```{r noeval2, eval = FALSE}
 drake_example("basic") # Write the code files.
-drake_examples() # List the other examples.
+drake_examples()       # List the other examples.
 vignette("quickstart") # This vignette
 ```
 
-# Setting up the basic example
+# Set up the basic example
 
 Let's establish the building blocks of a data analysis workflow.
 
@@ -90,7 +90,7 @@ reg2 <- function(d){
 }
 ```
 
-We need the source file `report.Rmd` for the end.k
+We need the source file `report.Rmd` for the final report.
 
 ```{r file}
 lines <- c(
@@ -118,7 +118,7 @@ lines <- c(
 writeLines(lines, "report.Rmd")
 ```
 
-# Workflow plan
+# The workflow plan data frame
 
 The workflow plan lists the intermediate steps of your project.
 
@@ -127,12 +127,12 @@ load_basic_example()
 my_plan
 ```
 
-Each row is an intermediate step, and each **command** generates a **target**. A target is an output R object (cached when generated) or output file (specified with single quotes), and a command just an ordinary piece of R code (not necessarily a single function call). As input, commands may take objects imported from your workspace, targets generated by other commands, or initial input files. These dependencies give your project an underlying network.
+Each row is an intermediate step, and each **command** generates a single **target**. A target is an output R object (cached when generated) or an output file (specified with single quotes), and a command just an ordinary piece of R code (not necessarily a single function call). Commands make use of R objects imported from your workspace, targets generated by other commands, and initial input files. These dependencies give your project an underlying network representation.
 
 ```{r graph1quick, eval = FALSE}
 # Hover, click, drag, zoom, and pan.
 config <- drake_config(my_plan)
-vis_drake_graph(config, width = "100%", height = "500px")
+vis_drake_graph(config, width = "100%", height = "500px") # Also drake_graph()
 ```
 
 <iframe
@@ -141,9 +141,7 @@ width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
 
-Also see the alias `drake_graph()`. Functions `dataframes_graph()`, `render_drake_graph()`, and `drake_config()` allow for faster and more customized regraphing.
-
-You can also check the dependencies of individual targets.
+You can also check the dependencies of individual targets and imported functions.
 
 ```{r checkdeps}
 deps(reg2)
@@ -153,7 +151,7 @@ deps(my_plan$command[1]) # Files like report.Rmd are single-quoted.
 deps(my_plan$command[nrow(my_plan)])
 ```
 
-List all the reproducibly-tracked objects and files, including imports and targets.
+List all the reproducibly-tracked objects and files.
 
 ```{r tracked}
 tracked(my_plan, targets = "small")
@@ -161,7 +159,7 @@ tracked(my_plan, targets = "small")
 tracked(my_plan)
 ```
 
-Check for cycles, missing input files, and other pitfalls.
+Check for circular reasoning, missing input files, and other pitfalls.
 
 ```{r check}
 check_plan(my_plan)
@@ -169,7 +167,7 @@ check_plan(my_plan)
 
 # Generate the workflow plan
 
-The data frame `my_plan` would be a pain to write by hand, so `drake` has functions to help you.
+The workflow plan data frame `my_plan` would be a pain to write by hand, so `drake` has functions to help you.
 
 ```{r datasets}
 my_datasets <- plan_drake(
@@ -213,9 +211,9 @@ results <- plan_summaries(summary_types, analyses = my_analyses,
 results
 ```
 
-The `gather` feature groups summaries into a smaller number of more manageable targets. I shut it off here to make the data frames more readable.
[…]
