### 2017-12-03 8eede73
Proofread the storage vignette
