+++ vignettes/storage.Rmd
@@ -35,12 +35,12 @@ knitr::opts_chunk$set(
 
 # Storage basics
 
-When you run `make()`, `drake` puts your imports and output targets in a hidden cache, or storage location.
+When you run `make()`, `drake` stores your imports and output targets in a hidden cache.
 
 
 ```{r basic_storage}
 library(drake)
-load_basic_example()
+load_basic_example(verbose = FALSE)
 config <- make(my_plan, verbose = FALSE)
 ```
 
@@ -69,14 +69,14 @@ cache <- get_cache() # Get the default cache from the last build.
 
 class(cache)
 
-cache$list() # functionality from storr
+cache$list() # Functionality from storr
 
-cache$get("small") # functionality from storr
+cache$get("small") # Functionality from storr
 ```
 
 # Hash algorithms
 
-The key to [storr](https://github.com/richfitz/storr)'s internals is the concept of  [hashing](https://en.wikipedia.org/wiki/Hash_function). [Storr](https://github.com/richfitz/storr) uses hashes to label what they store, and `drake` leverages these hashes to figure out what is up to date and what needs to be (re)built. A hash is like a fingerprint for a piece of data, so the hash should change if the dataset changes. Regardless of the data's size, the hash always has same number of characters.
+The concept of [hashing](https://en.wikipedia.org/wiki/Hash_function) is central to [storr](https://github.com/richfitz/storr)'s internals. [Storr](https://github.com/richfitz/storr) uses hashes to label stored objects, and `drake` leverages these hashes to figure out which targets are up to date and which ones are outdated. A hash is like a target's fingerprint, so the hash changes when the target changes. Regardless of the target's size, the hash is always the same number of characters.
 
 ```{r hashes}
 library(digest) # package for hashing objects and files
@@ -88,7 +88,7 @@ digest(smaller_data) # compute the hash
 digest(larger_data)
 ```
 
-However, different hash algorithms vary in output length.
+However, different hash algorithms vary in length.
 
 ```{r compare_algo_lengths}
 digest(larger_data, algo = "sha512")
@@ -102,10 +102,10 @@ digest(larger_data, algo = "murmur32")
 
 # Which hash algorithm should you choose?
 
-Hashing is expensive, and unsurprisingly, shorter hashes are usually faster to compute. So why not always use `murmur32`? One reason is the risk of collisions: when two different objects have the same hash. In general, shorter hashes have higher risks of collisions. We want our fingerprints to be unique. On the other hand, a longer hash is not always the answer. Besides speed, the decision depends on how we use the output. `Drake` and [storr](https://github.com/richfitz/storr) both use hash keys as names for internal cache files, and in general, file names should respect the 260-character cap on Windows file paths. That is why `drake` uses a shorter hash algorithm for internal cache-related file names and a longer hash algorithm for everything else.
+Hashing is expensive, and unsurprisingly, shorter hashes are usually faster to compute. So why not always use `murmur32`? One reason is the risk of collisions: that is, when two different objects have the same hash. In general, shorter hashes have more frequent collisions. On the other hand, a longer hash is not always the answer. Besides the loss of speed, `drake` and [storr](https://github.com/richfitz/storr) sometimes use hash keys as file names, and long hashes could violate the 260-character cap on Windows file paths. That is why `drake` uses a shorter hash algorithm for internal cache-related file names and a longer hash algorithm for everything else.
 
 ```{r justified_hash_choices}
-default_short_hash_algo() # for drake
+default_short_hash_algo()
 
 default_long_hash_algo()
 
@@ -116,7 +116,7 @@ long_hash(cache)
 
 # Select the hash algorithms of the default cache
 
[…]
