### 2017-12-03 457375d
Proofread the timing vignette
and factor storage operations back into timing.
