### 2017-12-05 66687b5
Belatedly repair devtools::run_examples(run = FALSE)
