+++ R/knitr.R
@@ -22,7 +22,8 @@
 #' # Find the dependencies of the compiled output target, 'report.md'.
 #' knitr_deps("report.Rmd")
 #' make(my_plan) # Run the project.
-#' knitr_deps("'report.md'") # Work on the Rmd source, not the output.
+#' # Work only on the Rmd source, not the output.
+#' # knitr_deps("'report.md'") # error # nolint
 #' }
 knitr_deps <- function(target){
   if (!length(target)){
