+++ R/cache.R
@@ -9,6 +9,7 @@
 #' you want to know
 #' @examples
 #' \dontrun{
+#' clean(destroy = TRUE)
 #' # Get/create a new drake/storr cache.
 #' cache <- recover_cache()
 #' # Show the file path of the cache.
@@ -47,6 +48,7 @@ cache_path <- function(cache = NULL){
 #' running version of drake.
 #' @examples
 #' \dontrun{
+#' clean(destroy = TRUE)
 #' # No cache is available.
 #' get_cache() # NULL
 #' load_basic_example() # Load drake's canonical example.
@@ -83,7 +85,8 @@ get_cache <- function(
 #' @param verbose, whether to print the file path of the cache.
 #' @examples
 #' \dontrun{
-#' x <- this_cache() # The cache does not exist yet.
+#' clean(destroy = TRUE)
+#' try(x <- this_cache(), silent = FALSE) # The cache does not exist yet.
 #' load_basic_example() # Load drake's canonical example.
 #' make(my_plan) # Run the project, build the targets.
 #' y <- this_cache() # Now, there is a cache.
@@ -136,8 +139,11 @@ this_cache <- function(
 #' @param ... other arguments to the cache constructor
 #' @examples
 #' \dontrun{
+#' clean(destroy = TRUE)
+#' unlink("not_hidden", recursive = TRUE)
 #' cache1 <- new_cache() # Creates a new hidden '.drake' folder.
 #' cache2 <- new_cache(path = "not_hidden", short_hash_algo = "md5")
+#' clean(destroy = TRUE, cache = cache2)
 #' }
 #' # Create a storr_environment() cache from a custom environment.
 #' e <- new.env() # Create a new environment.
@@ -202,6 +208,7 @@ new_cache <- function(
 #' @param verbose logical, whether to print the file path of the cache.
 #' @examples
 #' \dontrun{
+#' clean(destroy = TRUE)
 #' load_basic_example() # Load drake's canonical example.
 #' make(my_plan) # Run the project, build all the targets.
 #' x <- recover_cache(".drake") # Recover the project's storr cache.
@@ -271,6 +278,7 @@ default_cache_path <- function(){
 #'
 #' @examples
 #' \dontrun{
+#' clean(destroy = TRUE)
 #' load_basic_example() # Load drake's canonical example into the workspace.
 #' config <- make(my_plan) # Run the project, build all the targets.
 #' # Locate the drake/storr cache of the project
