### 2017-12-05 21dcdf0
Add more @return tags
