### 2017-12-05 #156 @wlandau-lilly
@jules32 and @benmarwick, could we touch base about timing? [Drake](https://github.com/wlandau-lilly/drake) is large and developing fast, so I do understand that reviews may be more difficult than is typical.
