### 2017-12-05 6175387
Wrap each example in a call to test_with_dir()
In run_examples("drake"), if the \dontrun{} blocks are run,
then drake wreaked havoc on the user's workspace and file system.
Now, every test with side effects is wrapped in a call to
test_with_dir() to quarantine the side effects. The users
can still run the examples interactively, as they should.
But now, if all the examples are run in batch mode with
\dontrun{} ignored, the right things are protected.
