+++ R/build.R
@@ -11,6 +11,7 @@
 #' @param config internal configuration list
 #' @examples
 #' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
 #' # This example is not really a user-side demonstration.
 #' # It just walks through a dive into the internals.
 #' # Populate your workspace and write 'report.Rmd'.
@@ -27,6 +28,7 @@
 #' # Should now include 'small'
 #' cached()
 #' readd(small)
+#' })
 #' }
 drake_build <- function(target, meta, config){
   config$hook(
