### 2017-12-06 #156 @jules32
Hi @wlandau-lilly et al,

I am going to get started on Thursday and this weekend since I was out of
the office last week. Looking forward to getting to know this package!

On Tue, Dec 5, 2017 at 7:53 AM, Maëlle Salmon <notifications@github.com>
wrote:

> I forgot to update the thread when @jules32 <https://github.com/jules32>
> contacted me to say she'd get the review in before Dec the 11th, sorry.
>
> @benmarwick <https://github.com/benmarwick> any update?
>
> Thanks to both reviewers and thanks @wlandau-lilly
> <https://github.com/wlandau-lilly> for your patience. :-)
>
