+++ man/time_stamps.Rd
@@ -28,6 +28,6 @@ config <- drake_config(my_plan) # Master internal configuration list
 time_stamps(config)
 # Now look in '.drake/ts' for dummy timestamp files.
 })
-} 
+}
 }
 \keyword{internal}
