### 2017-12-06 5f2cbb5
Try @jimhester's suggestion in travis issue 8864
travis-ci/travis-ci#8864
https://github.com/travis-ci/travis-ci/issues/8864#issuecomment-349684435
