### 2017-12-06 31e88a3
Use stable R in appveyor
