+++ .travis.yml
@@ -1,12 +1,12 @@
 # R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
-language: r
-r:
-  - release
+language: R
 sudo: false
 cache: packages
 
 # Added manually by Will Landau
+
+warnings_are_errors: true
 
 r_packages:
   - covr
