+++ appveyor.yml
@@ -40,3 +40,10 @@ artifacts:
 
   - path: '\*_*.zip'
     name: Bits
+
+environment:
+  global:
+    WARNINGS_ARE_ERRORS: 1
+
+  matrix:
+    - R_VERSION: stable
