### 2017-12-06 5c3e225
Change appveyor to R-release
