+++ appveyor.yml
@@ -46,4 +46,4 @@ environment:
     WARNINGS_ARE_ERRORS: 1
 
   matrix:
-    - R_VERSION: devel
+    - R_VERSION: release
