### 2017-12-06 ee0e4a6
Add option to suppress sessionInfo()
Rate-limiting step for small make()s.
Trying to speed up tests.
