### 2017-12-06 f804fa1
Eliminate another call to sessionInfo()
Should speed things up.
