### 2017-12-07 7eae23f
Replace some Filter()s with lightly_parallelize()s
Should speed up larger projects.
