### 2017-12-07 11b8467
Replace more lapply()s with lightly_parallelize()s
