### 2017-12-07 a61350b
Replace Filter()s with parallel_filter()s
