+++ R/check.R
@@ -75,9 +75,9 @@ assert_standard_columns <- function(config){
 missing_input_files <- function(config) {
   missing_files <- V(config$graph)$name %>%
     setdiff(y = config$plan$target) %>%
-    Filter(f = is_file) %>%
+    parallel_filter(f = is_file, jobs = config$jobs) %>%
     drake_unquote %>%
-    Filter(f = function(x) !file.exists(x))
+    parallel_filter(f = function(x) !file.exists(x), jobs = config$jobs)
   if (length(missing_files))
     warning("missing input files:\n", multiline_message(missing_files),
       call. = FALSE)

[…]
