### 2017-12-07 9db3765
Lightly parallelize instead of Vectorize in places
