### 2017-12-07 d84f15c
Lightly parallelize assign_to_envir()
