### 2017-12-07 43c19f8
Spell check README.md
I thought devtools::spell_check() covered that already...
