### 2017-12-08 b88aae2
Standardize the titles of the help files
