### 2017-12-08 d7ad0a4
Remove migrate message
