+++ R/migrate.R
@@ -45,10 +45,6 @@
 migrate_drake_project <- function(
   path = drake::default_cache_path(), jobs = 1
 ){
-  message(
-    "Note: drake project migration will be removed ",
-    "by the year 2020, probably by the year 2019."
-  )
   cache <- should_migrate(path = path)
   if (is.null(cache)){
     return(invisible(TRUE))
