### 2017-12-08 4852cd8
Boost speed a little
Replace some calls to cache$list() with
lightly_parallelize(), parallel_filter(),
and cache$exists().
