+++ R/package.R
@@ -1,5 +1,5 @@
 #' Drake is a pipeline toolkit
-#' (\url{https://github.com/pditommaso/awesome-pipeline}) 
+#' (\url{https://github.com/pditommaso/awesome-pipeline})
 #' and a scalable, R-focused solution for reproducibility
 #' and high-performance computing.
 #' @docType package
