### 2017-12-08 982154d
Fix README typo
