### 2017-12-08 fa188c2
Begin preparing for true arbitrary storr cache support
