+++ R/distributed.R
@@ -16,12 +16,18 @@
 #' })
 #' }
 prepare_distributed <- function(config){
-  this_cache_path <- cache_path(config$cache)
+  if (!file.exists(config$cache_path)){
+    dir.create(config$cache_path)
+  }
+  saveRDS(
+    config$fetch_cache,
+    file = file.path(config$cache_path, "fetch_cache.rds")
+  )
   if (identical(globalenv(), config$envir)){
     save(
       list = ls(config$envir, all.names = TRUE),
       envir = config$envir,
-      file = globalenv_file(this_cache_path)
+      file = globalenv_file(config$cache_path)
     )
   }
   config$cache$set(key = "envir", value = config$envir, namespace = "config")
@@ -58,7 +64,11 @@ build_distributed <- function(target, meta_list, cache_path){
 }
 
 recover_drake_config <- function(cache_path){
-  cache <- this_cache(cache_path, verbose = FALSE)
+  fetch_cache <- tryCatch(
+    readRDS(file.path(cache_path, "fetch_cache.rds")),
+    error = error_null
+  )
+  cache <- this_cache(cache_path, verbose = FALSE, fetch_cache = fetch_cache)
   config <- read_drake_config(cache = cache)
   if (identical(globalenv(), config$envir)){
     dir <- cache_path(cache = cache)
