+++ R/cache.R
@@ -16,7 +16,7 @@
 #' # Show the file path of the cache.
 #' cache_path(cache = cache)
 #' # In-memory caches do not have file paths.
-#' mem <- new_cache(type = "storr_environment") # or just storr_environment()
+#' mem <- storr_environment()
 #' cache_path(cache = mem)
 #' })
 #' }
@@ -65,14 +65,18 @@ get_cache <- function(
   path = getwd(),
   search = TRUE,
   verbose = 1,
-  force = FALSE
+  force = FALSE,
+  fetch_cache = NULL
 ){
   if (search){
     path <- find_cache(path = path)
   } else {
     path <- default_cache_path()
   }
-  this_cache(path = path, force = force, verbose = verbose)
+  this_cache(
+    path = path, force = force, verbose = verbose,
+    fetch_cache = fetch_cache
+  )
 }
 
 #' @title Function \code{this_cache}
@@ -101,19 +105,23 @@ get_cache <- function(
 #' })
 #' }
 this_cache <- function(
-  path = drake::default_cache_path(), force = FALSE, verbose = TRUE
+  path = drake::default_cache_path(), force = FALSE, verbose = TRUE,
+  fetch_cache = NULL
 ){
   if (is.null(path) || !file.exists(path)){
     return(NULL)
   }
   console_cache(path = path, verbose = verbose)
-  type <- type_of_cache(path)
-  if (is.null(type)) {
-    type <- default_cache_type()
+  if (is.null(fetch_cache)){
+    cache <- drake_fetch_rds(path)
+  } else {
+    cache <- eval(parse(text = fetch_cache))
   }
-  cache_fetcher <- paste0("get_", type, "_cache")
-  cache <- get(cache_fetcher, envir = getNamespace("drake"))(
-    path = path
+  configure_cache(
+    cache = cache,
+    long_hash_algo = "md5",
+    overwrite_hash_algos = FALSE,
+    clear_progress = FALSE
   )
   if (!force){
     assert_compatible_cache(cache = cache)
@@ -121,21 +129,30 @@ this_cache <- function(
   cache
 }
 
+#' @export
+drake_fetch_rds <- function(path){
+  if (!file.exists(path)) {
+    return(NULL)
+  }
+  hash_algo_file <- file.path(path, "config", "hash_algorithm")
+  hash_algo <- scan(hash_algo_file, quiet = TRUE, what = character())
+  storr::storr_rds(
+    path = path,
+    mangle_key = TRUE,
+    hash_algorithm = hash_algo
+  )
+}
+
 #' @title Function \code{new_cache}
-#' @description Make a new drake cache. Could be any
-#' type of cache in \code{\link{cache_types}()}.
+#' @description Make a new \code{storr_rds()} \code{drake} cache.
 #' @export
 #' @return A newly created drake cache as a storr object.
 #' @seealso \code{\link{default_short_hash_algo}},
 #' \code{\link{default_long_hash_algo}},
-#' \code{\link{make}}, \code{\link{cache_types}},
-#' \code{\link{in_memory_cache_types}}
+#' \code{\link{make}}
 #' @param path file path to the cache if the cache
 #' is a file system cache.
 #' @param verbose logical, whether to print out the path of the cache.
-#' @param type character scalar, type of the drake cache.
-#' Must be among the list of supported caches
-#' in \code{\link{cache_types}()}.
 #' @param short_hash_algo short hash algorithm for the cache.
 #' See \code{\link{default_short_hash_algo}()} and
 #' \code{\link{make}()}
@@ -152,32 +169,25 @@ this_cache <- function(
 #' cache2 <- new_cache(path = "not_hidden", short_hash_algo = "md5")
 #' clean(destroy = TRUE, cache = cache2)
 #' }
[…]
