+++ R/make.R
@@ -289,6 +289,7 @@ make <- function(
   verbose = 1,
   hook = default_hook,
   cache = drake::get_cache(verbose = verbose, force = force),
+  fetch_cache = NULL,
   parallelism = drake::default_parallelism(),
   jobs = 1,
   packages = rev(.packages()),
@@ -349,6 +350,7 @@ make <- function(
       recipe_command = recipe_command,
       clear_progress = TRUE,
       cache = cache,
+      fetch_cache = fetch_cache,
       timeout = timeout,
       cpu = cpu,
       elapsed = elapsed,
