### 2017-12-08 d8a3acc
Get ready for RSQLite cache tests
