+++ man/make.Rd
@@ -76,6 +76,15 @@ and \code{function(code){message("Avoiding the code")}} is a bad hook.}
 See also \code{\link{get_cache}()}, \code{\link{this_cache}()},
 and \code{\link{recover_cache}()}}
 
+\item{fetch_cache}{character vector containing lines of code.
+The purpose of this code is to fetch the \code{storr} cache
+with a command like \code{storr_rds()} or \code{storr_dbi()},
+but customized.
+This is necessary if you are using both custom caches
+and distributed parallelism (\code{parallelism = "future_lapply"}
+or \code{"Makefile"}) because the distributed R sessions
+need to know how to load the cache.}
+
 \item{parallelism}{character, type of parallelism to use.
 To list the options, call \code{\link{parallelism_choices}()}.
 For detailed explanations, see \code{?\link{parallelism_choices}},
