+++ NAMESPACE
@@ -42,7 +42,6 @@ export(drake_build)
 export(drake_config)
 export(drake_example)
 export(drake_examples)
-export(drake_fetch_rds)
 export(drake_gc)
 export(drake_graph)
 export(drake_palette)
