### 2017-12-08 0c08f2a
Write files for dbi cache test
