### 2017-12-08 29f0f76
Sketch a test of the dbi cache
Having problems with single-quoted keys...
