+++ tests/testthat/test-dbi-cache.R
@@ -1,110 +1,37 @@
-drake_context("memory cache")
+drake_context("dbi cache")
 
-test_with_dir("storr_environment is usable", {
-  x <- storr_environment(hash_algorithm = "murmur32") %>%
+test_with_dir("storr_dbi is usable", {
+  skip_if_not_installed("DBI")
+  skip_if_not_installed("RSQLite")
+
+  mydb <- DBI::dbConnect(RSQLite::SQLite(), "my-db.sqlite")
+  cache <- storr_dbi(
+    "dattbl", "keystbl", con = mydb, hash_algorithm = "murmur32") %>%
     configure_cache(
       long_hash_algo = "sha1",
       overwrite_hash_algos = TRUE
     )
+
+  scenario <- get_testing_scenario()
+  e <- eval(parse(text = scenario$envir))
+  jobs <- scenario$jobs
+  parallelism <- scenario$parallelism
+  
+  con <- load_basic_example(envir = e)
+  con$cache$destroy()
+  my_plan <- e$my_plan
+  config <- drake_config(
+    my_plan, envir = e,
+    jobs = jobs, parallelism = parallelism,
+    cache = cache
+  )
+  testrun(config)
+  
   expect_false(file.exists(default_cache_path()))
-  expect_equal(short_hash(x), "murmur32")
-  expect_equal(long_hash(x), "sha1")
-  expect_error(drake_session(cache = x))
-  pln <- plan_drake(y = 1)
-  config <- make(pln, cache = x, verbose = FALSE, session_info = FALSE)
-  expect_equal(cached(cache = x), "y")
+  expect_equal(short_hash(cache), "murmur32")
+  expect_equal(long_hash(cache), "sha1")
+  expect_equal(sort(built(cache = cache)), sort(my_plan$target))
   expect_false(file.exists(default_cache_path()))
   expect_equal(outdated(config), character(0))
-  expect_false(file.exists(default_cache_path()))
-})
-
-test_with_dir("arbitrary storr in-memory cache", {
-  expect_false(file.exists(default_cache_path()))
-  parallelism <- default_parallelism()
-  jobs <- 1
-  envir <- eval(parse(text = get_testing_scenario()$envir))
-  cache <- storr::storr_environment(hash_algorithm = "murmur32")
-  load_basic_example(envir = envir, cache = cache)
-  my_plan <- envir$my_plan
-  con <- make(
-    my_plan,
-    envir = envir,
-    cache = cache,
-    parallelism = parallelism,
-    jobs = jobs,
-    verbose = FALSE
-  )
-  envir$reg2 <- function(d){
-    d$x3 <- d$x ^ 3
-    lm(y ~ x3, data = d)
-  }
-  expect_false(file.exists(default_cache_path()))
-  expect_equal(short_hash(con$cache), "murmur32")
-  expect_equal(long_hash(con$cache), default_long_hash_algo())
-
-  x <- predict_runtime(con)
-  y <- rate_limiting_times(con, from_scratch = TRUE)
-  expect_true(length(x) > 0)
-  expect_true(nrow(y) > 0)
-
-  expect_equal(cached(verbose = FALSE), character(0))
-  targets <- my_plan$target
-  expect_true(all(targets %in% cached(cache = cache, verbose = FALSE)))
-  expect_false(file.exists(default_cache_path()))
-
-  expect_error(drake_session(verbose = FALSE))
-  expect_true(is.list(drake_session(cache = cache, verbose = FALSE)))
-  expect_false(file.exists(default_cache_path()))
-
-  expect_equal(length(imported(verbose = FALSE)), 0)
-  expect_true(length(imported(cache = cache, verbose = FALSE)) > 0)
[…]
