+++ tests/testthat/test-basic.R
@@ -5,12 +5,11 @@ test_with_dir("basic example works", {
   e <- eval(parse(text = scenario$envir))
   jobs <- scenario$jobs
   parallelism <- scenario$parallelism
-
+  
   load_basic_example(envir = e)
   my_plan <- e$my_plan
   config <- drake_config(my_plan, envir = e,
-    jobs = jobs, parallelism = parallelism,
-    verbose = FALSE)
+    jobs = jobs, parallelism = parallelism)
 
   expect_true(is.list(dependency_profile(
     target = "'report.md'", config = config)))
