+++ DESCRIPTION
@@ -50,12 +50,14 @@ Imports:
   visNetwork,
   withr
 Suggests: 
-    abind,
-    future.batchtools,
-    MASS,
-    methods,
-    rmarkdown,
-    tibble
+  abind,
+  DBI,
+  future.batchtools,
+  MASS,
+  methods,
+  RSQLite,
+  rmarkdown,
+  tibble
 VignetteBuilder: knitr
 URL: https://github.com/wlandau-lilly/drake
 BugReports: https://github.com/wlandau-lilly/drake/issues
