### 2017-12-09 5d457fb
Dumb down test-dbi-cache.R
Does not work with any kind of parallelism...
