+++ tests/testthat/test-basic.R
@@ -5,7 +5,7 @@ test_with_dir("basic example works", {
   e <- eval(parse(text = scenario$envir))
   jobs <- scenario$jobs
   parallelism <- scenario$parallelism
-  
+
   load_basic_example(envir = e)
   my_plan <- e$my_plan
   config <- drake_config(my_plan, envir = e,
