### 2017-12-09 0cdbfbf
Explain custom storr caches and pitfalls
