+++ man/get_cache.Rd
@@ -27,7 +27,7 @@ running version of drake.}
 \item{fetch_cache}{character vector containing lines of code.
 The purpose of this code is to fetch the \code{storr} cache
 with a command like \code{storr_rds()} or \code{storr_dbi()},
-but customized.}
+but customized. This feature is experimental.}
 }
 \value{
 A drake/storr cache, if available. \code{NULL} otherwise.
