### 2017-12-09 922ceac
Try to attenuate verbosity in Makefile testing scenario
