+++ tests/testthat/test-future.R
@@ -5,19 +5,17 @@ test_with_dir("future package functionality", {
   scenario <- get_testing_scenario()
   e <- eval(parse(text = scenario$envir))
   load_basic_example(envir = e)
-  expect_silent(
-    withr::with_options(
-      new = list(mc.cores = 2), code = {
-        config <- make(
-          e$my_plan,
-          envir = e,
-          parallelism = "future_lapply",
-          jobs = 1,
-          verbose = FALSE,
-          session_info = FALSE
-        )
-      }
-    )
+  withr::with_options(
+    new = list(mc.cores = 2), code = {
+      config <- make(
+        e$my_plan,
+        envir = e,
+        parallelism = "future_lapply",
+        jobs = 1,
+        verbose = FALSE,
+        session_info = FALSE
+      )
+    }
   )
   expect_equal(
     outdated(config),
