+++ inst/examples/basic/report.Rmd
@@ -4,13 +4,9 @@ author: You
 output: html_document
 ---
 
-Look how I read outputs from the drake cache."
-Drake notices that `small`, `coef_regression2_small`,
-and `large` are dependencies of the
-future compiled output report file target, `report.md`.
-Just be sure that the workflow plan command for the target `'report.md'`
-has an explicit call to `knit()`, something like `knit('report.Rmd')` or
-`knitr::knit(input = 'report.Rmd', quiet = TRUE)`.
+In the basic example,
+targets `small`, `coef_regression2_small`, and `large`
+are dependencies of the knitted output, `report.md`.
 
 ```{r example_chunk}
 library(drake)
