### 2017-12-09 6bd4cf2
Fix winbuilder note about URLs
