### 2017-12-09 76b246d
Move CONTRIBUTING.md to the top level
Supports ropensci/onboarding#156.
