### 2017-12-09 60f37e1
Improve statement of need at the beginning of the README
