+++ README.md
@@ -1,3 +1,5 @@
+# Data frames in R for [Make](http://kbroman.org/minimal_make/)
+
 <p align="center">
   <img src="./images/logo-readme.png" alt="">
 </p>
@@ -12,12 +14,100 @@
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/wlandau-lilly/drake?branch=master&svg=true)](https://ci.appveyor.com/project/wlandau-lilly/drake)
 [![codecov.io](https://codecov.io/github/wlandau-lilly/drake/coverage.svg?branch=master)](https://codecov.io/github/wlandau-lilly/drake?branch=master)
 
-# Data frames in R for [Make](http://kbroman.org/minimal_make/)
+# What gets done stays done.
 
-Drake is a workflow manager and build system for
+Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus):
 
-1. [Reproducibility](https://CRAN.R-project.org/view=ReproducibleResearch).
-2. [High-performance computing](https://CRAN.R-project.org/view=HighPerformanceComputing).
+1. Deploy your code.
+2. Wait for it to finish.
+3. Discover an issue.
+4. Repeat from scratch.
+
+But `drake` tracks changes. When you update your code or data, `drake` figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is solid.
+
+```r
+load_basic_example()
+
+# First round.
+make(my_plan) 
+
+## target large
+## target small
+## target regression1_large
+## target regression1_small
+## target regression2_large
+## target regression2_small
+## target coef_regression1_large
+## target coef_regression1_small
+## target coef_regression2_large
+## target coef_regression2_small
+## target summ_regression1_large
+## target summ_regression1_small
+## target summ_regression2_large
+## target summ_regression2_small
+## target 'report.md'
+
+# Change some code.
+reg2 <- function(d){    
+  d$x4 <- d$x ^ 4
+  lm(y ~ x4, data = d)
+}
+
+# Less work for the second round.
+make(my_plan)
+
+## target regression2_large
+## target regression2_small
+## target coef_regression2_large
+## target coef_regression2_small
+## target summ_regression2_large
+## target summ_regression2_small
+## target 'report.md'
+
+# No work if nothing changed.
+make(my_plan)
+
+## All targets are already up to date.
+```
+
+# Stay reproducible.
+
+The R community likes to emphasize reproducibility. Most people think it means [scientific replicability](https://en.wikipedia.org/wiki/Replication_crisis), literate programming with [knitr](https://yihui.name/knitr/), or version control with [git](https://git-scm.com/book/en/v2/Getting-Started-About-Version-Control). But internal consistency is important too. Reproducibility carries the promise that your output matches the code and data it came from. Ordinarily, you might have to rerun everything from scratch just to be sure. But with `drake`, you can just check that all your targets are up to date.
+
+```r
+make(my_plan)
+
+## All targets are already up to date.
+
+config <- drake_config(my_plan)
+outdated(config)
+
+## character(0)
+```
+
+# Agressively scale up.
+
+Not every project can run in one R session on your laptop. Some projects require multiple cores, and some need large high-performance computing systems. But parallel computing is hard. Your tables and figures depend on your analysis results, your analyses depend on your datasets, and not everything can run at the same time. `Drake` automatically links your work together a network graph to figure out what it can process in parallel.
+
+```r
+# Change some code.
+reg2 <- function(d){    
+  d$x3 <- d$x ^ 3
+  lm(y ~ x3, data = d)
+}
+
+# Plot an interactive graph.
+config <- drake_config(my_plan)
+vis_drake_graph(config)
+```
+
+![](./images/graph.png)
+
[…]
