### 2017-12-09 9fa2743
Align drake.Rmd vignette with README.md
