+++ vignettes/drake.Rmd
@@ -28,13 +28,81 @@ knitr::opts_chunk$set(
 
 ![](logo-vignettes.png)
 
-Reproducibility carries the promise that your output matches your input. Your tables, figures, reports, and intermediate objects should stay up to date with all the underlying data and code. To ensure internal consistency, `drake` detects and refreshes the outdated components of your project. The same solution unlocks powerful high-performance computing functionality.
+# What gets done stays done.
 
-# Reproducibility with `drake`
+Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus):
+
+1. Deploy your code.
+2. Wait for it to finish.
+3. Discover an issue.
+4. Repeat from scratch.
+
+But `drake` tracks changes. When you update your code or data, `drake` figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is solid.
+
+```{r drakemake}
+load_basic_example()
+
+# First round.
+make(my_plan) 
+
+# Change some code.
+reg2 <- function(d){    
+  d$x4 <- d$x ^ 4
+  lm(y ~ x4, data = d)
+}
+
+# Less work for the second round.
+make(my_plan)
+
+# No work if nothing changed.
+make(my_plan)
+```
+
+# Stay reproducible.
+
+The R community likes to emphasize reproducibility. Most people think it means [scientific replicability](https://en.wikipedia.org/wiki/Replication_crisis), literate programming with [knitr](https://yihui.name/knitr/), or version control with [git](https://git-scm.com/book/en/v2/Getting-Started-About-Version-Control). But internal consistency is important too. Reproducibility carries the promise that your output matches the code and data it came from. Ordinarily, you might have to rerun everything from scratch just to be sure. But with `drake`, you can just check that all your targets are up to date.
+
+```{r drakemake2}
+make(my_plan)
+
+config <- drake_config(my_plan, verbose = FALSE)
+outdated(config)
+```
+
+# Agressively scale up.
+
+Not every project can run in one R session on your laptop. Some projects require multiple cores, and some need large high-performance computing systems. But parallel computing is hard. Your tables and figures depend on your analysis results, your analyses depend on your datasets, and not everything can run at the same time. `Drake` automatically links your work together a network graph to figure out what it can process in parallel.
+
+```{r visgraphdrake, eval = FALSE}
+# Change some code.
+reg2 <- function(d){    
+  d$x3 <- d$x ^ 3
+  lm(y ~ x3, data = d)
+}
+
+# Plot an interactive graph. Hover, click, drag, etc.
+config <- drake_config(my_plan)
+vis_drake_graph(config)
+```
+
+<iframe
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/reg2.html"
+width = "100%" height = "600px" allowtransparency="true"
+style="border: none; box-shadow: none">
+</iframe>
+
+Within each column above, the nodes are conditionally independent given their dependencies. Each `make()` walks through the columns from left to right and applies parallel processing within each column. If any nodes are already up to date, `drake` looks downstream to maximize the number of outdated targets in a parallelizable stage. To show the parallelizable stages of the next `make()` programmatically, use the `parallel_stages()` function.
+
[…]
