### 2017-12-09 0e05dce
Expand scaling section
