+++ vignettes/drake.Rmd
@@ -43,10 +43,10 @@ But `drake` tracks changes. When you update your code or data, `drake` figures o
 load_basic_example()
 
 # First round.
-make(my_plan) 
+make(my_plan)
 
 # Change some code.
-reg2 <- function(d){    
+reg2 <- function(d){
   d$x4 <- d$x ^ 4
   lm(y ~ x4, data = d)
 }
@@ -69,13 +69,26 @@ config <- drake_config(my_plan, verbose = FALSE)
 outdated(config)
 ```
 
-# Agressively scale up.
+# Aggressively scale up.
 
-Not every project can run in one R session on your laptop. Some projects require multiple cores, and some need large high-performance computing systems. But parallel computing is hard. Your tables and figures depend on your analysis results, your analyses depend on your datasets, and not everything can run at the same time. `Drake` automatically links your work together a network graph to figure out what it can process in parallel.
+Not every project can run in one R session on your laptop. Some projects need more speed or computing power. Some require a few local cores, and some need large high-performance computing systems. But parallel computing is hard. Your tables and figures depend on your analysis results, your analyses depend on your datasets, and some tasks must finish before others even begin. But `drake` knows what to do. Parallelism is implicit and automatic.
+
+```{r implicitparallel, eval = FALSE}
+# Use the spare cores on your local machine.
+make(my_plan, jobs = 4)
+
+# Scale up to a supercomputer.
+drake_batchtools_tmpl_file("slurm") # https://slurm.schedmd.com/
+library(future.batchtools)
+batchtools_slurm(template = "batchtools.slurm.tmpl", workers = 100)
+make(my_plan, parallelism = "future_lapply")
+```
+
+During parallel processing, the targets wait for their dependencies because `drake` uses a network graph.
 
 ```{r visgraphdrake, eval = FALSE}
 # Change some code.
-reg2 <- function(d){    
+reg2 <- function(d){
   d$x3 <- d$x ^ 3
   lm(y ~ x3, data = d)
 }
