### 2017-12-09 #156 @wlandau-lilly
@benmarwick, your advice is exactly what I needed! Even before the inception of `drake`, I have had a strong ambition to challenge the R community's current conventions around reproducible data analysis and research. I want to make `drake` more relatable, understandable, usable, and widespread, and it will be gratifying to elaborate on the practical niche. Thank you for steering me back on course.

I have already started working on the changes you requested. I expect to have ample time next week, but I will be on vacation from December 16 through January 3, and I will be totally off the grid and unreachable from December 25 through January 3. I will eagerly resume work on January 4.
