### 2017-12-09 ef9dd37
Work on writing
