### 2017-12-09 c464aac
Talk about related work more thoroughly in the README
