### 2017-12-09 e59de87
Talk about knitr in the README
