### 2017-12-09 5a78930
Add sketch of concrete example vignette
