+++ DESCRIPTION
@@ -50,14 +50,15 @@ Imports:
   visNetwork,
   withr
 Suggests: 
-  abind,
-  DBI,
-  future.batchtools,
-  MASS,
-  methods,
-  RSQLite,
-  rmarkdown,
-  tibble
+    abind,
+    DBI,
+    future.batchtools,
+    MASS,
+    methods,
+    RSQLite,
+    rmarkdown,
+    tibble,
+    knitr
 VignetteBuilder: knitr
 URL: https://github.com/wlandau-lilly/drake
 BugReports: https://github.com/wlandau-lilly/drake/issues
