### 2017-12-09 7d74cf5
Complete draft of application.Rmd vignette.
