+++ vignettes/application.Rmd
@@ -2,7 +2,10 @@
 title: "A concrete example of drake in action"
 author: "William Michael Landau"
 date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
+output:
+  rmarkdown::html_vignette:
+    toc: true
+    number_sections: true
 vignette: >
   %\VignetteIndexEntry{application}
   %\VignetteEngine{knitr::rmarkdown}
@@ -10,8 +13,10 @@ vignette: >
 ---
 
 ```{r, echo = F}
-suppressMessages(suppressWarnings(library(drake)))
-suppressMessages(suppressWarnings(library(magrittr)))
+suppressPackageStartupMessages(library(drake))
+suppressPackageStartupMessages(library(Ecdat))
+suppressPackageStartupMessages(library(ggplot2))
+suppressPackageStartupMessages(library(knitr))
 unlink(".drake", recursive = TRUE)
 clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
@@ -22,8 +27,199 @@ knitr::opts_chunk$set(
 )
 ```
 
+![](logo-vignettes.png)
+
+To generate the [code files for this example](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application), run `drake_example("application")`. Then, walk through `run.R` yourself.
+
+# Research question
+
+Which factors are most associated with the productivity of a US State?
+
+# Data
+
+The `Produc` dataset from the [Ecdat pacakge](https://cran.r-project.org/package=Ecdat) contains data on the Gross State Product from 1970 to 1986. It has the following variables. See the references later in this report for more details.
+
+- `gsp`: gross state product.
+- `state`: the state.
+- `year`: the year.
+- `pcap`: private capital stock.
+- `hwy`: highway and streets.
+- `water`: water and sewer facilities.
+- `util`: other public buildings and structures.
+- `pc`: public capital.
+- `emp`: labor input measured by the employment in non-agricultural payrolls.
+- `unemp`: state unemployment rate.
+
+```{r masterdata}
+library(Ecdat)
+data(Produc)
+head(Produc)
+```
+
+# Analysis
+
+The goal of the analysis is to find the covariates in the data that best predict gross state product. We attempt to predict `gsp` using the rest of the variables in `Produc`. Each model is a multiple regression of `gsp` on three covariates, and we apply all 84 such models. For simplicity, we rank the models based on their root mean squared prediction errors.
+
+First, we load the packages we need. `Drake` is aware of all the packages you load with `library()` or `require()`.
+
+```{r masterpkgs}
+library(drake)
+library(Ecdat) # econometrics datasets
+library(knitr)
+library(ggplot2)
+```
+
+Next, set up our workflow plan data frame in pieces. We start with the models.
+
+```{r mastermodels}
+predictors <- setdiff(colnames(Produc), "gsp")
+combos <- t(combn(predictors, 3))
+head(combos)
+targets <- apply(combos, 1, paste, collapse = "_")
+commands <- apply(combos, 1, function(row){
+  covariates <- paste(row, collapse = " + ")
+  formula <- paste0("as.formula(\"gsp ~ ", covariates, "\")")
+  command <- paste0("lm(", formula, ", data = Produc)")
+})
+model_plan <- data.frame(target = targets, command = commands)
+
+head(model_plan)
+```
+
[…]
