### 2017-12-09 565a4da
Fix minor build checks
