+++ DESCRIPTION
@@ -58,8 +58,7 @@ Suggests:
   methods,
   RSQLite,
   rmarkdown,
-  tibble,
-  knitr
+  tibble
 VignetteBuilder: knitr
 URL: https://github.com/wlandau-lilly/drake
 BugReports: https://github.com/wlandau-lilly/drake/issues
