### 2017-12-09 5cd47fa
Mention application.Rmd in the README
