+++ vignettes/application.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "A concrete example of drake in action"
+title: "A practical example of drake in action"
 author: "William Michael Landau"
 date: "`r Sys.Date()`"
 output:
