### 2017-12-09 8a6e8d0
Tweak application.Rmd
