### 2017-12-09 #156 @wlandau-lilly
# Response to @benmarwock's review

It was gratifying to work on this response, and the changes to `drake` were straightforward. For the past year, I have been struggling to find the best way to talk about `drake`. Your feedback allowed me to make substantial progress.

> It could be more clear what drake does that no other package already does, Or why a user should use drake rather than make/remake/code chunks in Rmd/etc.

I agree. Please see the new "Similar work" section of the README (as well as the end of the new [application.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/application.Rmd)). I now compare `drake` to [Make](https://www.gnu.org/software/make/), [remake](https://github.com/richfitz/remake), and [knitr](https://yihui.name/knitr/options/).

> The four points in 'similar work' section of the ropensci submission should also be in the readme.

Done. I added these points and expanded on them in the README.

> That said, some of these are a matter of style, e.g. YAML vs data frames for config, rather that outright novel function. Are there any projects using drake that a user can inspect to see real-world applications? Publications that can be cited?

I have tried to search for real-world examples of `drake` in the wild, but I have not had success so far. I think it may be too early to see publicly-released projects and publications that use `drake`. However, I do know that @kendonB, @japperdapper, and @AlexAxthelm are heavily using `drake` for their projects. I cannot share any of my own because they are all company confidential.

In the new "Similar work" section of the README, I now refer to real-world applications of [Make](https://www.gnu.org/software/make/) and [remake](https://github.com/richfitz/remake), and I argue that `drake` improves on both these tools for R users. I hope that helps. In the coming years, I will continue to search for publicly-available `drake`-powered projects, and I will keep a running list in the README.

> My sense is that many packages aimed at workflow management or improving reproducibility are quite idiomatic. This makes it hard for the common-or-garden variety R-user to see how they fit into their own ways of working. If you can help to bridge that gap between your idioms and the average user, that would make this pkg much more useful and valuable to the community.

I completely agree! I have been struggling with communication this entire time. The first three sections of the current README and [drake.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/drake.Rmd) are new, and I think they are substantial improvements. I try to introduce `drake` using plain language, and I argue that it makes life easier.

>  I had no problem running these. However, I found the quickstart and examples difficult to relate to. For example, who writes their Rmd report in the console and passes it to an object as a character vector? That seems unnatural, to me at least, where the Rmd file is my main notebook and workbench. It would be easier to follow if the substance of the analysis was narrated in a little more detail. Perhaps a tiny actual research question with actual data would make this example more accessible? Perhaps also a comparison with a simple makefile to show makefile users (the main audience for this pkg) how to accomplish the same with drake (and why drake would be preferable). This would help a reader see how their existing workflow could be translated to the drake system. The drake system is a very comprehensive universe of functions, and new users will need a bit more guidance to see analogues between what drake does and what they're already using.

Absolutely! I was too entrenched in the details to realize this. I have added a new [application.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/application.Rmd) for exactly this purpose, and I paired it with [example code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) that the user can generate with `drake_example("application")`. Here, I define a research question and use real data to address it. I also comment on how [Make](https://www.gnu.org/software/make/) would be unwise for that particular use case.

Regarding the use of `knitr`, please see the new [knitr subsection](https://github.com/wlandau-lilly/drake/blob/master/README.md#knitr) of the README.

> Could CONTRIBUTING.md be located at the top level of the repo for better visibility?

Done.

> I see that you do not plan to submit to JOSS at the moment, so this is just an incidental comment: It is easy to imagine research applications for drake, it is a very solid contribution to an active area of workflow and provenance tracking tools. However, the research application would be more obvious if the readme referred to some actual real-world uses of the package. For example, a list of domain-specific research project repos where drake is used (i.e. by biologists, economists, whatever), or a list of publications reporting results that were generated or enabled using drake. Currently it looks like drake has great potential, but hasn't actually been used in any real-world applications. Perhaps it has, but it's not clear from the pkg docs. Examples of use would help potential users better understand how drake can help them.

I absolutely do plan to submit to JOSS in the future. Now is not the right time for me, however, and I am especially glad I received your feedback on the package itself first. And as I mentioned before, I am currently having trouble finding real-world applications of `drake` in the wild. I will continue searching, and I will gather and list them in the README when I see them.

Now I have a question: @maelle, may I return to this thread at a later date to fast-track a JOSS submission?
