### 2017-12-09 8770ab0
Simplify mention of graph
