### 2017-12-09 #156 @maelle
Thanks for your review, @benmarwick! Are you happy with the changes?

@wlandau-lilly, thanks for answering @benmarwick's review promptly. Three points from me:

* The JOSS submission would only need a paper.md and archived version of the repository. We do not need that before the end of onboarding, which will be at a later date. 😉

* Why not create a website for the package using `pkgdown`? It'll make the vignettes easier to browse. See http://enpiar.com/2017/11/21/getting-down-with-pkgdown/

* Very naive question, does each command need to be something as basic as `summary` or could it be sourcing a larger script containing several regression calls?
