### 2017-12-09 #156 @maelle
Another argument in favour of `pkgdown`: you could create a grouping of functions as in [this example](https://dirkschumacher.github.io/ompr/reference/index.html) which is what you have in your README now but without the documentation of each function accessible by one click.

I've also just noticed this phrasing "Most people think that means". Even if you have data underlying this, I think it looks a bit agressive here, maybe replace it "It does not only mean". :-)
