### 2017-12-09 6bcfca7
Tone down a strong claim
