### 2017-12-09 #156 @wlandau-lilly
@maelle,

>The JOSS submission would only need a paper.md and archived version of the repository. We do not need that before the end of onboarding, which will be at a later date. :wink:

Very much appreciated!

> Why not create a website for the package using pkgdown? It'll make the vignettes easier to browse. See http://enpiar.com/2017/11/21/getting-down-with-pkgdown/

Good suggestion. `Drake` heavily relies on its vignettes, and `pkgdown` is a community standard for documentation. I expect to begin work on this soon, but if it is okay with you, could we wait until after onboarding to launch it? Otherwise, I would expect the URL to change after transferring the repo.

> Very naive question, does each command need to be something as basic as summary or could it be sourcing a larger script containing several regression calls?

I am glad you asked! `Drake` commands can be arbitrary R code (although I would avoid unquoted formulas because they may throw off the static code analysis that detects dependencies. This would not break `make()`, but it may create false positive messages about missing import objects). So yes, you could have a large script containing several regression calls separated by `;` or `\n`. This is yet another advantage over [remake](https://github.com/richfitz/remake), which requires all commands to be single function calls with no nesting (except for `I()`, which declares string literals).

Inside `drake`, each command is wrapped in a protective function call in order to quarantine the side effects, so in general, only the return value should have an effect on the rest of `make()` (see wlandau-lilly/drake#39).

Large commands are not always good practice because they can make workflow plan data frames difficult to print properly. (Full disclosure: `gather_plan()` creates super long commands, so I am guilty.) I remember submitting a feature request to [tibble](https://github.com/tidyverse/tibble) to allow individual columns to be truncated, but I cannot seem to find the issue.

Every project needs a balance between having too many targets and assigning too much work to any individual target. The new [application.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/application.Rmd) implicitly hints at a possible explosion in the number of targets for massive studies with crushing combinatorics. There is no one-size-fits-all solution.

> I've also just noticed this phrasing "Most people think that means". Even if you have data underlying this, I think it looks a bit agressive here, maybe replace it "It does not only mean". :-)

You are right. In wlandau-lilly/drake@6bcfca7a12aafdcdb50cf5bb11904a8c3eaaac52, I just changed the sentence to "The R community likes to emphasize reproducibility, which one could interpret to mean..."
