### 2017-12-09 72e2aab
Correct spelling of a cited author
