### 2017-12-10 2409b4c
Rearrange _pkgdown.yml
