+++ _pkgdown.yml
@@ -35,18 +35,25 @@ navbar:
 articles:
 - title: General vignettes
   contents:
-  - '`application`'
-  - '`caution`'
   - '`drake`'
+  - '`application`'
   - '`quickstart`'
+  - '`caution`'
 - title: Specialized walkthroughs
   contents:
   - '`debug`'
   - '`graph`'
   - '`parallelism`'
-  - '`storage`'
   - '`timing`'
+  - '`storage`'
 reference:
+- title: Help
+  contents:
+  - '`drake_example`'
+  - '`drake_examples`'
+  - '`drake-package`'
+  - '`drake_tip`'
+  - '`load_basic_example`'
 - title: Basic
   contents:
   - '`drake_config`'
@@ -54,21 +61,58 @@ reference:
   - '`make`'
   - '`make_targets`'
   - '`make_with_config`'
-- title: Cache configuration
-  - '`available_hash_algos`'
-  - '`cache_namespaces`'
-  - '`cache_path`'
-  - '`cleaned_namespaces`'
-  - '`configure_cache`'
-  - '`default_cache_path`'
-  - '`default_long_hash_algo`'
-  - '`default_short_hash_algo`'
-  - '`long_hash`'
-  - '`new_cache`'
-  - '`recover_cache`'
-  - '`short_hash`'
-  - '`target_namespaces`'
-  - '`this_cache`'
+- title: Workflow plan data frames
+  contents:
+  - '`analysis_wildcard`'
+  - '`dataset_wildcard`'
+  - '`evaluate_plan`'
+  - '`expand_plan`'
+  - '`gather_plan`'
+  - '`plan_analyses`'
+  - '`plan_drake`'
+  - '`plan_summaries`'
+- title: Network graphs and visualization
+  contents:
+  - '`build_drake_graph`'
+  - '`drake_palette`'
+  - '`dataframes_graph`'
+  - '`legend_nodes`'
+  - '`prune_drake_graph`'
+  - '`render_drake_graph`'
+  - '`vis_drake_graph`'
+- title: Debugging and testing
+  contents:
+  - '`check_plan`'
+  - '`default_trigger`'
+  - '`dependency_profile`'
+  - '`deps`'
+  - '`diagnose`'
+  - '`knitr_deps`'
+  - '`migrate_drake_project`'
+  - '`missed`'
+  - '`outdated`'
+  - '`possible_targets`'
+  - '`tracked`'
+  - '`triggers`'
+- title: High-performance computing
+  contents:
+  - '`default_Makefile_args`'
+  - '`default_Makefile_command`'
+  - '`default_parallelism`'
+  - '`default_recipe_command`'
+  - '`drake_batchtools_tmpl_file`'
+  - '`Makefile_recipe`'
+  - '`max_useful_jobs`'
+  - '`next_stage`'
+  - '`parallelism_choices`'
+  - '`parallel_stages`'
+  - '`r_recipe_wildcard`'
+  - '`shell_file`'
+- title: Timing
+  contents:
+  - '`build_times`'
+  - '`predict_runtime`'
+  - '`rate_limiting_times`'
 - title: Cache usage
   contents:
   - '`built`'
@@ -90,41 +134,21 @@ reference:
   - '`read_drake_plan`'
   - '`readd`'
   - '`rescue_cache`'
-- title: Debugging and testing
-  contents:
-  - '`check_plan`'
-  - '`default_trigger`'
-  - '`dependency_profile`'
-  - '`deps`'
-  - '`diagnose`'
-  - '`knitr_deps`'
-  - '`migrate_drake_project`'
-  - '`missed`'
-  - '`outdated`'
-  - '`possible_targets`'
-  - '`tracked`'
[…]
