### 2017-12-10 e20b989
Make README friendlier
