+++ vignettes/drake.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "drake"
-subtitle: "data frames in R for Make"
+subtitle: "stay reproducible and save time"
 author: "William Michael Landau"
 date: "`r Sys.Date()`"
 output:
@@ -28,6 +28,10 @@ knitr::opts_chunk$set(
 
 ![](logo-vignettes.png)
 
+# Why use drake
+
+`Drake` saves time. It notices any previous work you did and only refreshes the pieces that are out of date or missing. That way, your data analysis project **thinks before it acts**, and your results stay consistent with the code and data they came from. The time you spend learning `drake` will pay off when you no longer have to rerun your work from scratch. The time savings are even greater if you turn on the [parallel computing](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd), which is friendly, comprehensive, and compatible with small and large projects alike.  `Drake` is more dependable, R-focused, scalable, well-documented, and actively maintained than [other tools with similar functionality](#similar-work).
+
 # What gets done stays done.
 
 Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus):
@@ -37,9 +41,10 @@ Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.o
 3. Discover an issue.
 4. Restart from scratch.
 
-But `drake` tracks changes. When you update your code or data, `drake` figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is steadier and smoother.
+`Drake` is more discerning. When you update your code or data, it figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is steadier and smoother.
 
 ```{r drakemake}
+library(drake)
 load_basic_example()
 
 # First round.
@@ -187,6 +192,53 @@ reg2 <- function(d){
 
 make(my_plan)
 ```
+
+# Similar work
+
+## GNU Make
+
+The original idea of a time-saving reproducible build system extends back at least as far as [GNU Make](https://www.gnu.org/software/make/), which still aids the work of [data scientists](http://blog.kaggle.com/2012/10/15/make-for-data-scientists/) as well as the original user base of complied language programmers. In fact, the name "drake" stands for "Data Frames in R for Make". [Make](http://kbroman.org/minimal_make/) is used widely in reproducible research. Below are some examples from [Karl Broman's website](http://kbroman.org/minimal_make/).
+
+- Bostock, Mike (2013). "A map of flowlines from NHDPlus." https://github.com/mbostock/us-rivers. Powered by the Makefile at https://github.com/mbostock/us-rivers/blob/master/Makefile.
+- Broman, Karl W (2012). "Halotype Probabilities in Advanced Intercross Populations." *G3* 2(2), 199-202. https://doi.org/10.1534/g3.111.001818. Powered by the `Makefile` at https://github.com/kbroman/ailProbPaper/blob/master/Makefile.
+- Broman, Karl W (2012). "Genotype Probabilities at Intermediate Generations in the Construction of Recombinant Inbred Lines." *Genetics 190(2), 403-412. https://doi.org/10.1534/genetics.111.132647. Powered by the Makefile at https://github.com/kbroman/preCCProbPaper/blob/master/Makefile.
+- Broman, Karl W and Kim, Sungjin and Sen, Saunak and Ane, Cecile and Payseur, Bret A (2012). "Mapping Quantitative Trait Loci onto a Phylogenetic Tree." *Genetics* 192(2), 267-279. https://doi.org/10.1534/genetics.112.142448. Powered by the `Makefile` at https://github.com/kbroman/phyloQTLpaper/blob/master/Makefile.
+
+There are several reasons for R users to prefer `drake` instead.
+
[…]
