+++ README.md
@@ -1,4 +1,4 @@
-# Data frames in R for [Make](https://www.gnu.org/software/make/)
+# drake: stay reproducible and save time
 
 <p align="center">
   <img src="./images/logo-readme.png" alt="">
@@ -14,6 +14,10 @@
 [![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/wlandau-lilly/drake?branch=master&svg=true)](https://ci.appveyor.com/project/wlandau-lilly/drake)
 [![codecov.io](https://codecov.io/github/wlandau-lilly/drake/coverage.svg?branch=master)](https://codecov.io/github/wlandau-lilly/drake?branch=master)
 
+# Why use drake
+
+`Drake` saves time. It notices any previous work you did and only refreshes the pieces that are out of date or missing. That way, your data analysis project **thinks before it acts**, and your results stay consistent with the code and data they came from. The time you spend learning `drake` will pay off when you no longer have to rerun your work from scratch. The time savings are even greater if you turn on the [parallel computing](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd), which is friendly, comprehensive, and compatible with small and large projects alike.  `Drake` is more dependable, R-focused, scalable, well-documented, and actively maintained than [other tools with similar functionality](#similar-work).
+
 # What gets done stays done.
 
 Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus):
@@ -23,12 +27,13 @@ Too many data analysis projects follow a [Sisyphean loop](https://en.wikipedia.o
 3. Discover an issue.
 4. Restart from scratch.
 
-But `drake` tracks changes. When you update your code or data, `drake` figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is steady and smooth.
+`Drake` is more discerning. When you update your code or data, it notices the changes. It figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is steady and smooth.
 
 ```r
+library(drake)
 load_basic_example()
 
-# First round.
+# First round: builds 15 targets.
 make(my_plan) 
 
 ## target large
@@ -47,7 +52,7 @@ make(my_plan)
 ## target summ_regression2_small
 ## target 'report.md'
 
-# Change some code.
+# Change some of your code.
 reg2 <- function(d){    
   d$x4 <- d$x ^ 4
   lm(y ~ x4, data = d)
@@ -133,186 +138,19 @@ install_github("wlandau-lilly/drake", build = TRUE)        # Development version
 - You must properly install `drake` using `install.packages()`, `devtools::install_github()`, or similar. It is not enough to use `devtools::load_all()`, particularly for the parallel computing functionality, in which multiple R sessions initialize and then try to `require(drake)`.
 - For `make(..., parallelism = "Makefile")`, Windows users need to download and install [`Rtools`](https://cran.r-project.org/bin/windows/Rtools/).
 
-# Where to begin
-
-Outline your work in a data frame.
-
-```r
-library(drake)
-load_basic_example()
-my_plan
-```
-
-```r
-##                    target                                      command
-## 1             'report.md'             knit('report.Rmd', quiet = TRUE)
-## 2                   small                                  simulate(5)
-## 3                   large                                 simulate(50)
-## 4       regression1_small                                  reg1(small)
-## 5       regression1_large                                  reg1(large)
-## 6       regression2_small                                  reg2(small)
[…]
