### 2017-12-10 be2ac41
Repair _pkgdown.yml
