### 2017-12-10 e6214ba
Reorder vignettes in _pkgdown.yml
