### 2017-12-10 ebded86
Tweak "why drake" section
