### 2017-12-10 ff2cc44
Rephrase beginning
