### 2017-12-10 d83fd95
Tweak "why drake" paragraph
Do not anthropomorphise projects
