### 2017-12-10 6c73487
Improve the application.Rmd vignette and example
Mostly improved the writing.
Related: @benmarwick's review at ropensci/onboarding#156.
