+++ vignettes/application.Rmd
@@ -29,15 +29,22 @@ knitr::opts_chunk$set(
 
 ![](logo-vignettes.png)
 
-To generate the [code files for this example](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application), run `drake_example("application")`. Then, walk through `run.R` yourself.
+The following data analysis workflow a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you with to tinker it with yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application), with `drake_example("application")`. Then, walk through `run.R` yourself.
 
-# Research question
+# Objective and methods
 
-Which factors are most associated with the productivity of a US State?
+The goal is to search for factors closely associated with the productivity of states in the USA around the 1970s and 1980s. For the sake of simplicity, we use gross state product as a metric of productivity, we restrict ourselves to multiple linear regression models with three variables. For each of the 84 possible models, we fit the data and then evaluate the root mean squared prediction error (RMSPE).
+
+$$
+\begin{aligned}
+\text{RMSPE} = \sqrt{(\text{y} - \widehat{y})^T(y - \widehat{y})}
+\end{aligned}
+$$
+Here, $y$ is the vector of observed gross state products in the data, and $\widehat{y}$ is the vector of predicted gross state products under one of the models. We take the best variables to be the triplet in the model with the lowest RMSPE.
 
 # Data
 
-The `Produc` dataset from the [Ecdat package](https://cran.r-project.org/package=Ecdat) contains data on the Gross State Product from 1970 to 1986. It has the following variables. See the references later in this report for more details.
+The `Produc` dataset from the [Ecdat package](https://cran.r-project.org/package=Ecdat) contains data on the Gross State Product from 1970 to 1986. Each row is a single observation on a single state for a single year. The dataset has the following variables as columns. See the references later in this report for more details.
 
 - `gsp`: gross state product.
 - `state`: the state.
@@ -58,9 +65,7 @@ head(Produc)
 
 # Analysis
 
-The goal of the analysis is to find the covariates in the data that best predict gross state product. We attempt to predict `gsp` using the rest of the variables in `Produc`. Each model is a multiple regression of `gsp` on three covariates, and we apply all 84 such models. For simplicity, we rank the models based on their root mean squared prediction errors.
-
-First, we load the packages we need. `Drake` is aware of all the packages you load with `library()` or `require()`.
+First, we load the required packages. `Drake` is aware of all the packages you load with `library()` or `require()`.
 
 ```{r masterpkgs}
 library(drake)
@@ -69,7 +74,7 @@ library(knitr)
 library(ggplot2)
 ```
 
-Next, set up our workflow plan data frame in pieces. We start with the models. Each model has 3 predictors, and we try all 84 possible models.
+Next, set up our workflow plan data frame in stages. We start with the models. Each model has 3 predictors, and we try all 84 possible models.
 
 ```{r mastermodels}
 predictors <- setdiff(colnames(Produc), "gsp")
@@ -131,6 +136,8 @@ output_plan <- plan_drake(
   file_targets = TRUE,
   strings_in_dots = "literals"
 )
+
+head(output_plan)
 ```
 
 At this point, we can gather together the whole workflow plan.
@@ -175,7 +182,7 @@ library(ggplot2)
 plot_rmspe(rmspe = results)
 ```
 
-And here are the best models.
+And here are the best models. The best variables are in the top row under `X1`, `X2`, and `X3`.
 
 ```{r masterbestmodels}
 head(results[order(results$rmspe, decreasing = FALSE), ])
