### 2017-12-10 613f381
Fix more prose
