+++ vignettes/application.Rmd
@@ -29,11 +29,11 @@ knitr::opts_chunk$set(
 
 ![](logo-vignettes.png)
 
-The following data analysis workflow is a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you with to tinker with it yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`.
+The following data analysis workflow is a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you wish to tinker with it yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`.
 
 # Objective and methods
 
-The goal is to search for factors closely associated with the productivity of states in the USA around the 1970s and 1980s. For the sake of simplicity, we use gross state product as a metric of productivity, we restrict ourselves to multiple linear regression models with three variables. For each of the 84 possible models, we fit the data and then evaluate the root mean squared prediction error (RMSPE).
+The goal is to search for factors closely associated with the productivity of states in the USA around the 1970s and 1980s. For the sake of simplicity, we use gross state product as a metric of productivity, and we restrict ourselves to multiple linear regression models with three variables. For each of the 84 possible models, we fit the data and then evaluate the root mean squared prediction error (RMSPE).
 
 $$
 \begin{aligned}
