+++ inst/examples/application/report.Rmd
@@ -9,7 +9,7 @@ suppressPackageStartupMessages(library(drake))
 suppressPackageStartupMessages(library(Ecdat))
 ```
 
-The following data analysis workflow a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you with to tinker it with yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application), with `drake_example("application")`. Then, walk through `run.R` yourself, which creates the results and knits this report. This `report.Rmd` file will not knit on its own until `run.R` has generated enough output.
+The following data analysis workflow a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you wish to tinker with it yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`. This `report.Rmd` file will not knit on its own until `run.R` has generated enough output.
 
 # Objective and methods
 
