### 2017-12-10 3b62a3d
Tweak more wording
