### 2017-12-10 260d277
Tweak Makefile section of vignette
