### 2017-12-12 #156 @jules32
Hi @wlandau-lilly et al,

Sorry for the delay here. I'm based in Santa Barbara, California, and with the huge wildfire that is ongoing, we've left town in the last few days.

I did have a look, and had a lot of the same thoughts as @benmarwick in trying to understand how
 the intended users of`drake` would know that it was the right tool for them. I know you are addressing some of Ben's comments before going on holiday; I'll have more comments for you in the New Year too.

Cheers,
Julie
