### 2017-12-12 f3d8c98
Define get_from_meta() and set_in_meta()
