+++ R/meta.R
@@ -114,3 +114,30 @@ file_hash <- function(target, config, size_cutoff = 1e5) {
     config$cache$get(key = target, namespace = "kernels")
   }
 }
+
+get_from_meta <- function(key, metaspace, cache){
+  if (!inherits(cache, "storr")){
+    stop("cache is missing.")
+  }
+  meta <- safe_get(key = key, namespace = "meta", config = list(cache = cache))
+  if (metaspace %in% names(meta)){
+    meta[[metaspace]]
+  } else {
+    NULL
+  }
+}
+
+set_in_meta <- function(key, value, metaspace, cache){
+  if (!inherits(cache, "storr")){
+    stop("cache is missing.")
+  }
+  if (cache$exists(key = key, namespace = "meta")){
+    meta <- cache$get(key = key, namespace = "meta")
+    meta[[metaspace]] <- value
+    cache$set(key = key, value = meta, namespace = "meta")
+  } else {
+    meta <- list(key, value)
+    names(meta) <- c("target", metaspace)
+    cache$set(key = key, value = meta, namespace = "meta")
+  } 
+}
