+++ R/migrate.R
@@ -64,7 +64,6 @@ migrate_drake_project <- function(
   config$verbose <- TRUE
   config$trigger <- "any"
   config$execution_graph <- config$graph
-  config$store_meta <- TRUE
   config$lazy_load <- FALSE
   config$session_info <- TRUE
   config$outdated <- legacy_outdated(config) %>%
