+++ R/build_times.R
@@ -116,6 +116,9 @@ to_build_duration <- function(x){
 time_columns <- c("elapsed", "user", "system")
 
 append_times_to_meta <- function(target, start, meta, config){
+  if (any(is.na(start))){
+    return(meta)
+  }
   build_times <- (proc.time() - start) %>%
     runtime_entry(target = target, imported = meta$imported)
   meta$build_times <- build_times
