### 2017-12-12 3d2e441
Put commands and dependency hashes in "meta" namespace
