+++ tests/testthat/test-full-build.R
@@ -85,7 +85,7 @@ test_with_dir("scratch build with custom filesystem cache.", {
   meta_list <- read_drake_meta(cache = cache)
   expect_true(length(meta_list) == 0)
   expect_true(length(cached(cache = cache, namespace = "meta")) == 0)
-  
+
   clean(destroy = TRUE, cache = cache)
   expect_false(file.exists(path))
 })
