### 2017-12-12 da0cbaf
Disconnect all drivers
