### 2017-12-12 0acd9c5
Add back clear_progress arg to make()
