### 2017-12-12 accbcd7
Try to fix Windows builds
Apparently, Windows stores some NA times from Sys.time()
