### 2017-12-12 6f5c71a
Try to fix Appveyor again
it all has to do with storing times
