### 2017-12-12 c9b3aa8
Add log_progress flag
