+++ tests/testthat/test-cache.R
@@ -15,11 +15,15 @@ test_with_dir("clean() works if there is no cache already", {
   expect_false(file.exists(default_cache_path()))
 })
 
-test_with_dir("bad/corrupt caches", {
+test_with_dir("bad/corrupt caches, no progress", {
   expect_null(drake_fetch_rds("sldkfjlke"))
   expect_warning(new_cache(type = "nope"))
   x <- plan_drake(a = 1)
-  make(x, verbose = FALSE, session_info = FALSE)
+  make(x, verbose = FALSE, session_info = FALSE, log_progress = FALSE)
+  expect_equal(get_cache()$list(namespace = "progress"), character(0))
+  clean()
+  make(x, verbose = FALSE, session_info = FALSE, log_progress = TRUE)
+  expect_equal(get_cache()$list(namespace = "progress"), "a")
   path <- file.path(default_cache_path(), "config", "hash_algorithm")
   expect_true(file.exists(path))
   unlink(path)
