+++ R/build.R
@@ -53,8 +53,11 @@ build_in_hook <- function(target, meta, config) {
   start <- proc.time()
   meta <- finish_meta(
     target = target, meta = meta, config = config)
-  config$cache$set(key = target, value = "in progress",
-    namespace = "progress")
+  set_progress(
+    target = target,
+    value = "in progress",
+    config = config
+  )
   console(imported = meta$imported, target = target, config = config)
   if (meta$imported) {
     value <- imported_target(target = target, meta = meta,
