+++ R/config.R
@@ -39,7 +39,7 @@
 #' @param elapsed same as for \code{\link{make}}
 #' @param retries same as for \code{\link{make}}
 #' @param force same as for \code{\link{make}}
-#' @param clear_progress logical, whether to clear
+#' @param log_progress logical, whether to clear
 #' the cached progress of the targets readable by
 #' @param graph igraph object representing the workflow plan network.
 #' Overrides \code{skip_imports}.
@@ -97,7 +97,7 @@ drake_config <- function(
   elapsed = timeout,
   retries = 0,
   force = FALSE,
-  clear_progress = FALSE,
+  log_progress = FALSE,
   graph = NULL,
   trigger = drake::default_trigger(),
   imports_only = FALSE,
@@ -128,7 +128,7 @@ drake_config <- function(
   # A storr_rds() cache should already have the right hash algorithms.
   cache <- configure_cache(
     cache = cache,
-    clear_progress = clear_progress,
+    log_progress = log_progress,
     overwrite_hash_algos = FALSE,
     jobs = jobs
   )
@@ -154,7 +154,7 @@ drake_config <- function(
     seed = seed, trigger = trigger,
     timeout = timeout, cpu = cpu, elapsed = elapsed, retries = retries,
     imports_only = imports_only, skip_imports = skip_imports,
-    skip_safety_checks = skip_safety_checks,
+    skip_safety_checks = skip_safety_checks, log_progress = log_progress,
     lazy_load = lazy_load, session_info = session_info
   )
 }
