### 2017-12-12 02f9058
Update docs based on recent refactoring
