### 2017-12-12 ae61f08
Add namespace arg to loadd() and readd()
