### 2017-12-12 8d69258
Try to fix code coverage
