### 2017-12-12 #156 @maelle
Thanks @jules32!

I assigned a 3d reviewer, @jeroen, to have a look at the implementation, not the interface. ☺
