### 2017-12-12 0e3ab27
Clarify the metadata in dependency_profile()
