+++ man/dependency_profile.Rd
@@ -24,6 +24,12 @@ of the returned list should agree: for example,
 \code{cached_dependency_hash} and
 \code{current_dependency_hash}.
 }
+\details{
+The dependency profile is based on cached metadata
+taken \emph{just before} the target was built, not
+immediately afterwards. If the metadata seems
+out of date, that is the reason why.
+}
 \examples{
 \dontrun{
 test_with_dir("Quarantine side effects.", {
