### 2017-12-12 d0fc6dc
Deprecate plan_drake() in favor of drake_plan()
.......
