+++ R/debug.R
@@ -26,7 +26,7 @@ dbug_envir <- function(envir) {
 }
 
 dbug_plan <- function() {
-  plan_drake(list = c(
+  drake_plan(list = c(
     `'intermediatefile.rds'` = "saveRDS(combined, \"intermediatefile.rds\")",
     yourinput = "f(1+1)",
     nextone = "myinput + g(7)",
