+++ tests/testthat/test-edge-cases.R
@@ -1,7 +1,7 @@
 drake_context("edge cases")
 
 test_with_dir("config and make without safety checks", {
-  x <- plan_drake(file = readRDS("my_file.rds"))
+  x <- drake_plan(file = readRDS("my_file.rds"))
   expect_warning(tmp <- config(x, verbose = FALSE))
   expect_silent(
     tmp <- drake_config(x, skip_safety_checks = TRUE, verbose = FALSE))
@@ -9,7 +9,7 @@ test_with_dir("config and make without safety checks", {
 })
 
 test_with_dir("Strings stay strings, not symbols", {
-  x <- plan_drake(a = "A", strings_in_dots = "literals")
+  x <- drake_plan(a = "A", strings_in_dots = "literals")
   expect_silent(make(x, verbose = FALSE, session_info = FALSE))
 })
 
@@ -21,7 +21,7 @@ test_with_dir("error handlers", {
 })
 
 test_with_dir("error when file target names do not match actual filenames", {
-  x <- plan_drake(y = 1, file_targets = TRUE)
+  x <- drake_plan(y = 1, file_targets = TRUE)
   expect_warning(expect_error(make(x, verbose = FALSE, session_info = FALSE)))
 })
 
@@ -67,7 +67,7 @@ test_with_dir("target conflicts with previous import", {
 })
 
 test_with_dir("can use semicolons and multi-line commands", {
-  plan <- plan_drake(list = c(x = "a<-1; a", y = "b<-2\nb"))
+  plan <- drake_plan(list = c(x = "a<-1; a", y = "b<-2\nb"))
   make(plan, verbose = FALSE, session_info = FALSE)
   expect_false(any(c("a", "b") %in% ls()))
   expect_true(all(cached(x, y, search = FALSE)))
@@ -78,7 +78,7 @@ test_with_dir("true targets can be functions", {
   generator <- function() return(function(x) {
     x + 1
   })
-  plan <- plan_drake(myfunction = generator(), output = myfunction(1))
+  plan <- drake_plan(myfunction = generator(), output = myfunction(1))
   config <- make(plan, verbose = FALSE, session_info = FALSE)
   expect_equal(readd(output), 2)
   expect_true(
