+++ tests/testthat/test-random.R
@@ -10,7 +10,7 @@ test_with_dir("Random targets are reproducible", {
   parallelism <- scenario$parallelism
   jobs <- scenario$jobs
 
-  data <- plan_drake(
+  data <- drake_plan(
     x = runif(20),
     y = runif(20),
     z = rnorm(20),
