+++ R/knitr.R
@@ -3,7 +3,7 @@
 #' @seealso \code{\link{deps}},
 #' \code{\link{make}}, \code{\link{load_basic_example}}
 #' @description Find the dependencies of a dynamic report. To
-#' enable drake to watch for these dependencies, your plan_drake
+#' enable drake to watch for these dependencies, your drake_plan
 #' plan command to compile this report must make direct use of
 #' \code{knitr::knit()}.
 #' That is, it must look something like \code{knit('your_report.Rmd')}
