### 2017-12-12 ae7ae61
Fix vignette and add man page
