+++ tests/testthat/test-deprecate.R
@@ -27,9 +27,11 @@ test_with_dir("drake_plan deprecation", {
   pl3 <- expect_warning(plan_drake(x = 1, y = x))
   expect_equal(pl2, pl3)
   expect_warning(drake::plan())
-  expect_warning(drake::plan(x = y, file_targets = TRUE))
+  expect_warning(drake::plan_drake())
   expect_warning(drake::workflow())
   expect_warning(drake::workplan())
+  expect_warning(drake::plan(x = y, file_targets = TRUE))
+  expect_warning(drake::plan_drake(x = y, file_targets = TRUE))
   expect_warning(drake::workflow(x = y, file_targets = TRUE))
   expect_warning(drake::workplan(x = y, file_targets = TRUE))
   expect_warning(check(drake_plan(a = 1)))
