### 2017-12-13 #156 @maelle
Yes! After discussing with @jeroen and @jules32 the new deadline is Jan the 4th after your vacation. Sorry about the process length but it'll have been worth it I think with such a reviewers dream team! 🦄🦄🦄

Will you soon have time for JOSS paper.md? See their instructions, it's really a short paper.
