### 2017-12-13 #156 @wlandau-lilly
That works for me. I will return refreshed and ready to respond.

I just read http://joss.theoj.org/about#author_guidelines, and it turns out that I completely misunderstood JOSS! I assumed that I would need to write a full-length journal article and that the expectations and process would be similar to [JSS](https://www.jstatsoft.org), etc.

I think a JOSS submission should be possible early next year, but it will take some time.  Given all the great feedback I am about to receive through rOpenSci, I would rather wait. The `paper.md` will be quick to write, but my company requires a disclosure process for academic journals, and there is not enough time left in the year to initiate a new disclosure. Also, each iteration of `paper.md` will need to be reviewed and approved all over again. But I can minimize the bureaucratic red tape if I am overprepared.
