### 2017-12-13 bcfdfe4
Change abstract-like summary
