### 2017-12-13 3f5f5f0
Shorten some prose
