### 2017-12-13 7e73493
Avoid redundant wording
