### 2017-12-13 31f68dc
Change some word choices
