### 2017-12-13 e8c633a
Contemplate memoise
