### 2017-12-13 88ec8e6
Tweak section on memoise
