### 2017-12-13 17bba1a
Touch up storage vignette
