### 2017-12-13 b95f922
Improve stub of dbi cache stuff
