### 2017-12-13 a5b51f7
Make the top-level summary more understandable
