### 2017-12-14 c5bad7d
Clean up build logic
Make sure saved metadata is up to date and corresponds
to the target after it is built. Except for meta$missing.
