### 2017-12-14 e76a3ef
Fill in details in read_drake_meta()
