### 2017-12-14 7c49398
Document read_drake_meta()
