+++ R/read.R
@@ -403,11 +403,12 @@ read_drake_graph <- function(
 #' with \code{drake:::meta()} and then
 #' \code{drake:::finish_meta()}. This metadata corresponds
 #' to the state of the target immediately after it was built
-#' or imported in the last \code{\link{make}()}.
+#' or imported in the last \code{\link{make}()} that
+#' did not skip it.
 #' The exception to this is the \code{$missing} element
 #' of the metadata, which indicates if the target/import
 #' was missing just \emph{before} it was built.
-#' @seealso \code{\link{make}}
+#' @seealso \code{\link{dependency_profile}}, \code{\link{make}}
 #' @export
 #' @return The cached master internal configuration list
 #' of the last \code{\link{make}()}.
@@ -470,5 +471,8 @@ read_drake_meta <- function(
     jobs = jobs
   )
   names(out) <- targets
+  if (length(out) == 1){
+    out <- out[[1]]
+  }
   out
 }
