### 2017-12-14 15b0478
Avoid ambiguity: begin with "The drake R package..."
