### 2017-12-14 b13a7ed
Bugfix in file_hash(): only read cached file hash if it exists
