### 2017-12-15 0298e1f
Be kinder to memoise
