+++ man/dependency_profile.Rd
@@ -32,12 +32,6 @@ con <- make(my_plan) # Run the project, build the targets.
 # Get some example dependency profiles of targets.
 dependency_profile("small", config = con)
 dependency_profile("'report.md'", config = con)
-clean(large) # Clear out most of the information about the target 'large'.
-# The dependency hashes are still there.
-dependency_profile("large", config = con)
-# Should agree with `$hashes_of_dependencies`.
-con$cache$get_hash("simulate",
-  namespace = "kernels")
 })
 }
 }
