### 2017-12-15 9f05dc0
Steamline visualization internals
