+++ NAMESPACE
@@ -141,6 +141,7 @@ importFrom(igraph,adjacent_vertices)
 importFrom(igraph,degree)
 importFrom(igraph,delete_vertices)
 importFrom(igraph,edge)
+importFrom(igraph,induced_subgraph)
 importFrom(igraph,is_dag)
 importFrom(igraph,make_empty_graph)
 importFrom(igraph,plot.igraph)
