### 2017-12-15 87efb90
Rearrange graph/vis testing
