### 2017-12-15 53aa5b5
Remove superfluous functions
