### 2017-12-15 762d12f
Clean up a cache internal function
