+++ R/hash.R
@@ -206,20 +206,21 @@ default_long_hash_algo <- function(cache = NULL) {
 }
 
 check_storr_short_hash <- function(cache, chosen_algo){
-  if ("storr" %in% class(cache)){
-    true_algo <- cache$driver$hash_algorithm
-    if (true_algo != chosen_algo){
-      warning(
-        "The storr-based cache actually uses ", true_algo,
-        " for the short hash algorithm, but ", chosen_algo,
-        " was also supplied. Reverting to ", true_algo, ".",
-        call. = FALSE
-      )
-      cache$set(
-        key = "short_hash_algo",
-        value = true_algo,
-        namespace = "config"
-      )
-    }
+  if (!inherits(cache, "storr")){
+    return()
+  }
+  true_algo <- cache$driver$hash_algorithm
+  if (!identical(true_algo, chosen_algo)){
+    warning(
+      "The storr-based cache actually uses ", true_algo,
+      " for the short hash algorithm, but ", chosen_algo,
+      " was also supplied. Reverting to ", true_algo, ".",
+      call. = FALSE
+    )
+    cache$set(
+      key = "short_hash_algo",
+      value = true_algo,
+      namespace = "config"
+    )
   }
 }
