### 2017-12-19 f1c5c1a
Update acknowledgements
