+++ inst/examples/gsp/report.Rmd
@@ -9,7 +9,11 @@ suppressPackageStartupMessages(library(drake))
 suppressPackageStartupMessages(library(Ecdat))
 ```
 
-The following data analysis workflow a toy example, but it is based on real data, and it demonstrates the utility of `drake` in real-world projects. If you wish to tinker with it yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`. This `report.Rmd` file will not knit on its own until `run.R` has generated enough output.
+The following data analysis workflow shows off `drake`'s ability to generate lots of reproducibly-tracked tasks with ease. The same technique would be cumbersome, even intractable, with [GNU Make](https://www.gnu.org/software/make/).
+
+If you wish to tinker with the example yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`.
+
+This `report.Rmd` file will not knit on its own until `run.R` has generated enough output.
 
 # Objective and methods
 
