### 2017-12-19 367cd82
Repair gsp vignette
