### 2017-12-19 7031c7e
Change comments in gsp run.R
