### 2017-12-19 cf3d593
Begin stub of a new example and vignette
Supports ropensci/onboarding#156.
This new example (example-package.Rmd) will explore
some trends in R package downloads over time.
It will feature a changing project with changing data,
and drake will make sure that things never need to
restart from scratch.
