+++ inst/examples/packages/clean.R
@@ -1,0 +1,2 @@
+drake::clean(destroy = TRUE)
+unlink(c("report.html", "figure"), recursive = TRUE)
