+++ inst/examples/packages/report.Rmd
@@ -1,0 +1,103 @@
+---
+title: "A practical example of drake in action"
+author: "William Michael Landau"
+output: html_document
+---
+
+```{r setup, echo = FALSE}
+suppressPackageStartupMessages(library(drake))
+suppressPackageStartupMessages(library(Ecdat))
+```
+
+The following data analysis workflow shows off `drake`'s ability to generate lots of reproducibly-tracked tasks with ease. The same technique would be cumbersome, even intractable, with [GNU Make](https://www.gnu.org/software/make/).
+
+If you wish to tinker with the example yourself, you can generate the [underlying code files](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/application) with `drake_example("application")` and then step through `run.R`.
+
+This `report.Rmd` file will not knit on its own until `run.R` has generated enough output.
+
+# Objective and methods
+
+The goal is to search for factors closely associated with the productivity of states in the USA around the 1970s and 1980s. For the sake of simplicity, we use gross state product as a metric of productivity, and we restrict ourselves to multiple linear regression models with three variables. For each of the 84 possible models, we fit the data and then evaluate the root mean squared prediction error (RMSPE).
+
+$$
+\begin{aligned}
+\text{RMSPE} = \sqrt{(\text{y} - \widehat{y})^T(y - \widehat{y})}
+\end{aligned}
+$$
+Here, $y$ is the vector of observed gross state products in the data, and $\widehat{y}$ is the vector of predicted gross state products under one of the models. We take the best variables to be the triplet in the model with the lowest RMSPE.
+
+# Data
+
+The `Produc` dataset from the [Ecdat package](https://cran.r-project.org/package=Ecdat) contains data on the Gross State Product from 1970 to 1986. Each row is a single observation on a single state for a single year. The dataset has the following variables as columns. See the references later in this report for more details.
+
+- `gsp`: gross state product.
+- `state`: the state.
+- `year`: the year.
+- `pcap`: private capital stock.
+- `hwy`: highway and streets.
+- `water`: water and sewer facilities.
+- `util`: other public buildings and structures.
+- `pc`: public capital.
+- `emp`: labor input measured by the employment in non-agricultural payrolls.
+- `unemp`: state unemployment rate.
+
+```{r masterdata}
+library(Ecdat)
+data(Produc)
+head(Produc)
+```
+
+# Results
+
+Here are the root mean squared prediction errors of all the models.
+
+```{r rmspeplot}
+results <- readd(rmspe)
+
+loadd(plot_rmspe)
+
+library(ggplot2)
+plot_rmspe(rmspe = results)
+```
+
+And here are the best models. The best variables are in the top row under `X1`, `X2`, and `X3`.
+
+```{r bestmodels}
+head(results[order(results$rmspe, decreasing = FALSE), ])
+```
+
+# Comparison with Make
+
+If we were using [Make](https://www.gnu.org/software/make/) for this project, the analogous [Makefile](https://www.gnu.org/software/make/) would look something like this sketch.
+
+<pre><code>#!/bin/bash
+
+models = model_state_year_pcap model_state_year_hwy ... # 84 of these
+
+model_%
+    Rscript -e 'saveRDS(lm(...), ...)'
+
+rmspe_%: model_%
+    Rscript -e 'saveRDS(get_rmspe(...), ...)'
+
+rmspe: rmspe_%
+    Rscript -e 'saveRDS(rbind(...), ...)'
+
+rmspe.pdf: rmspe
+    Rscript -e 'ggsave(plot_rmspe(rmspe), "rmspe.pdf")'
+
+report.md: report.Rmd
+    Rscript -e 'knitr::knit("report.Rmd")'
+</code></pre>
+
[…]

[…]
