### 2017-12-20 c94cf22
Repair graphing parallel stages
