### 2017-12-20 2d695f8
Complete the "packages" example
