### 2017-12-20 838afe7
Add packages.html to images
Prepare for an example-packages.Rmd vignette.
