### 2017-12-20 8db42c7
Add cranlogs to suggests
