+++ vignettes/example-packages.Rmd
@@ -12,6 +12,10 @@ vignette: >
 ```{r setup, include = FALSE}
 suppressMessages(suppressWarnings(library(drake)))
 suppressMessages(suppressWarnings(library(magrittr)))
+suppressMessages(suppressWarnings(library(cranlogs)))
+suppressMessages(suppressWarnings(library(knitr)))
+suppressMessages(suppressWarnings(library(ggplot2)))
+suppressMessages(suppressWarnings(library(plyr)))
 clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
 knitr::opts_chunk$set(
@@ -19,34 +23,41 @@ knitr::opts_chunk$set(
   error = TRUE,
   warning = TRUE
 )
+reportfile <- file.path("examples", "packages", "report.Rmd") %>%
+  system.file(package = "drake", mustWork = TRUE)
+file.copy(reportfile, getwd())
 ```
 
-# This analysis explores some trends in R package downloads over time.
-# The data change often, but drake makes sure
-# the project does not restart from scratch.
-#
-# Also see the example-packages.Rmd vignette,
-# https://github.com/wlandau-lilly/drake/blob/master/vignettes/example-packages.Rmd
+This data analysis project explores some trends in R package downloads over time.
+The data change often, and `drake` makes sure the project does not restart from scratch
+on a typical runthrough. 
 
+First, we load the required packages. `Drake` knows about the packages you install and load.
+
+```{r pkgspkgs}
 library(drake)
 library(cranlogs)
 library(knitr)
 library(ggplot2)
 library(plyr)
+```
 
-# We want to explore the daily downloads from these packages.
+We want to explore the daily downloads from these packages.
 
+```{r packagelist}
 package_list <- c(
   "knitr",
   "Rcpp",
   "ggplot2"
 )
+```
 
-# We plan to use the cranlogs package.
-# The data frames `older` and `recent` will
-# contain the number of daily downloads for each package
-# from the RStudio CRAN mirror.
+We plan to use the [cranlogs package](https://github.com/metacran/cranlogs).
+The data frames `older` and `recent` will
+contain the number of daily downloads for each package
+from the RStudio CRAN mirror.
 
+```{r datadataplan}
 data_plan <- drake_plan(
   recent = cran_downloads(packages = package_list, when = "last-month"),
   older = cran_downloads(
@@ -57,17 +68,26 @@ data_plan <- drake_plan(
   strings_in_dots = "literals"
 )
 
-# We want to summarize each set of
-# the download statistics a couple different ways.
+data_plan
+```
 
+We want to summarize each set of
+download statistics a couple different ways.
+
+```{r outputtypespackages}
 output_types <- drake_plan(
   averages = make_my_table(dataset__),
   plot = make_my_plot(dataset__)
 )
 
-# We need to define functions to summarize
-# and plot the data.
+output_types
+```
 
+
+We need to define functions to summarize
+and plot the data.
+
+```{r summplotdatapackages}
 make_my_table <- function(downloads){
   ddply(downloads, "package", function(package_downloads){
     data.frame(mean_downloads = mean(package_downloads$count))
@@ -78,67 +98,107 @@ make_my_plot <- function(downloads){
   ggplot(downloads) +
     geom_line(aes(x = date, y = count, group = package, color = package))
 }
+```
 
-# Below, the symbols `recent` and `older`
-# each substitute in for the `dataset__` wildcard.
-# Thus, `output_plan` has four rows.
+Below, the targets `recent` and `older`
+each take turns substituting the `dataset__` wildcard.
+Thus, `output_plan` has four rows.
 
+```{r outputplanpackages}
 output_plan <- plan_analyses(
   plan = output_types,
[…]
