### 2017-12-20 5142a85
Tweak knitr chunk opts
