### 2017-12-20 661c00a
Explain a way around triggers
