### 2017-12-20 91f660b
Add an entry for the packages example
