### 2017-12-20 642f18e
Include logo and correct graph link
