### 2017-12-20 ddb5529
Change order of pkgs
