+++ vignettes/example-packages.Rmd
@@ -13,10 +13,10 @@ vignette: >
 
 ```{r setup, include = FALSE}
 suppressMessages(suppressWarnings(library(drake)))
-suppressMessages(suppressWarnings(library(magrittr)))
 suppressMessages(suppressWarnings(library(cranlogs)))
-suppressMessages(suppressWarnings(library(knitr)))
 suppressMessages(suppressWarnings(library(ggplot2)))
+suppressMessages(suppressWarnings(library(knitr)))
+suppressMessages(suppressWarnings(library(magrittr)))
 suppressMessages(suppressWarnings(library(plyr)))
 clean(destroy = TRUE, verbose = FALSE)
 unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
@@ -39,8 +39,8 @@ First, we load the required packages. `Drake` knows about the packages you insta
 ```{r pkgspkgs}
 library(drake)
 library(cranlogs)
-library(knitr)
 library(ggplot2)
+library(knitr)
 library(plyr)
 ```
 
