### 2017-12-20 #156 @wlandau-lilly
I have been thinking more about `drake`'s accessibility to new users, especially @benmarwick's comment that it is difficult to relate to the vignettes. I did some expanding and refactoring, and as of now, two of the vignettes concentrate on more down-to-earth examples. Both run quickly to avoid bottlenecking the package quality checks, and the statistical methodology is elementary to keep things clear and simple. Each is of these vignettes is paired with a set of example code files (available via `drake_example("packages")` and `drake_example("gsp")`).

1. [example-packages.Rmd](https://wlandau-lilly.github.io/drake/articles/example-packages.html) is a tiny analysis of R package download trends using [cranlogs](https://github.com/metacran/cranlogs). In this situation, some of the input data needs to be refreshed every day, and `drake` brings the project up to date without restarting everything from scratch.
2. [example-gsp.Rmd](https://wlandau-lilly.github.io/drake/articles/example-gsp.html) focuses on a real econometrics dataset. The goal here is to show that `drake` easily scales up with the number of targets, but [GNU Make](https://www.gnu.org/software/make/) does not.
