### 2017-12-20 74b09f1
Add a wordlist and lang for spell checking
