+++ DESCRIPTION
@@ -68,4 +68,5 @@ Suggests:
 VignetteBuilder: knitr
 URL: https://github.com/wlandau-lilly/drake
 BugReports: https://github.com/wlandau-lilly/drake/issues
+Language: en-US
 RoxygenNote: 6.0.1
