### 2017-12-20 2e992ff
Print out whole_plan once more time
