### 2017-12-20 #156 @maelle
Cool -- I was also wondering (and have not checked myself sorry) whether the R podcast episode about `drake` is listed in the documentation? It provides some useful context&history.
