### 2017-12-20 #156 @wlandau-lilly
Good question, @maelle. I have not mentioned it in the documentation, and I am still trying to decide whether I will.
