### 2017-12-20 e644038
Reword packages example
