+++ inst/examples/packages/clean.R
@@ -1,2 +1,2 @@
 drake::clean(destroy = TRUE)
-unlink(c("report.html", "figure"), recursive = TRUE)
+unlink(c("report.html", "figure", "*.pdf"), recursive = TRUE)
