### 2017-12-21 d5a25c4
Fix CRAN checks for links
