### 2017-12-22 6e4cf3e
Mention R podcast
Supports ropensci/onboarding#156
