+++ README.md
@@ -139,7 +139,7 @@ install_github("wlandau-lilly/drake", build = TRUE)        # Development version
 
 # Documentation
 
-`Drake` has a [documentation website](https://wlandau-lilly.github.io/drake/). The [reference section](https://wlandau-lilly.github.io/drake/reference/index.html) organizes the functions by purpose,  and the articles are tutorials taken from the [package vignettes](https://github.com/wlandau-lilly/drake/tree/master/vignettes).
+`Drake` has a [documentation website](https://wlandau-lilly.github.io/drake/). The [reference section](https://wlandau-lilly.github.io/drake/reference/index.html) organizes the functions by purpose,  and the articles below are tutorials taken from the [package vignettes](https://github.com/wlandau-lilly/drake/tree/master/vignettes). For more context and history, listen to a full-length interview about `drake` in [episode 22 of the R Podcast](https://www.r-podcast.org/episode/022-diving-in-to-drake-with-will-landau/).
 
 - [Get started](https://wlandau-lilly.github.io/drake/articles/drake.html)
 - [Example: R package download trends](https://wlandau-lilly.github.io/drake/articles/example-packages.html)
