### 2017-12-22 259bc93
Try local svg in readme
