### 2017-12-22 e50284f
Use rawgit for README logo
