### 2018-01-01 #156 @jeroen
I'm asking some help from @HenrikBengtsson to review the parallelism components.
