### 2018-01-04 9e7a725
Elaborate on zombies for #184
