### 2018-01-04 fbeb8f7
Fix NEWS.md for pkgdown site
