### 2018-01-04 b29b41c
Call batchtools_* functions the proper way in examples (#185)
Still need to address nested parallelism.
