+++ inst/examples/sge/run.R
@@ -6,7 +6,11 @@ library(future.batchtools)
 library(drake)
 
 # Use future::plan(multicore) instead for a dry run.
-future::plan(batchtools_sge(template = "sge-simple.tmpl"))
+future::plan(
+  batchtools_sge,
+  template = "sge-simple.tmpl",
+  workers = 4
+)
 
 load_basic_example()
 make(
