+++ man/drake_batchtools_tmpl_file.Rd
@@ -39,7 +39,7 @@ drake_example("slurm") # Writes a new 'slurm' folder with more files.
 # Run the basic example with a
 # SLURM-powered parallel backend. Requires SLURM.
 library(future.batchtools)
-# future::plan(batchtools_slurm(template = "batchtools.slurm.tmpl")) # nolint
+# future::plan(batchtools_slurm, template = "batchtools.slurm.tmpl") # nolint
 # make(my_plan, parallelism = "future_lapply") # nolint
 })
 }
