### 2018-01-04 7cf1e5b
Rearrange nesting in parallelism vignette
