### 2018-01-04 #156 @jules32
Hi All!

Sorry for the silence over break! I've just had the chance to review drake, but it is just a partial review at this point, partly because I'm approaching the 3 hour mark, and partly because I am getting an error installing the current version from GitHub (note: I know this is because you've been working on it in my silence! You are probably already fixing it, but the error is below).

My partial review is focused on the README and website, which have greatly improved following @benmarwick's comments and how @wlandau-lilly addressed them. I will put my suggestions in the following comment. I can plan to coordinate further edits with on @jeron's timeline if that's easiest.

In December (for the original deadline) I did begin to look through drake v4.4.1.9000, and installed it from GitHub with no problem.

However, tonight, January 3 I installed `devtools::install_github("wlandau-lilly/drake", build = TRUE)`, which I believe is 4.4.1.9002 (not yet a GitHub release). I got the following install error due to `Ecdat`:

```
...
* preparing ‘drake’:
* creating vignettes ... ERROR
Quitting from lines 17-25 (example-gsp.Rmd)
Error: processing vignette 'example-gsp.Rmd' failed with diagnostics:
there is no package called 'Ecdat'
Execution halted
Installation failed: Command failed (1)
```
