### 2018-01-04 #156 @jules32
JL: I am reviewing `drake` as an interface reviewer, and from a beginner-friendly angle. I played around with `drake` prior to @benmarwick's comments above, and have also followed the work that @wlandau-lilly has put towards addressing them.

This review is in progress, as I've just focused on the documentation for now.

JL: This reads really well. I like the "Why use drake?" and "What gets done stays done" sections. I have a few wording suggestions for them at the end of this review. A big question I'm still left with after the README and the (super-helpful) website is what the functions that users will use as they get started, and over and over again. We have seen `make` and `drake_config`, but that's after a lot of other drake magic has gone on behind the scenes. Would it be possible/desirable to make a list (and linking to the website's reference page)? I know you're trying to cut down the README so some of this could go on the website's Get Started page perhaps.
JL: I don't think it's necessary to include options to install different tags/releases from GitHub. I think if a user wants that, they'll know where to look. To keep it simple, clear, and save space, I'd suggest:

```
# install drake from CRAN
install.packages("drake")

# Or the development version from GitHub:
# install.packages("devtools")
devtools::install_github("wlandau-lilly/drake", build = TRUE)
```

*JL: the following review is to be completed with drake v. 4.4.1.9002 or greater*

>##### For packages co-submitting to JOSS
>
> JL: from the above thread it seems like this is in progress so I will wait to evaluate this
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

*JL: the following review is to be completed with drake v. 4.4.1.9002 or greater*

---

In December, I was running `drake` on my machine, and had no problems running the examples. But I had trouble seeing how to move past the examples to see how drake could be used from the ground up, and how \*I* would use it. @wlandau-lilly has since made this a lot more clear, and I can see how this would be a good tool for more beginner-types to know about. These comments are kind of fine-tuning some of the work you've already done to help it resonate.

**README suggestions**

*What gets done stays done*

Having the Sisyphean loop example 1-4 is great: it's really helpful. It seems that with `drake`, this turns into something like:

1. Launch the code
1. Drake evaluates and rebuilds anything that has changed since the last run through

Is that true, and would that we worth itemizing like that in the README? And then here are some suggestions for commenting the example, which is a bit obvious but can be a bit easier to follow:

```
Example: `my_plan` lists 15 targets (analysis steps that have specific commands), and `drake` will evaluate them with its `make` function.

# Load drake's basic example and examine my_plan's analysis
library(drake) # install.packages("drake")
load_basic_example(verbose = FALSE)
head(my_plan)

##                    target                                      command
## 1             'report.md'             knit('report.Rmd', quiet = TRUE)
## 2                   small                                  simulate(5)
## 3                   large                                 simulate(50)
## 4       regression1_small                                  reg1(small)
## 5       regression1_large                                  reg1(large)
## 6       regression2_small                                  reg2(small)

# First round: drake builds all 15 targets.
make(my_plan)

## target large
## target small
## target regression1_large
## target regression1_small
## target regression2_large
## target regression2_small
## target coef_regression1_large
## target coef_regression1_small
## target coef_regression2_large
## target coef_regression2_small
## target summ_regression1_large
## target summ_regression1_small
## target summ_regression2_large
## target summ_regression2_small
## target 'report.md'

# Then, you change the reg2 function; this will affect all regression2 targets.
reg2 <- function(d){
  d$x4 <- d$x ^ 4
  lm(y ~ x4, data = d)
}

# Second round: drake only builds what was updated.
make(my_plan)

## target regression2_large
## target regression2_small
## target coef_regression2_large
## target coef_regression2_small
## target summ_regression2_large
## target summ_regression2_small
## target 'report.md'

# And if nothing was updated, drake doesn't try to rebuild.
make(my_plan)

## All targets are already up to date.
```

**website: wlandau-lilly.github.io/drake**

Mostly, these are small things that might be fixed with v. 4.4.1.9002 or greater.

1. Something is wrong on the [Reference](https://wlandau-lilly.github.io/drake/reference/index.html) page; instead of a short descriptor of the function, it repeats the function name after the word "Function". This was also the case in the R help for v. 4.4.1.9000.

2. [Get Started](https://wlandau-lilly.github.io/drake/articles/drake.html) page:

  - I'd suggest starting this page with the "Where to Begin" part since the rest of it is on the homepage.
  - As you saw from my suggestion to the example code on the README, I really liked seeing the output from `load_basic_example(verbose = FALSE); my_plan` on this page. Seeing inside the `my_plan` variable is when I could really see myself using drake in my own workflow. It also lets us see reg2 before the example's second round.

3. The [packages example](https://wlandau-lilly.github.io/drake/articles/example-packages.html) also really hit drake home for me. I like seeing `drake_plan()` being used to to assign the targets and commands that we've seen in `load_basic_example()`.

4. Small thing, but when I look at the website on Chrome, in the tab it is labeled "Data Frames in R for Make — Drake". I know that that is the origin of the name, but I agree with @maelle's comment above that it's not intuitive (especially as a person who has too many tabs open all the time). If it is possible to have it say "drake" that would be awesome.
