### 2018-01-04 #156 @maelle
Happy New Year everyone in this thread!

@jules32 Thanks a lot for your review!

I'm trying to find a technical reviewer who'd have time to review `drake` before Jeroen's available again, I'll update this thread once I know more. Sorry for the long process, @wlandau-lilly , and thanks again for your work on the package since the submission!
