### 2018-01-04 16a4ea5
Address @jules32's comments about installation
- build = TRUE was excessive
- comment on the "Suggests:" packages

Supports ropensci/onboarding#156
