### 2018-01-04 #156 @wlandau-lilly
Thank you, @jules32! I am eager to address your comments, and I expect to have a proper response within the next couple days. For now, I will comment on the installation issue you mentioned, which I believe I addressed just now via https://github.com/wlandau-lilly/drake/commit/16a4ea5dcd8017dbe2c756e3e4d301565f3fd4ee. The [Ecdat package](https://cran.r-project.org/package=Ecdat) is only required for a vignette, so I listed it to the `Suggests:` field of the [DESCRIPTION](https://github.com/wlandau-lilly/drake/blob/master/DESCRIPTION). Rather than move it to `Imports:` or `Depends:`, I simply removed `build = TRUE` from the call to `install_github()`. (In hindsight, `build = TRUE` seems excessive anyway.) I also added special instructions for building the vignettes.

> - If you want to build [the vignettes](https://github.com/wlandau-lilly/drake/tree/master/vignettes) when you install the development version, you must
>     1. Set the `build` argument to `TRUE` in `install_github()`.
>     2. Install all the packages in the `Suggests:` field of the [DESCRIPTION file](https://github.com/wlandau-lilly/drake/blob/master/DESCRIPTION), including [cranlogs](https://cran.r-project.org/package=cranlogs) and [Ecdat](https://cran.r-project.org/package=Ecdat). All these packages are available through the [Comprehensive R Archive Network (CRAN)](https://cran.r-project.org), and you can install them with `install.packages()`.
