### 2018-01-04 b577e7c
Implement @jules32's suggestions about the README
Supports ropensci/onboarding#156
