### 2018-01-05 ce4addc
Avoid repeating the README in drake.Rmd
Supports ropensci/onboarding#156
