+++ vignettes/drake.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "drake"
-subtitle: "save time and stay reproducible"
+subtitle: "where to begin"
 author: "William Michael Landau"
 date: "`r Sys.Date()`"
 output:
@@ -28,102 +28,13 @@ knitr::opts_chunk$set(
 
 ![](logo-vignettes.png)
 
-# Why use drake?
-
-The `drake` R package is a workflow manager and computational engine for data science projects. Its primary objective is to keep your results up to date with the code and data they come from. When it runs your project, `drake` detects any pre-existing output and refreshes the pieces that are outdated or missing. Not every runthrough starts from scratch, and your final answers are reproducible. With a user-friendly R-focused interface, [comprehensive documentation](https://wlandau-lilly.github.io/drake/), and [extensive implicit parallel computing support](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd), `drake` surpasses the analogous functionality in similar tools such as  [Make](www.gnu.org/software/make/), [remake](https://github.com/richfitz/remake), [memoise](https://github.com/r-lib/memoise), and [knitr](https://github.com/yihui/knitr).
-
-# What gets done stays done.
-
-Too many data science projects follow a [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus):
-
-1. Launch the code.
-2. Wait for it to finish.
-3. Discover an issue.
-4. Restart from scratch.
-
-`Drake` is more discerning. When you update your code or data, it figures out exactly what needs building, and it builds it in the correct order. The next runthrough is shorter, and your progress is steadier and smoother.
-
-```{r drakemake}
-library(drake)
-load_basic_example()
-
-# First round.
-make(my_plan)
-
-# Change some code.
-reg2 <- function(d){
-  d$x4 <- d$x ^ 4
-  lm(y ~ x4, data = d)
-}
-
-# Less work for the second round.
-make(my_plan)
-
-# No work if nothing changed.
-make(my_plan)
-```
-
-# Stay reproducible.
-
-The R community likes to emphasize reproducibility, which one could interpret to mean [scientific replicability](https://en.wikipedia.org/wiki/Replication_crisis), literate programming with [knitr](https://yihui.name/knitr/), or version control with [git](https://git-scm.com/book/en/v2/Getting-Started-About-Version-Control). But internal consistency is important too. Reproducibility carries the promise that your output matches the code and data it came from. Ordinarily, you might have to rerun everything from scratch just to be sure. But with `drake`, you can just check that all your targets are up to date.
-
-```{r drakemake2}
-make(my_plan)
-
-config <- drake_config(my_plan, verbose = FALSE)
-outdated(config)
-```
-
-# Aggressively scale up.
-
-Not every project can complete in a single R session on your laptop. Some projects need more speed or computing power. Some require a few local processor cores, and some need large high-performance computing systems. But parallel computing is hard. Your tables and figures depend on your analysis results, and your analyses depend on your datasets, so some tasks must finish before others even begin. But `drake` knows what to do. Parallelism is implicit and automatic. See the [parallelism vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/parallelism.Rmd) for all the details.
-
-```{r implicitparallel, eval = FALSE}
-# Use the spare cores on your local machine.
-make(my_plan, jobs = 4)
-
-# Scale up to a supercomputer.
-drake_batchtools_tmpl_file("slurm") # https://slurm.schedmd.com/
-library(future.batchtools)
-future::plan(
-  batchtools_slurm,
-  template = "batchtools.slurm.tmpl",
-  workers = 100
-)
-make(my_plan, parallelism = "future_lapply")
[…]
