### 2018-01-05 42d4a7a
Switch steps in installation instructions
proper order
