### 2018-01-05 b6ecdef
Use a list instead of a table for function list
The table rendered poorly for the pkgdown site.
