### 2018-01-05 1f7cb75
Add more descriptive titles in the help files.
Address @jules32's comments about the repeated "Function X" titles.
Tedious, but worth it.
Supports ropensci/onboarding#156.
