+++ R/plot_graph.R
@@ -1,7 +1,7 @@
-#' @title Function \code{vis_drake_graph}
-#' @description Plot the dependency structure of your drake_plan.
-#' IMPORTANT: you must be in the root directory of your project.
-#' To save time for repeated plotting, this function is divided into
+#' @title Show an interactive visual network representation
+#' of your drake project.
+#' @description To save time for repeated plotting,
+#' this function is divided into
 #' \code{\link{dataframes_graph}()} and \code{\link{render_drake_graph}()}.
 #' @export
 #' @aliases drake_graph
@@ -161,9 +161,11 @@ vis_drake_graph <- function(
 #' @export
 drake_graph <- vis_drake_graph
 
-#' @title Function \code{render_drake_graph}
-#' @description render a graph from the data frames
-#' generated by \code{\link{dataframes_graph}()}
+#' @title Render a visualization using the data frames
+#' generated by \code{\link{dataframes_graph}()}.
+#' @description This function is called inside
+#' \code{\link{vis_drake_graph}()}, which typical users
+#' call more often.
 #' @export
 #' @return A visNetwork graph.
 #'
