### 2018-01-05 9fe8eff
Tweak wording of "Why use drake?" section
