### 2018-01-05 #156 @wlandau-lilly
@maelle I was hoping to complete this before [rstudio::conf(2018)](https://www.rstudio.com/conference/), but I guess it can't be helped. And I realize that the winter holidays are not the right time to do work. Thank you for trying.
