### 2018-01-05 ae02966
Add comments to installation instructions
