### 2018-01-05 f76f6e2
Tweak wording of installation instructions
