### 2018-01-05 #156 @wlandau-lilly
# Response to @jules32's partial review from January 3

Your feedback and advice are extremely helpful, and your encouragement is gratifying. I want to reach as many new users as possible, so I care a lot about outside feedback on the documentation. Thank you for your efforts.

I agree with all your suggestions from January 3, and I believe I addressed them all in https://github.com/wlandau-lilly/drake/commit/16a4ea5dcd8017dbe2c756e3e4d301565f3fd4ee through https://github.com/wlandau-lilly/drake/commit/f76f6e2127ac8003a8c5417a667ae9b9141ae15a. Please let me know if you think I missed anything.

> A big question I'm still left with after the README and the (super-helpful) website is what the functions that users will use as they get started, and over and over again. We have seen make and drake_config, but that's after a lot of other drake magic has gone on behind the scenes. Would it be possible/desirable to make a list (and linking to the website's reference page)? I know you're trying to cut down the README so some of this could go on the website's Get Started page perhaps.

The [Documentation section of the README](https://github.com/wlandau-lilly/drake/tree/6678e28e3ae1729eb9f04d8014d9ad942880e008#documentation) now includes a list of the 10 most important functions, given roughly in the order I expect a user to call them, and it also refers to the [reference section of the documentation website](https://wlandau-lilly.github.io/drake/reference/index.html). The [documentation website's main page](https://wlandau-lilly.github.io/drake/#documentation) has an identical section.

> JL: I don't think it's necessary to include options to install different tags/releases from GitHub. I think if a user wants that, they'll know where to look.

Done. Like the rest of the changes to the README, this change is also reflected on the [documentation website](https://wlandau-lilly.github.io).

> **For packages co-submitting to JOSS**
>
> JL: from the above thread it seems like this is in progress so I will wait to evaluate this

I have just begun my company's scientific disclosure process to release my JOSS manuscript. It is essentially the ["Why use drake?" section of the README](https://github.com/wlandau-lilly/drake/tree/ae02966cc85963b3f0fb01d98721301631d92d00#why-use-drake) with some added metadata and references.

>Having the Sisyphean loop example 1-4 is great: it's really helpful. It seems that with drake, this turns into something like:
>
> 1. Launch the code
> 2. Drake evaluates and rebuilds anything that has changed since the last run through
>
> Is that true, and would that we worth itemizing like that in the README?

Yes, absolutely. The continuity of the format highlights the contrast between approaches. I made the change.

> And then here are some suggestions for commenting the example, which is a bit obvious but can be a bit easier to follow: ...

Narration is definitely helpful here. I have added very similar comments to the code.

> Something is wrong on the Reference page; instead of a short descriptor of the function, it repeats the function name after the word "Function". This was also the case in the R help for v. 4.4.1.9000.

That was just my own laziness. I went back and changed all the titles to be informative. The new reference page and help files are now improved.

> Get Started page:
>
> - I'd suggest starting this page with the "Where to Begin" part since the rest of it is on the homepage.

Come to think of it, the "Where to begin" section is the only unique part of the [Get Started page](https://wlandau-lilly.github.io/drake/articles/drake.html). I have now removed all the sections that were repeated on the [main page](https://wlandau-lilly.github.io/drake) and README. You can also see the changes in the [underlying drake.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/f76f6e2127ac8003a8c5417a667ae9b9141ae15a/vignettes/drake.Rmd).

> Small thing, but when I look at the website on Chrome, in the tab it is labeled "Data Frames in R for Make — Drake". I know that that is the origin of the name, but I agree with @maelle's comment above that it's not intuitive (especially as a person who has too many tabs open all the time). If it is possible to have it say "drake" that would be awesome.

Done. The title you saw was automatically generated by [pkgdown](https://github.com/hadley/pkgdown). The change required [some more post-processing](https://github.com/wlandau-lilly/drake/blob/9a1deb3f1a48d65e62cf606f1e6d8ab524bd3bfb/build-site.R#L14), but I agree that it was necessary.
