### 2018-01-05 b477c25
some typos
