+++ man/evaluate_plan.Rd
@@ -10,7 +10,7 @@ evaluate_plan(plan, rules = NULL, wildcard = NULL, values = NULL,
 }
 \arguments{
 \item{plan}{workflow plan data frame, similar to one produced by
-\code{ink{drake_plan}}}
+\code{\link{drake_plan}}}
 
 \item{rules}{Named list with wildcards as names and vectors of
 replacements
@@ -40,7 +40,7 @@ A workflow plan data frame with the wildcards evaluated.
 The commands in workflow plan data frames can have
 wildcard symbols that can stand for datasets, parameters, function
 arguments, etc. These wildcards can be evaluated over a set of
-possible values using \code{evaluate}.
+possible values using \code{evaluate_plan}.
 }
 \details{
 Specify a single wildcard with the \code{wildcard}
