### 2018-01-05 9520cd8
Address #187
Convert `values` to a character vector
inside expand_plan() and evaluate_plan()
