### 2018-01-08 1c9c67b
Add JOSS submission docs: paper.md, paper.bib
Compile the pdf with `pandoc --bibliography paper.bib paper.md -o paper.pdf`
