+++ paper.bib
@@ -1,0 +1,43 @@
+@misc{
+  drake,
+  author = {William Michael Landau},
+  title = {{drake: an R-focused pipeline toolkit for reproducibility and high-performance computing}},
+  year = {2018},
+  howpublished = {https://CRAN.R-project.org/package=drake}
+}
+
+@misc{
+  knitr,
+  author = {Yihui Xie},
+  year = 2017,
+  title = {{knitr: A General-Purpose Package for Dynamic Report Generation in R}},
+  note = {R package version 1.17},
+  howpublished = {https://CRAN.R-project.org/package=knitr}
+}
+
+@book{
+  Make,
+  title = {{GNU Make, Version 3.77}},
+  publisher = {{Free Software Foundation}},
+  author = {Richard Stallman},
+  isbn = {1882114809},
+  year = {1998}
+}
+
+@misc{
+  memoise,
+  title = {{memoise: Memoisation of Functions}},
+  author = {Hadley Wickham and Jim Hester and Kirill Müller and Daniel Cook},
+  year = {2017},
+  note = {R package version 1.1.0},
+  howpublished = {https://CRAN.R-project.org/package=memoise}
+}
+
+@misc{
+  remake,
+  title = {{remake: Make-like build management}},
+  author = {Rich FitzJohn},
+  note = {R package version 0.3.0},
+  year = {2017},
+  howpublished = {GitHub repository, https://github.com/richfitz/remake}
+}
