### 2018-01-08 #156 @wlandau-lilly
FYI: I just added the JOSS submission docs in https://github.com/wlandau-lilly/drake/commit/1c9c67b6495330d5674315bc019e5473d0e7a4ab. You can compile the pdf with `pandoc --bibliography paper.bib paper.md -o paper.pdf`. Since a DOI for a GitHub repo needs a tag/release, I will wait to generate one for `drake` until all the reviewers here approve the rest of the package.
