### 2018-01-08 #156 @wlandau-lilly
Also, I would like the DOI to correspond to a version of `drake` whose logo and links refer to https://github.com/ropensci/drake (see also https://github.com/wlandau-lilly/drake/pull/176) and the CRAN release of version 5.0.0.
