### 2018-01-08 #156 @maelle
@jules32 recruited a technical reviewer! :tada: Thanks a ton! Welcome, @gothub, and thanks a lot! Is 2018/01/22 a good deadline for your review?

Here is [our review template](https://github.com/ropensci/onboarding/blob/master/reviewer_template.md) and our [reviewing guide](https://github.com/ropensci/onboarding/blob/master/reviewing_guide.md)
