### 2018-01-09 #156 @wlandau-lilly
Yes, thank you @gothub!

By the way: @benmarwick, do you think I sufficiently addressed your review? I made several improvements since my first response on December 9.
