### 2018-01-10 2ca610f
Try to add footer template for pkgdown site
