### 2018-01-10 ea0af78
Move pkgdown templates
