### 2018-01-10 81f0b72
Tweaks to README wording
