### 2018-01-10 5c40983
Bring all pkgdown files outside docs/
