### 2018-01-10 3ffacba
Simplify usage of drake_build()
No metadata needs to be supplied now.
Automatically computed if not already supplied.
