### 2018-01-10 ebdbe54
Contemplate drake_build() in the debug.Rmd vignette
