### 2018-01-11 bc69c7d
Update and clean up the README
