### 2018-01-16 8aa6b63
Implement drake_hash_log()
and use personal email instead of work email.
Never know if I'll need to change jobs.
