+++ R/hash.R
@@ -228,3 +228,117 @@ check_storr_short_hash <- function(cache, chosen_algo){
     )
   }
 }
+
+#' @title Get the hash keys of all the targets and imports
+#' in the cache.
+#' @description A hash is a fingerprint of an object's value.
+#' Together, the hash keys of all your targets and imports
+#' represent the state of your project.
+#' Use \code{drake_hash_log()} to generate a data frame
+#' with the hash keys of all the targets and imports
+#' stored in your cache.
+#' This function is particularly useful if you are
+#' storing your drake project in a version control repository.
+#' The cache has a lot of tiny files, so you should not put it
+#' under version control. Instead, save the output
+#' of \code{drake_hash_log()} as a text file after each \code{\link{make}()},
+#' and put the text file under version control.
+#' That way, you have a changelog of your project's results.
+#' See the examples below for details.
+#' @details Depending on your project's
+#' history, the targets may be different than the ones
+#' in your workflow plan data frame.
+#' Also, the keys depend on the short hash algorithm
+#' of your cache (default: \code{\link{default_short_hash_algo}()}).
+#' @seealso \code{\link{default_short_hash_algo}},
+#' \code{\link{default_long_hash_algo}},
+#' \code{\link{short_hash}},
+#' \code{\link{long_hash}},
+#' \code{\link{cached}},
+#' \code{\link{get_cache}}
+#' @export
+#' @return Data frame of the hash keys of the targets and imports
+#' in the cache
+#' @param cache drake cache. See \code{\link{new_cache}()}.
+#' If supplied, \code{path} and \code{search} are ignored.
+#' @param path Root directory of the drake project,
+#' or if \code{search} is \code{TRUE}, either the
+#' project root or a subdirectory of the project.
+#' @param search logical. If \code{TRUE}, search parent directories
+#' to find the nearest drake cache. Otherwise, look in the
+#' current working directory only.
+#' @param verbose whether to print console messages
+#' @param jobs number of jobs/workers for parallel processing
+#' @param targets_only logical, whether to output information
+#' only on the targets in your workflow plan data frame.
+#' If \code{targets_only} is \code{FALSE}, the output will
+#' include the hashes of both targets and imports. 
+#' @examples
+#' \dontrun{
+#' test_with_dir("Quarantine side effects.", {
+#' # Load drake's canonical example.
+#' load_basic_example()
+#' # Run the project, build all the targets.
+#' make(my_plan)
+#' # Get a data frame of all the hash keys.
+#' # If you want a changelog, be sure to do this after every make().
+#' hashes <- drake_hash_log()
+#' head(hashes)
+#' # Save the hash log as a flat text file.
+#' write.table(
+#'   x = hashes,
+#'   file = "hash_log.txt",
+#'   quote = FALSE,
+#'   row.names = FALSE
+#' )
+#' # At this point, put hash_log.txt under version control
+#' # (e.g. with 'git add hash_log.txt') alongside your code.
+#' # Now, every time you run your project, your commit history
+#' # of hash_lot.txt is a changelog of the project's results.
+#' # It shows which targets and imports changed on every commit.
+#' # It is extremely difficult to track your results this way
+#' # by putting the raw '.drake/' cache itself under version control.
+#' })
+#' }
+drake_hash_log <- function(
+  path = getwd(),
+  search = TRUE,
+  cache = drake::get_cache(path = path, search = search, verbose = verbose),
+  verbose = TRUE,
+  jobs = 1,
+  targets_only = FALSE
+){
+  if (is.null(cache)){
+    return(
+      data.frame(
[…]
