+++ NAMESPACE
@@ -44,6 +44,7 @@ export(drake_example)
 export(drake_examples)
 export(drake_gc)
 export(drake_graph)
+export(drake_hash_log)
 export(drake_meta)
 export(drake_palette)
 export(drake_plan)
