### 2018-01-16 3fd77d6
Reverse columns in drake_hash_log()
That way, if you print a flat text file,
the spacing just works out. It's visually pleasing.
