+++ R/hash.R
@@ -311,9 +311,9 @@ drake_hash_log <- function(
   if (is.null(cache)){
     return(
       data.frame(
-        name = character(0),
+        hash = character(0),
         type = character(0),
-        hash = character(0)
+        name = character(0)
       )
     )
   }
@@ -340,5 +340,5 @@ single_hash_log <- function(key, cache){
   )
   imported <- ifelse(is.na(imported), TRUE, imported)
   type <- ifelse(imported, "import", "target")
-  data.frame(name = key, type = type, hash = hash, stringsAsFactors = FALSE)
+  data.frame(hash = hash, type = type, name = key, stringsAsFactors = FALSE)
 }
