### 2018-01-16 0816381
Change cache logging
- drake_hash_log() is now drake_cache_log()
- added drake_cache_log_file() and make(..., cache_log_file = TRUE)
