+++ R/hash.R
@@ -229,33 +229,38 @@ check_storr_short_hash <- function(cache, chosen_algo){
   }
 }
 
-#' @title Get the hash keys of all the targets and imports
-#' in the cache.
-#' @description A hash is a fingerprint of an object's value.
+#' @title Get a table that represents the state of the cache.
+#' @description
+#' This functionality is like
+#' \code{\link{make}(..., cache_log_file = TRUE)},
+#' but separated and more customizable. Hopefully, this functionality
+#' is a step toward better data versioning tools.
+#' @details A hash is a fingerprint of an object's value.
 #' Together, the hash keys of all your targets and imports
 #' represent the state of your project.
-#' Use \code{drake_hash_log()} to generate a data frame
+#' Use \code{drake_cache_log()} to generate a data frame
 #' with the hash keys of all the targets and imports
 #' stored in your cache.
 #' This function is particularly useful if you are
 #' storing your drake project in a version control repository.
 #' The cache has a lot of tiny files, so you should not put it
 #' under version control. Instead, save the output
-#' of \code{drake_hash_log()} as a text file after each \code{\link{make}()},
+#' of \code{drake_cache_log()} as a text file after each \code{\link{make}()},
 #' and put the text file under version control.
 #' That way, you have a changelog of your project's results.
 #' See the examples below for details.
-#' @details Depending on your project's
+#' Depending on your project's
 #' history, the targets may be different than the ones
 #' in your workflow plan data frame.
 #' Also, the keys depend on the short hash algorithm
 #' of your cache (default: \code{\link{default_short_hash_algo}()}).
-#' @seealso \code{\link{default_short_hash_algo}},
+#' @seealso \code{\link{drake_cache_log_file}}
+#' \code{\link{cached}},
+#' \code{\link{get_cache}},
+#' \code{\link{default_short_hash_algo}},
 #' \code{\link{default_long_hash_algo}},
 #' \code{\link{short_hash}},
-#' \code{\link{long_hash}},
-#' \code{\link{cached}},
-#' \code{\link{get_cache}}
+#' \code{\link{long_hash}}
 #' @export
 #' @return Data frame of the hash keys of the targets and imports
 #' in the cache
@@ -282,17 +287,17 @@ check_storr_short_hash <- function(cache, chosen_algo){
 #' make(my_plan)
 #' # Get a data frame of all the hash keys.
 #' # If you want a changelog, be sure to do this after every make().
-#' hashes <- drake_hash_log()
-#' head(hashes)
+#' cache_log <- drake_cache_log()
+#' head(cache_log)
 #' # Save the hash log as a flat text file.
 #' write.table(
-#'   x = hashes,
-#'   file = "hash_log.txt",
+#'   x = cache_log,
+#'   file = "drake_cache.log",
 #'   quote = FALSE,
 #'   row.names = FALSE
 #' )
-#' # At this point, put hash_log.txt under version control
-#' # (e.g. with 'git add hash_log.txt') alongside your code.
+#' # At this point, put drake_cache.log under version control
+#' # (e.g. with 'git add drake_cache.log') alongside your code.
 #' # Now, every time you run your project, your commit history
 #' # of hash_lot.txt is a changelog of the project's results.
 #' # It shows which targets and imports changed on every commit.
@@ -300,7 +305,7 @@ check_storr_short_hash <- function(cache, chosen_algo){
 #' # by putting the raw '.drake/' cache itself under version control.
 #' })
 #' }
-drake_hash_log <- function(
+drake_cache_log <- function(
   path = getwd(),
   search = TRUE,
   cache = drake::get_cache(path = path, search = search, verbose = verbose),
[…]
