+++ tests/testthat/test-cache.R
@@ -41,7 +41,7 @@ test_with_dir("bad/corrupt caches, no progress", {
 })
 
 test_with_dir("non-existent caches", {
-  expect_equal(0, nrow(drake_hash_log()))
+  expect_equal(0, nrow(drake_cache_log()))
   check_storr_short_hash(NULL, "BLAH")
   expect_equal(find_cache(), NULL)
   expect_equal(find_project(), NULL)
@@ -119,13 +119,16 @@ test_with_dir("cache functions work", {
   config$session_info <- TRUE
   testrun(config)
 
-  # drake_hash_log() # nolint
-  all_hashes <- drake_hash_log()
-  some_hashes <- drake_hash_log(targets_only = TRUE)
+  # drake_cache_log() # nolint
+  all_hashes <- drake_cache_log()
+  some_hashes <- drake_cache_log(targets_only = TRUE)
   expect_equal(ncol(all_hashes), ncol(some_hashes))
   n_a <- nrow(all_hashes)
   n_s <- nrow(some_hashes)
   expect_true(n_a > n_s && n_s > 0)
+  expect_false(file.exists("log.txt"))
+  drake_cache_log_file(file = "log.txt")
+  expect_true(file.exists("log.txt"))
 
   # drake_gc() should not remove any important targets/imports.
   x <- cached()
