### 2018-01-16 8546cb2
Try to fix appveyor build until pillar release is built on CRAN for windows
