+++ appveyor.yml
@@ -46,4 +46,4 @@ environment:
     WARNINGS_ARE_ERRORS: 1
 
   matrix:
-    - R_VERSION: release
+    - R_VERSION: oldrel
