### 2018-01-16 a4846ca
Skip a test if pillar is not installed.
Should fix appveyor build errors
