### 2018-01-16 a1f7901
Move the cache log code to its own file
