+++ man/drake_cache_log_file.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hash.R
+% Please edit documentation in R/cache_log.R
 \name{drake_cache_log_file}
 \alias{drake_cache_log_file}
 \title{Generate a flat text log file
