+++ man/drake_cache_log.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/hash.R
+% Please edit documentation in R/cache_log.R
 \name{drake_cache_log}
 \alias{drake_cache_log}
 \title{Get a table that represents the state of the cache.}
