### 2018-01-16 e2bfe94
Drake has more than nine vignettes
