+++ vignettes/quickstart.Rmd
@@ -426,7 +426,11 @@ As you have seen with `reg2()`, `drake` reacts to changes in dependencies. In ot
 1. Commands in your workflow plan data frame.
 1. Global variables mentioned in the commands or imported functions.
 1. The output values of upstream targets.
-1. For [dynamic knitr reports](https://yihui.name/knitr/) (with `knit('report.Rmd')` or something similar as a command in your workflow plan data frame), targets and imports mentioned in calls to `readd()` and `loadd()` in the code chunks to be evaluated. `Drake` treats these targets and imports as dependencies of the compiled output target (say, `'report.md'`).
+1. For [dynamic knitr reports](https://yihui.name/knitr/) generated with the right commands, targets and imports mentioned in calls to `readd()` and `loadd()` in the code chunks to be evaluated. `Drake` treats these targets and imports as dependencies of the compiled output target (say, `'report.md'` or `report.html`). **To activate this feature, the command in your workflow plan data frame must call `knitr::knit()` or `rmarkdown::render()`. Examples of acceptable commands:
+    - `knit('report.Rmd')`
+    - `knitr::knit(input = 'report.Rmd', quiet = TRUE)`
+    - `render('report.Rmd')`
+    - `rmarkdown::('report.Rmd', output_file = "report.html")`
 
 To enhance reproducibility beyond the scope of drake, you might consider [packrat](https://rstudio.github.io/packrat) and [Docker](https://www.docker.com/). [Packrat](https://rstudio.github.io/packrat) creates a tightly-controlled local library of packages to extend the shelf life of your project. And with [Docker](https://www.docker.com/), you can execute your project on a [virtual machine](https://en.wikipedia.org/wiki/Virtual_machine) to ensure platform independence. Together, [packrat](https://rstudio.github.io/packrat) and [Docker](https://www.docker.com/) can help others reproduce your work even if they have different software and hardware.
 

[…]
