### 2018-01-17 7e811b0
Fix #189
Clarify the different roles of file imports, R environment imports,
and targets. Explanations are in the README, drake.Rmd vignette,
and quickstart vignette.
