### 2018-01-17 7a98925
Mention knit() and render() in the caution vignette
knit() should come from knitr and render() should come from rmarkdown.
Users should not define their own versions of these functions
or load different versions from other packages.
