### 2018-01-19 6a7fdab
Use testthat::skip_if_not_installed more often
