+++ tests/testthat/test-workflow-plan.R
@@ -84,11 +84,8 @@ test_with_dir("make() and check_plan() trim outer whitespace in target names", {
 })
 
 test_with_dir("make() plays nicely with tibbles", {
-  for (pkg in c("pillar", "tibble")){
-    if (!(pkg %in% rownames(installed.packages()))){
-      skip("tibble or a tibble dependency is not installed.")
-    }
-  }
+  skip_if_not_installed("pillar")
+  skip_if_not_installed("tibble")
   x <- tibble::tribble(~target, ~command, "nothing", 1)
   expect_silent(check_plan(x, verbose = FALSE))
   expect_silent(make(x, verbose = FALSE, session_info = FALSE))
