### 2018-01-19 b20c7d6
Fix #191
