### 2018-01-19 651888d
Update drake_plan() examples
