### 2018-01-19 c535a6b
Tweak an example
