### 2018-01-19 1013eab
Fix a new_cache() example
