### 2018-01-19 b803edf
Suppress drake_session() calls
