### 2018-01-19 67454c2
Fix typo in pkgdown.R
