### 2018-01-19 350eed4
Tweak pkgdown.R
