+++ pkgdown.R
@@ -27,6 +27,6 @@ to <- paste(
 )
 x <- gsub(pattern = from, replacement = to, x = x)
 
-writeLines(text = x, con = index_file)
+tmp <- writeLines(text = x, con = index_file)
 from <- list.files(site_dir, all.files = TRUE, full.names = TRUE)
-file.copy(from = from, to = dest, overwrite = TRUE, recursive = TRUE)
+tmp <- file.copy(from = from, to = dest, overwrite = TRUE, recursive = TRUE)
