### 2018-01-19 21285ff
Don't copy hidden files to docs
