+++ pkgdown.R
@@ -28,5 +28,5 @@ to <- paste(
 x <- gsub(pattern = from, replacement = to, x = x)
 
 tmp <- writeLines(text = x, con = index_file)
-from <- list.files(site_dir, all.files = TRUE, full.names = TRUE)
+from <- list.files(site_dir, full.names = TRUE)
 tmp <- file.copy(from = from, to = dest, overwrite = TRUE, recursive = TRUE)
