### 2018-01-19 51be0a5
Suppress return values in pkgdown.R
