### 2018-01-19 d050e10
Clean up .Rbuildignore
