### 2018-01-20 866e973
Fix #188
CRAN builds for 3.3.3 on Mac OS have not updated since Jan 9, but I am not going to hold my breath.
