+++ NAMESPACE
@@ -135,8 +135,8 @@ importFrom(crayon,make_style)
 importFrom(digest,digest)
 importFrom(evaluate,try_capture_stack)
 importFrom(formatR,tidy_source)
-importFrom(future,future_lapply)
 importFrom(future,plan)
+importFrom(future.apply,future_lapply)
 importFrom(grDevices,col2rgb)
 importFrom(grDevices,rgb)
 importFrom(igraph,V)
