### 2018-01-22 #156 @gothub
@maelle here is my review:

- [p] **A statement of need** clearly stating problems the software is designed to solve and its target audience in README
The purpose of the package is clearly stated. I am not a statistician but rather a software engineer who has worked with  environmental scientists and computational ecologists, so my evaluation is based on that perspective. Therefore I may not fully verify the claims made about the usefulness of this software for the target audience.
There were quite a few 'suggested' packages, but the software is well designed to use the standard R methods to detect and use optional packages if available.
The vignettes are clearly written and appear to describe the package functionality throughly.
The 'drake' vignette provides an overview of the package, but I believe that the 'quickstart' vignette is also required to read in order to understand the scope of the package. The link in the 'drake' vignette to the 'quickstart' points to the documentation website, although the equivalent 'quickstart' vignette is included with the R package.
I'm not familiar with the types of workflows that the package processes, where
dataset replicates and analyses are automatically generated ('quickstart' vignette, section 4), so I'm not able to fully evaluate these vignettes. The 'basic example' (i.e. from 'load_basic_example') is used throughout the documentation, so the basic setup of a 'template' workflow is not described fully in the vignettes or R command documentation. there are also other examples available that are more detailed and these can be listed with 'drake_examples()'. It would be very useful to provide a list of these examples in this vignette with a very brief description of the examples. I found these additional examples very helpful  (to be more helpful than then the basic example) as they used actual data with a processing goal explained.

The function documentation is complete and detailed with examples. Some of the examples are not run (by roxygen). Does devtools::check() evaluate these examples?
- [p] **Examples** for all exported functions in R Help that run successfully locally
As mentioned, some of the examples are not run (i.e. '\dontrun'), and I have not searched for and tested all 'not run' examples. I ran checks manually for some of the functions, without errors: drake_config(), drake_plan, make(), outdated()

>##### For packages co-submitting to JOSS
>
>- [ ] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [ ] **A short summary** describing the high-level functionality of the software
>- [ ] **Authors:**  A list of authors with their affiliations
>- [ ] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [ ] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

The package installed simply with 'install.packages()' and did not require any additional setup or
configuration.
- [p] **Functionality:** Any functional claims of the software been confirmed.
The parallel execution functionality appears to be powerful and is clearly explained but I have not
run tests to confirmed processing time gains when using the parallel backends vs serial processing. If the parallel execution functionality is important to rOpenSci for inclusion of this package, then it may be
useful to include a long running example for use as a benchmark with execution times shown for each  backend and for serial processing. I'm not suggesting that this is necessary, but this could be suggested  by the editor if deemed a requirement.
- [x] **Automated tests:** Unit tests cover essential functions of the package  and a reasonable range of inputs and conditions. All tests pass on the local machine.  The unit tests ran successfully using devtools::test(). The tests appear to be well  constructed and extensive.

---

The documentation makes several statements that this package supports "reproducibility", and even mentions that the  reproducibility supported is different than scientific replicability. The meaning of this term as it relates to the package functionality is not clear to me. In my experience at NCEAS this term generally means that source data, processing artifacts and documentation of a workflow are persisted and available so that they  can be obtained by a researcher who could replicate any processing products on their own, without any guidance from the original researcher. Does 'drake' somehow assist in this type of reproducibility?

An alternative use of reproducibility might be for an individual researcher (or team) to always expect consistent results even when any of their source data or processing steps have been modified. Is this the claim that the documentation is making? If a processing workflow is configured in drake then consistent results will always be generated? If that is the case, then how is this more reliable than running all scripts manually,
even though this may be inefficient? I'm not trying to play devil's advocate here, just trying to clarify and understand the statements made in the documentation. In working with environmental scientists and ecologists for several years, I have found that the definition of reproducibility is fairly specific, so elaborating on the use of this term for the package may bring more clarity.
