### 2018-01-22 wlandau-lilly/drake#193 @pat-s
How to add .R files to drake_plan()
Hi there,

maybe I overlooked something but I fail to properly add `.R` files to `drake_plan` so that the dependencies are detected.

Reprex:

```r
analyze <- drake_plan(
  test.md = knit('03_scripts/01_test.Rmd', quiet = TRUE),
  test2 = source('03_scripts/test2.R'),
```

And in `test2.R` I have the following:

```r
loadd(test.md)
```

However, when visualizing with

```r
config <- drake_config(analyze)
vis_drake_graph(config, width = "100%", height = "500px")
```

dependencies are not detected for the .R file.
If I use a .Rmd file, everything works as expected.

Tried a few different approaches but could not successfully add a .R script. Help :smile:
