### 2018-01-22 wlandau-lilly/drake#193 @wlandau-lilly
Great question, @pat-s! Questions like these are starting to come up a lot, and yours is the first in an FAQ I am starting.

The feature you want is designed for `knitr` source documents only, such as `*.Rmd` and `*.Rnw` files (although the file extension itself does not matter). `Drake` knows to analyze a file when it sees `knit()` or `render()` in the command, so nothing special happens if you use `source()` or your file has a `.R` extension.

Why not detect `loadd()`/`readd()` dependencies in R scripts `source()`'d from commands? Because with `drake`, the focus is on your R session, not your R scripts. The idea is to `source()` the files beforehand so that you have a bunch of functions and small data objects in your workspace. `Drake` treats these objects as "imports", and it detects and analyzes them once they are in your environment. For example:

```r
ls()

## character(0)

source("my_functions.R")

ls() # The simulate() function is defined in my_functions.R and treated as an import.

## [1] simulate
```

`Drake` looks inside the bodies of imported functions for dependencies.

```r
simulate

## function(n) {
##   data.frame(x = stats::rnorm(n), y = rpois(n, 1))
## }

deps(simulate)

## [1] "data.frame"   "rpois"        "stats::rnorm"
```

Now that you loaded your imports with `source("my_functions.R")`, you can `make()` your targets.

```r
analyze <- data.frame(
  target = c("test_dataset", "'test.md'"),
  command = c(
    "simulate(5)",
    "knit('03_scripts/01_test.Rmd', output = \"test.md\", quiet = TRUE)"
  )
)

analyze

##         target                                                          command
## 1 test_dataset                                                      simulate(5)
## 2    'test.md' knit('03_scripts/01_test.Rmd', output = "test.md", quiet = TRUE)

make(analyze)
```

where `01_test.Rmd` might look like this.

<pre><code>
---
title: "Test Report"
author: Patrick Schratz
output: html_document
---

This report depends on `test_dataset`.

```{r example_chunk}
drake::readd(test_dataset)
```
</code></pre>

A second `make(analyze)` will rebuild `test.md` if you have changed

- the `simulate()` function (except for changes to whitespace or comments) or
- either command in the `analyze` data frame (except for whitespace or comments) or
- the contents of `03_scripts/01_test.Rmd`.

Besides `knitr` reports, the `loadd()` and `readd()` functions are only meant for informally exploring your results. If they are embedded in your commands or imported functions, this could create a dangerous circularity in your workflow.

Does that help? Does this way of doing things meet your needs, or does your use case require you to `source()` an R file from a command?
