### 2018-01-22 wlandau-lilly/drake#193 @pat-s
Hi @wlandau-lilly, thanks for the quick and extensive answer!

So if I understand correctly, all analysis scripts should be Rmarkdown files while .R scripts should only contain functions that are used as imports and then analyzed for their dependencies?

> Does that help? Does this way of doing things meet your needs, or does your use case require you to source() an R file from a command?

Right, in my case, I have a mixture of `.Rmd` and `.R` files.
The `.Rmd` files contain all the preprocessing and EDA stuff including post-analysis of modeling results.

However, the modeling stuff is stored in `.R` scripts as these can easily be run from the command line (e.g. using `Rscript`), also avoiding all the `knitr` output/overhead.
The modeling scripts are running on a server in parallel.
While I use `.Rmd` reports a lot, I still find the good old `.R` scripts easier to handle on servers.

As I said, these files contain modeling code and therefore depend on the preprocessing `.Rmd` file.
Hence, I wanted to insert something like `load(preprocessing)` so that `drake` recognizes that the scripts depend on the preprocessing script. Subsequently, if something changes in the preprocessing script, it should first rerun `preprocessing.Rmd`.

Afterwards, there are again additional `.Rmd` reports (for the post analysis) that depend on the `.R` files.
So my current approach looks like this:

```r
analyze <- drake_plan(
  preprocessing.md = knit('03_scripts/01_preprocessing.Rmd', quiet = TRUE), # 1

  EDA.md = knit('03_scripts/02_EDA.Rmd', quiet = TRUE), # 2
  study_area.md = knit('03_scripts/03_study_area.Rmd', quiet = TRUE), # 2

  brt_sp_non = source('03_scripts/server/brt_sp_non.R'), # 3
  [...]
  brt_nsp_nsp = source('03_scripts/server/brt_nsp_nsp.R'), # 3

  cv_vis.md = knit('03_scripts/04_spcv_vis.Rmd', quiet = TRUE) # 4
)
```

where

* `#4` depends on `#3` and `#1`
* `#3` depends on `#1`
* `#2` depends on `#1`

Hope that does not confuse you too much :smile: .
I think there will be plenty of users out there using either only .R files or a mixture of .Rmd and .R for their analysis.
Would be great if both file types could be used together.
