### 2018-01-22 wlandau-lilly/drake#193 @tiernanmartin
Just wanted to chime in my support for accommodating a mixture of `.R` and `.Rmd` files.

Similar to @pat-s , the distinction for my projects is usually:

- `.R` for computationally expensive processes (in my use case these are operations on medium-size spatial data, some of which need to be run on a server)
- `.Rmd` for simple data operations, reporting, and documentation

 `drake` has helped me overcome a major pain point in my project workflow, so thanks a lot for your effort on this package! 👏
