### 2018-01-22 51da28f
Place more obvious references to the basic example.
From @gothub's suggestions in ropensci/onboarding#156,
I point out in the documentation that load_basic_example()
is paired with the example code generated by drake_example("basic").
Also, move future.R to future_lapply.R.
