+++ R/build_times.R
@@ -21,7 +21,7 @@
 #' \dontrun{
 #' test_with_dir("Quarantine side effects.", {
 #' # Show the build times for the basic example.
-#' load_basic_example() # Load the example.
+#' load_basic_example() # Get the code with drake_example("basic").
 #' make(my_plan) # Build all the targets.
 #' build_times() # Show how long it took to build each target.
 #' })
