+++ R/deprecate.R
@@ -56,7 +56,7 @@ as_file <- function(x){
 #' @examples
 #' \dontrun{
 #' test_with_dir("Quarantine side effects.", {
-#' load_basic_example() # Load the canonical example
+#' load_basic_example() # Get the code with drake_example("basic").
 #' # Choose future's multicore parallel backend.
 #' library(future)
 #' future::plan(multicore) # Instead of backend(). Avoid drake::plan().
