### 2018-01-22 8d5987f
List examples in the README and drake.Rmd vignette
Copied from https://github.com/wlandau-lilly/drake/tree/master/inst/examples.
Supports @gothub's review at ropensci/onboarding#156.
