+++ man/load_basic_example.Rd
@@ -33,10 +33,12 @@ non-back-compatible cache from a previous version of drake.}
 A \code{\link{drake_config}()} configuration list.
 }
 \description{
-For a thorough walkthrough of how to set up this example, see the
-quickstart vignette: \code{vignette('quickstart')}. Alternatively,
-call \code{\link{drake_example}('basic')} to generate an R script
-that builds up this example step by step.
+Use \code{\link{drake_example}('basic')} to get the code
+for the basic example. The included R script is a detailed,
+heavily-commented walkthrough. The quickstart vignette at
+\url{https://github.com/wlandau-lilly/drake/blob/master/vignettes/quickstart.Rmd} # nolint
+and \url{https://wlandau-lilly.github.io/drake/articles/quickstart.html}
+also walks through the basic example.
 }
 \details{
 This function also writes/overwrites
