### 2018-01-22 #156 @gothub
@wlandau-lilly the example descriptions in drake.Rmd and README look great.
