### 2018-01-22 1a70dd8
Emphasize reproducibility
Supports ropensci/onboarding#156
