### 2018-01-23 #156 @wlandau-lilly
# Response to @gothub's review

I sincerely appreciate your timely and thorough review! I found your feedback extremely helpful, and I have responded to specific comments below.

> There were quite a few 'suggested' packages

Yes, most of these optional packages support the vignettes, examples, custom caches, and advanced high-performance computing options.

> The link in the 'drake' vignette to the 'quickstart' points to the documentation website, although the equivalent 'quickstart' vignette is included with the R package.

I have tweaked the beginning of the [drake.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/master/vignettes/drake.Rmd) so it links to both the [quickstart vignette source](https://github.com/wlandau-lilly/drake/blob/master/vignettes/quickstart.Rmd) and the [rendered online version](https://wlandau-lilly.github.io/drake/articles/quickstart.html).

> The 'basic example' (i.e. from 'load_basic_example') is used throughout the documentation, so the basic setup of a 'template' workflow is not described fully in the vignettes or R command documentation.

I know it is buried very deep, but the example from `load_basic_example()` is actually paired with its own template workflow. The code files are [located here](https://github.com/wlandau-lilly/drake/tree/master/inst/examples/basic), [referenced here](https://github.com/wlandau-lilly/drake/blob/master/inst/examples/README.md), and generated by `drake_example("basic")`. The R script sets up the code and carries out a heavily-commented tutorial that walks through some of `drake`'s main features. I do agree that I should make all this more available, so I have added extra references to the vignettes and examples. Now, wherever `load_basic_example()` is called, there is a comment that says `# Get the code with drake_example("basic")`.

> there are also other examples available that are more detailed and these can be listed with 'drake_examples()'. It would be very useful to provide a list of these examples in this vignette with a very brief description of the examples. I found these additional examples very helpful (to be more helpful than then the basic example) as they used actual data with a processing goal explained.

There is a listing of examples [here](https://github.com/wlandau-lilly/drake/tree/master/inst/examples), but again, it is buried and far from obvious. I have replicated this list in a [new section](https://github.com/wlandau-lilly/drake/blob/master/vignettes/drake.Rmd#built-in-example-projects) of the `drake.Rmd` vignette, and the [documentation section of the README](https://github.com/wlandau-lilly/drake#documentation) now describes the beginner-oriented examples.

> The function documentation is complete and detailed with examples. Some of the examples are not run (by roxygen). Does devtools::check() evaluate these examples?

I chose to suppress many of the examples because some are time-consuming and others require the optional packages. And yes, `devtools::check()` runs the examples, and it skips the code enclosed in `\dontrun{}`. You can run all the code, even the code inside `\dontrun{}`, with the `run` argument of `devtools::run_examples()`. I routinely check `devtools::run_examples(run = TRUE)` and `devtools::run_examples(run = FALSE)`.

> The parallel execution functionality appears to be powerful and is clearly explained but I have not run tests to confirmed processing time gains when using the parallel backends vs serial processing. If the parallel execution functionality is important to rOpenSci for inclusion of this package, then it may be useful to include a long running example for use as a benchmark with execution times shown for each backend and for serial processing. I'm not suggesting that this is necessary, but this could be suggested by the editor if deemed a requirement.

You bring up an important point, and I do plan to do serious performance studies in the future. @maelle, do you think we need a long-running example for rOpenSci?

>The documentation makes several statements that this package supports "reproducibility", and even mentions that the reproducibility supported is different than scientific replicability. The meaning of this term as it relates to the package functionality is not clear to me. In my experience at NCEAS this term generally means that source data, processing artifacts and documentation of a workflow are persisted and available so that they can be obtained by a researcher who could replicate any processing products on their own, without any guidance from the original researcher. Does 'drake' somehow assist in this type of reproducibility?

I am so glad you brought this up. The whole point of `drake` is to verify that you can re-create the results from the code and data. If `make()` says `"All targets are already up to date"`, we have tangible evidence in favor of this kind of replicability. And if some targets are out of sync, `drake` skips all the up-to-date work and restores reproducibility fast. I just added a note to the end of the "Stay reproducible" section of the README to emphasize this point.

Other reproducibility advantages:

- The parallel computing is implicit and automatic. It is not tangled up with the code that actually does the analysis. This makes the code highly portable, re-scalable, and in my experience, much easier to read than otherwise.
- `vis_drake_graph()` provides big-picture guidance as to how the pieces of the analysis fit together.
- Re-creating the results should always be a simple matter of calling `make()`.

> An alternative use of reproducibility might be for an individual researcher (or team) to always expect consistent results even when any of their source data or processing steps have been modified. Is this the claim that the documentation is making?

Not exactly. `Drake` is about detecting when the data/processing has been modified, when it has not, and what to do in each eventuality.

> If a processing workflow is configured in drake then consistent results will always be generated? If that is the case, then how is this more reliable than running all scripts manually, even though this may be inefficient?

For large projects, running all the scripts manually is difficult enough for `drake` to make a huge practical difference. I strongly believe that checking internal consistency early and often is super important for reproducibility.

You brought up another kind of consistency that I find interesting. If we repeat the analysis from scratch a large number of times, will we always get the same final answer? For stochastic methods like MCMC, this is another important consideration for reproducibility. However, it is outside the scope of `drake`. `Drake` does make some effort to preserve random seeds (see wlandau-lilly/drake#56 and wlandau-lilly/drake#109) but that hardly scratches the surface of the issue.
