### 2018-01-23 #156 @HenrikBengtsson
My $0.02 comment on:

> "... include a long running example for use as a benchmark with execution times shown for each backend and for serial processing ..."

It'll be impossible to find a one-size-fits-all use case that will be representative and that can be used to produce benchmark statistics that can be used as a guidance for a random user.  Each user/project will have its own use case/pipeline to benchmark - I think that's what should be benchmarked. And after the first round of parallelization comes tweaking/optimization of how to best structure the parallelization of your analysis - and there are often multiple ways.  Factors such as amount of RAM, size of files processed, other tenants running on the same system at the same time, and so on, also come in play.

So, I would be careful making benchmark claims for large workflows, especially since such "results" tend to stick in the community as the ultimate facts/truth of a particular tool.   It's easier to benchmark a very specific problems, but Drake is in the business of being a generic workflow manager with automatic parallelization as a feature.
