### 2018-01-23 #156 @wlandau-lilly
Your comments are gratifying, Henrik. I still plan to study performance and optimization at some point, but for right now, you are making my job seem less daunting.

When it comes to parallel computing, I claim that `drake` processes and caches targets in parallel (in stages) and in the correct order. I also claim that the work deploys to the user's parallel backend of choice. With the function `in_progress()` (or equivalently, `progress()` or `vis_drake_graph()`) and your system's job monitoring software, I think this should all be straightforward to verify for one of the built-in examples if you think there is still a need. The basic example is probably easiest because of `load_basic_example()`. You may need to artificially slow down the workflow, maybe by appending "; sleep(60)" to all the commands or making the input datasets large enough.

`Drake`'s contribution is figuring out the parallel arrangement of targets automatically. For example, a machine learning expert need not worry that the neural nets depend on the training data. `Drake` knows that the data must be processed first, and it knows which neural nets can be computed at the same time. Beyond that, it borrows its power from other tools. The `storr`, `parallel`, `future`, `future.apply`, `future.batchtools`, and `batchtools` packages do the heavy lifting.

Late last year, @kendonB heavily used `drake` to deploy a *massive* project to his organization's [SLURM](https://slurm.schedmd.com/) cluster (via `make(..., parallelism = "future_lapply")`). He uncovered [a huge batch of performance issues](https://github.com/wlandau-lilly/drake/issues?utf8=%E2%9C%93&q=is%3Aissue+kendonB+), and `drake` is a much better package because of his on-target feedback. I fully expect more bottlenecks to arise as more power users use the package en earnest.
