### 2018-01-23 wlandau-lilly/drake#193 @wlandau-lilly
@pat-s and @tiernanmartin, thank you for clarifying. I think I can do a better job of explaining now.

> So if I understand correctly, all analysis scripts should be Rmarkdown files while .R scripts should only contain functions that are used as imports and then analyzed for their dependencies?

You don't need any analysis scripts. `Drake` tries to do away with them altogether, focusing on your R session rather than your files. To `drake`, your "script" is all the functions in your environment and the commands in your workflow plan data frame. How you load things into your environment is up to you. This is a very unusual, even disconcerting way of thinking, but I think it makes projects cleaner and smoother in the long run.

For example, instead of

```r
analyze <- drake_plan(
  ...
  preprocessing.md = knit('03_scripts/01_preprocessing.Rmd', quiet = TRUE),
  brt_sp_non = source('03_scripts/server/brt_sp_non.R'),
  ...
)
make(analyze)
```

you might try something like

```r
source("03_scripts/server/brt_sp_non.R")
analyze <- drake_plan(
  ...,
  preprocessed_data = preprocess_data(read_my_data('data.csv')),
  brt_sp_non = build_brt_sp_non(preprocessed_data),
  ...
)
make(analyze)
```

and define functions `preprocess_data()`, `read_my_data()`, and `build_brt_sp_non()` in "03_scripts/server/brt_sp_non.R". `Drake` knows that `brt_sp-non` depends on `preprocessed_data` and `build_brt_sp_non()` because these symbols are part of the command.

The commands in your workflow plan data frame are just arbitrary chunks of R code that return values. Other than the special accommodations for `knit()`, `render()`, and single-quoted file targets, you can treat commands like ordinary bits of R code.

Here is an example of a workflow that juggles a bunch of numbers.

```r
my_plan <- drake_plan(
  a = 1 + 1,
  b = {
    x <- pi + a
    y <- sqrt(x)
    rand <- rnorm(10, sd = y)
    mean(rand)
  },
  c = a - 5,
  d = c(b, c)
)
config <- drake_config(my_plan)
vis_drake_graph(config)
```

![math](https://user-images.githubusercontent.com/22958003/35253858-860ac6dc-ffb5-11e7-9b96-32c03dccd3ef.png)

```r
make(my_plan)
readd(d)

## [1]  1.886363 -3.000000
```

Now, if you rely on R Markdown reports for sharing results, you have the option to create one and then `loadd()`/`readd()` non-file targets or imports into code chunks. But that is up to you.
