### 2018-01-23 wlandau-lilly/drake#193 @wlandau-lilly
I forgot to mention: `drake` analyzes functions and commands for dependencies, and it knows those functions can be nested. For example, `drake` knows `g()` is nested inside `f()`.

```r
library(drake)

my_plan <- drake_plan(
  a = f(1 + 1)
)

f <- function(x){
  g(x + 1)
}

g <- function(x){
  x + sqrt(4)
}

config <- drake_config(my_plan)
vis_drake_graph(config)
```

![simple](https://user-images.githubusercontent.com/22958003/35254445-3925ee66-ffb8-11e7-815a-abc6a35edc73.png)

```r
make(my_plan)

## ...
## target a

readd(a)

## [1] 5

make(my_plan)

## ...
## All targets are already up to date.

# Let's change function g().

g <- function(x){
  x + 3
}

# f() depends on g(), and target `a` depends on f().
# So target `a` is out of date, and `make()` recomputes it.
make(my_plan)

## ...
## target a

# The value of `a` changed.
readd(a)

## [1] 6
```
