### 2018-01-23 #156 @wlandau-lilly
@gothub, I realized that I did not adequately address your point about the lack of a template workflow. It is difficult to have a template when `drake` focuses on a project's R session rather than its file system. You need a workflow plan data frame with targets and commands, plus any supporting functions and files, but otherwise, there is a lot of flexibility. The built-in examples are probably the best reference.
