### 2018-01-23 wlandau-lilly/drake#193 @pat-s
Thanks @wlandau-lilly, I tried your suggestion with the following setup now:

```r
source("03_scripts/01_preprocessing.R")
source("03_scripts/server/modeling_functions.R")
methods <- drake_plan(
  preprocessed_data = preprocess(
    pathogens = "/data/patrick/raw/survey_data/diseases240112_mod.csv",
    points = "/data/patrick/mod/survey_data/points_mod.shp",
    slope = "/data/patrick/mod/DEM/slope/slope_5m.tif",
    ph = new("GDALReadOnlyDataset", "/data/patrick/raw/ph_europe/ph_cacl2"),
    lithology = "/data/patrick/raw/lithology/CT_LITOLOGICO_25000_ETRS89.shp",
    hail = "/data/patrick/raw/hail/Prob_GAM_square_area.tif",
    elevation = "/data/patrick/mod/DEM/dem_5m.tif",
    soil = "/data/patrick/raw/soil/ISRIC_world_soil_information/TAXNWRB_250m_ll.tif",
    study_area = "/data/patrick/raw/boundaries/basque-country/Study_area.shp"),

  brt_sp_non = brt_sp_nsp(data = preprocessed_data, iterations = 200),
  preprocess.md = knit("03_scripts/01_preprocessing.Rmd"),
  strings_in_dots = "literals"
```

I put the `.Rmd` file `01_preprocessing.Rmd` into a function called `preprocess()` that gets imported via `source("03_scripts/01_preprocessing.R")`.

I generate `preprocessed_data` which serves as input for `brt_sp_nsp()` which itself is defined in `source("03_scripts/server/modeling_functions.R")`.

However, when following your [quick example](https://wlandau-lilly.github.io/drake/articles/quickstart.html), to generate a report for all the preprocessing I have to add something like `preprocess.md = knit("03_scripts/01_preprocessing.Rmd")` that includes
`readd()` or `loadd()` for the datasets, .e.g. `readd(pathogens)`, because they already exist in the cache.
If I wouldn't do so, dependencies are not detected, right?

So currently I would need an `.R` file defining my function (e.g. `01_preprocessing.R`) and an additional `.Rmd` file for the report, both with 99% redundant code?

Hm, this really gets kinda complicated now and it seems that have to modify all my scripts..
Maybe I overcomplicate things here?
Actually, my analysis structure is "very simple" and I only would need a tool that takes care of the following structure.

* `01_preprocessing.Rmd` `#1`
* `02_EDA.Rmd` (depends on `#1`)
* `03_study_area.Rmd` (depends on `#1`)
* `04_scripts/20 files here.R` (depend on `#1`)
* `05_spcvis.Rmd` (depends on `#1` and `#4`)
* `06_post_analysis.Rmd` (depends on `#1` and `#4`)

`drake` is very function orientated as far as I see and less focused on resolving simple dependencies of scripts? I understand that there needs to be a clear link between all the scripts to be able to connect them.
However, wouldn't be something like the example below also desirable?

```r
drake_plan(
1 = report("01_preprocessing.Rmd", depends_on = NULL),
2 = report("02_EDA.Rmd", depends_on = 1),
3 = report("03_study_area.Rmd", depends_on = 1),
4 = script(list("04_scripts/20 files here.R"), depends_on = 1)
5 = report("05_spcvis.Rmd", depends on = c(1, 4)
)
```
