### 2018-01-23 #156 @maelle
Thanks a lot @gothub!

Rich discussion here! @wlandau-lilly please make sure to insert all clarifications from this thread into the documentation 😉

When you're done update this thread with a summary of answers to reviewers (a short one per review) in order to make it easier for them to sign off your changes even if they have little time. ☺ good job on using their comments until now!
