### 2018-01-23 wlandau-lilly/drake#193 @wlandau-lilly
I think we're making progress here.

> However, when following your quick example, to generate a report for all the preprocessing I have to add something like preprocess.md = knit("03_scripts/01_preprocessing.Rmd") that includes
readd() or loadd() for the datasets, .e.g. readd(pathogens), because they already exist in the cache.
If I wouldn't do so, dependencies are not detected, right?

> So currently I would need an .R file defining my function (e.g. 01_preprocessing.R) and an additional .Rmd file for the report, both with 99% redundant code?

No additional `.Rmd` report should be necessary. In the [quckstart vignette](https://wlandau-lilly.github.io/drake/articles/quickstart.html) you mentioned, the `report.Rmd` file is for post-processing, not preprocessing. Its purpose is to display and inspect the results at the very end of the analysis. I intend to make that clearer very soon. I recommend against `.Rmd` files for preprocessing since all you get is an final `.md` file rather than targets that would be useful to future commands. If I were you, I would completely let go of `01_preprocessing.Rmd`.

> drake is very function orientated as far as I see and less focused on resolving simple dependencies of scripts?

Yes, `drake` resolves dependencies of functions and commands, not scripts. R Markdown files are a special exception because people like to use them as final reports. I would encourage you to write your workflow kind of like this.

```r
drake_plan(
  preprocessed_data = preprocessing_01(),
  EDA = EDA_02(preprocessed_data),
  study_area = study_area_03(preprocessed_data),
  other_work = other_work_04(preprocessed_data),
  spcvis = spcvis_05(preprocessed_data, other_work)
)
```

You don't need anything like `depends_on` because drake knows the targets depend on the arguments supplied to functions in commands. For example, `drake` knows that `spcvis` depends on `preprocessed_data` and `other_work` because these two targets are passed as arguments to your  `spcvis_05()` function.
