### 2018-01-23 779a083
Address #193
There was some confusion about the role of
R Markdown reports in drake workflows.
They are not actually necessary, and
the only targets they generate are report files
such as `.md` and `.html` files. I have changed
the report in the basic example to try to explain.
@pat-s, I hope this helps.
