### 2018-01-23 wlandau-lilly/drake#193 @wlandau-lilly
Something else I forgot to clarify: if one of your commands is `knit('some_report.Rmd')` or  `render('some_report.Rmd')`, your target will be another file such as `some_report.md` or `some_report.html`. You will not be able to extract any of the data objects you make inside the code chunks of `some_report.Rmd`. In other words, `some_report.Rmd` is a dead end.
