### 2018-01-23 c21ea42
Expand reproducibility section of the README
I will keep trying to rephrase and re-explain drake's brand of reproducibility
for as long as the concept is lost in translation.
In this latest attempt, I try to adopt an independent reviewer's perspective
to address @gothub's comments from ropensci/onboarding#156.
