### 2018-01-23 020f4fb
Clarify high-performance computing claims
@gothub and @benmarwick commented on drake's
high-performance computing claims for ropensci/onboarding#156.
In this commit, I attempt to spell out what those claims are.
I have a new "Caveats" section of the parallelism vignette
and an analogous "practical utility of parallel computing"
subsection of the cautionary vignette. In addition, I made sure
to explain that the runtime prediction features were experimental
and the results have not been rigorously verified yet.
