+++ R/max_useful_jobs.R
@@ -1,9 +1,11 @@
-#' @title Determine the maximum number of useful jobs in the next call
+#' @title Suggest an upper bound on the jobs in the next call
 #' to \code{make(..., jobs = YOUR_CHOICE)}.
-#' @description Any additional jobs more than \code{max_useful_jobs(...)}
+#' @description The best number of \code{jobs}
+#' should be somewhere between 1 and \code{max_useful_jobs(...)}.
+#' Any additional jobs more than \code{max_useful_jobs(...)}
 #' should be superfluous, and could even slow you down for
 #' \code{make(..., parallelism = 'parLapply')}.
-#' Set the \code{imports} argument to change your assumptions about
+#' @details Set the \code{imports} argument to change your assumptions about
 #' how fast objects/files are imported.
 #' IMPORTANT: you must be in the root directory of your project.
 #' @export
