+++ vignettes/parallelism.Rmd
@@ -78,7 +78,7 @@ Be mindful of the maximum number of simultaneous parallel jobs you deploy. Conse
 
 ## Drake can report the maximum number of useful simultaneous jobs
 
-The `max_useful_jobs()` function analyzes your project and recommends a maximum value for the `jobs` argument to the next `make()` (or the `workers` argument to a backend function in [future](https://github.com/HenrikBengtsson/future)). 
+The `max_useful_jobs()` function analyzes your project and recommends a maximum value for the `jobs` argument to the next `make()` (or the `workers` argument to a backend function in [future](https://github.com/HenrikBengtsson/future)). This number returned by `max_useful_jobs()` is only an upper bound, not necessarily the number of `jobs` you should choose. 
 
 ```{r hpcquick, eval = FALSE}
 library(drake)
@@ -108,6 +108,20 @@ max_useful_jobs(config, imports = "files") # 4
 max_useful_jobs(config, imports = "all") # 8
 max_useful_jobs(config, imports = "none") # 4
 ```
+
+## Caveats
+
+`Drake` claims that it can
+
+1. Build and cache your targets in parallel (in stages).
+2. Build and cache your targets in the correct order, finishing dependencies before starting targets that depend on them.
+3. Deploy your targets to the parallel backend of your choice.
+
+However, the practical efficiency of the parallel computing functionality remains to be verified rigorously. Serious performance studies will be part of future work that has not yet been conducted at the time of writing. In addition, each project has its own best parallel computing set up, and the user needs to optimize it on a case-by-case basis. Some general considerations include the following.
+
+- The high overhead high scalability of distributed computing versus the low overhead and low scalability of local multicore computing.
+- The high memory usage of local multicore computing, especially `"mclapply"` parallelism, as opposed to distributed computing, which can spread the memory demands over the available nodes on a cluster.
+- The marginal gains of increasing the number of jobs indefinitely, especially in the case of local multicore computing if the number of cores is low.
 
 # Parallel backends
 
