### 2018-01-23 63d33ea
Add a new best practices vignette
Supports #193
