### 2018-01-23 c07f8d2
Repair best practices vignette
