+++ vignettes/best-practices.Rmd
@@ -51,13 +51,13 @@ my_plan
 
 `Drake` knows that `my_analysis` depends on `my_data` because `my_data` is an argument to `analyze_data()`, which is part of the command for `my_analysis`.
 
-```{r visbad, eval = FALSE}
+```{r visgood, eval = FALSE}
 config <- drake_config(my_plan)
 vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/good-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/63d33ea8/images/good-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -99,7 +99,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/bd8a086f/images/bad-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/63d33ea8/images/bad-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
