### 2018-01-23 738a26e
Recompute images with new titles
