+++ images/images.R
@@ -81,6 +81,7 @@ my_plan <- drake_plan(
 )
 config <- drake_config(my_plan)
 vis_drake_graph(
+  main = "Good workflow plan",
   config, file = "good-commands.html", selfcontained = TRUE,
   width = "100%", height = "500px"
 )
@@ -97,6 +98,7 @@ my_plan <- drake_plan(
 )  
 config <- drake_config(my_plan)
 vis_drake_graph(
+  main = "Bad workflow plan",
   config, file = "bad-commands.html", selfcontained = TRUE,
   width = "100%", height = "500px"
 )
