### 2018-01-23 0359069
Recompute images with better hover text
