### 2018-01-23 4d2ec5d
Update image hashes
