+++ vignettes/best-practices.Rmd
@@ -57,7 +57,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/738a26e8/images/good-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/0359069c/images/good-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -99,7 +99,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/738a26e8/images/bad-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/0359069c/images/bad-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
