### 2018-01-23 21f83e6
Add a new data.csv to best practices guide
