+++ vignettes/best-practices.Rmd
@@ -24,6 +24,7 @@ knitr::opts_chunk$set(
   error = TRUE,
   warning = TRUE
 )
+tmp <- file.create("data.csv")
 ```
 
 This vignette describes general best practices for creating, configuring, and running `drake` projects.
@@ -41,7 +42,7 @@ Then, set up your workflow plan data frame.
 
 ```{r storecode1}
 good_plan <- drake_plan(
-  my_data = get_data(),
+  my_data = get_data('data.csv'), # External files need to be in commands explicitly. # nolint
   my_analysis = analyze_data(my_data),
   my_summaries = summarize_results(my_data, my_analysis)
 )
@@ -107,7 +108,8 @@ style="border: none; box-shadow: none">
 Dangers:
 
 1. `my_analysis` will not update when `my_data` changes.
-2. In `make(bad_plan, jobs = 2)`, `drake` will try to build `my_data` and `my_analysis` at the same time even though `my_data` must be built first.
+2. `Drake` does not know about the `data.csv` file at all since it is not explicitly mentioned in a workflow plan command. So when `data.csv` changes, `make(bad_plan)` will not rebuild `my_data`.
+3. In `make(bad_plan, jobs = 2)`, `drake` will try to build `my_data` and `my_analysis` at the same time even though `my_data` must be built first.
 
 In addition, this `source()`-based approach is simply inconvenient. `Drake` rebuilds `my_data` every time `get_data.R` changes, even when those changes are just extra comments or blank lines. On the other hand, in the previous plan that uses `my_data = get_data()`, `drake` does not trigger rebuilds when comments or whitespace in `get_data()` are modified. `Drake` is R-focused, not file-focused. If you embrace this viewpoint, your work will be easier.
 
@@ -139,5 +141,11 @@ style="border: none; box-shadow: none">
 
 ```{r rmfiles_caution, echo = FALSE}
 clean(destroy = TRUE, verbose = FALSE)
-unlink(c("Makefile", "report.Rmd", "shell.sh", "STDIN.o*", "Thumbs.db"))
+file.remove()
+unlink(
+  c(
+    "data.csv", "Makefile", "report.Rmd",
+    "shell.sh", "STDIN.o*", "Thumbs.db"
+  )
+)
 ```
