### 2018-01-23 ccb4da8
Touch up best practices vignette
