+++ vignettes/best-practices.Rmd
@@ -58,7 +58,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/0359069c/images/good-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/21f83e6a/images/good-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
@@ -91,7 +91,6 @@ bad_plan <- drake_plan(
 bad_plan
 ```
 
-
 But now, the dependency structure of your work is broken. Your R script files are dependencies, but since `my_data` is not mentioned in a function or command, `drake` does not know that `my_analysis` depends on it.
 
 ```{r visbad, eval = FALSE}
@@ -100,7 +99,7 @@ vis_drake_graph(config)
 ```
 
 <iframe
-src = "https://cdn.rawgit.com/wlandau-lilly/drake/0359069c/images/bad-commands.html"
+src = "https://cdn.rawgit.com/wlandau-lilly/drake/21f83e6a/images/bad-commands.html"
 width = "100%" height = "600px" allowtransparency="true"
 style="border: none; box-shadow: none">
 </iframe>
