### 2018-01-23 wlandau-lilly/drake#193 @wlandau-lilly
@pat-s please see the new [best practices guide](https://wlandau-lilly.github.io/drake/articles/best-practices.html). I try to lay out the issue there as best I can, and I will likely add more in the future. I think I have done enough to close here, but let's keep talking on the thread.
