### 2018-01-23 #156 @wlandau-lilly
Also, I realized that I totally missed the point of one of your earlier questions.

> Very naive question, does each command need to be something as basic as summary or could it be sourcing a larger script containing several regression calls?

Commands can be arbitrary code chunks. For example:

```r
my_plan <- drake_plan(
  a = 1 + 1,
  b = {
    x <- pi + a
    y <- sqrt(x)
    rand <- rnorm(10, sd = y)
    mean(rand)
  },
  c = a - 5,
  d = c(b, c)
)
```

However, a command should **NOT** just `source('large_command.R')`. `Drake` is function-oriented, not file-oriented. I am beginning to see widespread confusion here (for example, wlandau-lilly/drake#193), and I am doing my best to fix it. I have a new [best practices guide](https://wlandau-lilly.github.io/drake/articles/best-practices.html) that spells out the pitfalls, and I plan to add more over the course of future development.
