### 2018-01-23 5a3de4a
Reorder cautionary notes about source()
