+++ vignettes/best-practices.Rmd
@@ -106,9 +106,9 @@ style="border: none; box-shadow: none">
 
 Dangers:
 
-1. `my_analysis` will not update when `my_data` changes.
-2. `Drake` does not know about the `data.csv` file at all since it is not explicitly mentioned in a workflow plan command. So when `data.csv` changes, `make(bad_plan)` will not rebuild `my_data`.
-3. In `make(bad_plan, jobs = 2)`, `drake` will try to build `my_data` and `my_analysis` at the same time even though `my_data` must be built first.
+1. In the first `make(bad_plan, jobs = 2)`, `drake` will try to build `my_data` and `my_analysis` at the same time even though `my_data` must finish before `my_analysis` begins.
+2. `Drake` is oblivious to `data.csv` since it is not explicitly mentioned in a workflow plan command. So when `data.csv` changes, `make(bad_plan)` will not rebuild `my_data`.
+3. `my_analysis` will not update when `my_data` changes.
 4. The return value of `source()` is formatted counter-intuitively. If `source('get_data.R')` is the command for `my_data`, then `my_data` will always be a list with elements `"value"` and `"visible"`. In other words, `source('get_data.R')$value` is really what you want.
 
 In addition, this `source()`-based approach is simply inconvenient. `Drake` rebuilds `my_data` every time `get_data.R` changes, even when those changes are just extra comments or blank lines. On the other hand, in the previous plan that uses `my_data = get_data()`, `drake` does not trigger rebuilds when comments or whitespace in `get_data()` are modified. `Drake` is R-focused, not file-focused. If you embrace this viewpoint, your work will be easier.
