### 2018-01-23 #156 @wlandau-lilly
# Summary of @wlandau-lilly's responses to reviewers

`Drake` is large and difficult to review, and I am grateful for all your hard work. `Drake` is a much better package because of your feedback.

## @benmarwick

### Changes to `drake`

- Add a [new section to the README](https://github.com/wlandau-lilly/drake/tree/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c#similar-work) that compares `drake` to similar work: [GNU Make](www.gnu.org/software/make/), [remake](https://github.com/richfitz/remake), [memoise](https://github.com/r-lib/memoise), and [knitr](https://yihui.name/knitr/).
- In this new section, argue that `drake` surpasses these tools for its intended use case.
- [List real-world projects](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#gnu-make) that use Make as a reproducible pipeline toolkit.
- [Reference a list](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#remake) of [examples of real-world remake projects](https://github.com/richfitz/remake/blob/master/README.md#real-world-examples).
- Add two realistic demonstrations: [example-gsp.Rmd](https://github.com/wlandau-lilly/drake/blob/master/vignettes/example-gsp.Rmd) and [example-packages.Rmd](https://github.com/wlandau-lilly/drake/blob/master/vignettes/example-packages.Rmd). Both are fabricated, but they use real data to answer real questions. They help users see the practical utility of `drake` for their own projects.
- [example-gsp.Rmd](https://github.com/wlandau-lilly/drake/blob/master/vignettes/example-gsp.Rmd) shows how `drake` surpasses Make.
- Move [CONTRIBUTING.md](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/CONTRIBUTING.md) to the top level for better visibility.
- Add a [paper.md](https://github.com/wlandau-lilly/drake/blob/master/paper.md) and [paper.bib](https://github.com/wlandau-lilly/drake/blob/master/paper.bib) for JOSS.

### Comments

- Unfortunately, I cannot yet find publicly-available real-world projects or publications using `drake`. I think it is still too early.
- I have now clarified `drake`'s high-performance computing claims [here](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/parallelism.Rmd#caveats).

## @jules32

### Changes to `drake`

- Change the [installation instructions](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#installation):
    - Remove `build = TRUE` from `install_github()` so the ["Suggests:" packages](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/DESCRIPTION#L57) such as `Ecdat` are not required.
    - Remove superfluous instructions for installing a GitHub tag/release.
- Add a [list of useful functions](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#documentation) and a link to the [documentation website's reference section](https://wlandau-lilly.github.io/drake/reference/index.html) at the top of the [README's "Documentation" section](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#documentation).
- Use a [numbered list here](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#what-gets-done-stays-done) to show how `drake` conquers the [Sisyphean loop](https://en.wikipedia.org/wiki/Sisyphus) of naive data science.
- Comment the code chunk in [this section](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#what-gets-done-stays-done) in the way @jules32 suggested.
- Add useful titles to functions in the help files and the [website's reference section](https://wlandau-lilly.github.io/drake/reference/index.html). (Previously, they all said "Function f" or equivalent.)
- Change the website tab title from "Data Frames in R for Make - Drake" to just "drake".
- Add a [paper.md](https://github.com/wlandau-lilly/drake/blob/master/paper.md) and [paper.bib](https://github.com/wlandau-lilly/drake/blob/master/paper.bib) for JOSS.

### Comments

- Previously, you mentioned that your review was in progress. What are your plans for the remainder?

## @gothub

### Changes to `drake`

- At the top of the [drake.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/drake.Rmd), reference both the [quickstart vignette](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/quickstart.Rmd) and the analogous [quickstart page on the documentation website](https://wlandau-lilly.github.io/drake/articles/quickstart.html).
- Reference `drake_example("basic")` wherever `load_basic_example()` is called in the documentation. That way, users know how to obtain the underlying code.
- Expose the annotated listing of `drake_examples()` [here in the README](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#documentation), [here in the drake.Rmd vignette](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/drake.Rmd#built-in-example-projects), and [here in the Get Started page](https://wlandau-lilly.github.io/drake/articles/drake.html#built-in-example-projects).
- Clarify exactly what `drake` is claiming about its parallel computing functionality: https://github.com/wlandau-lilly/drake/commit/020f4fb7e452765c9e138e07e5f64a28aa83711e. See [here in the parallelism vignette](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/parallelism.Rmd#caveats) and [here in the cautionary vignette](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/vignettes/caution.Rmd#the-practical-utility-of-parallel-computing).
- Expand the ["Stay Reproducible" section of the README](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#stay-reproducible) to concretely spell out `drake`'s role when it comes to reproducibility.

### Comments

- I have improved and expanded the ["Stay Reproducible" section of the README](https://github.com/wlandau-lilly/drake/blob/5a3de4ab0cdd5226c62c94c0c9a26432f7ad0a0c/README.md#stay-reproducible) since my last comment on the rOpenSci thread. I hope I have done a better job explaining how `drake` helps with reproducibility.
- With the `run` argument to `devtools::run_examples()`, you can override the `\dontrun{}` command and execute **all** the code in the examples. I routinely run `devtools::run_examples(run = TRUE)` and `devtools::run_examples(run = FALSE)`.
- Regarding reproducibility, you mentioned consistency, which I took to mean stability. If you run a project from scratch a large number of times, will new results always be consistent with the previous ones? It is an important concern, especially for stochastic methods such as Markov chain Monte Carlo, but it is outside the scope of `drake`.
