### 2018-01-23 #156 @maelle
@jules32 @benmarwick @gothub are you happy with the changes made?

Thanks @wlandau-lilly for the summary, and great on having used all feedback! 😸
