### 2018-01-23 91b07f5
Change quickstart vignette subtitle
