### 2018-01-23 #156 @wlandau-lilly
Thank you, @benmarwick! I am so glad you think the changes are making a difference.

By the way, I just updated the review summaries in light of  wlandau-lilly/drake#195.
