### 2018-01-24 3a900c8
Rephrase a cautionary note
