### 2018-01-24 #156 @jules32
Yes, I'm happy too. You've addressed the ideas I had and think that this is good to continue moving forward. Nice work!
