### 2018-01-24 #156 @wlandau-lilly
Thank you so much, @gothub and @jules32! It is exciting to have confirmation from all three reviewers!

@gothub, I really like your question. I will try to break down the answer.

### R scripts

As you say, we need any code that writes functions, generates workflow plans, loads packages, and calls `make()`.

### The workflow plan

In the way I use `drake` (and the way I picture others using it), the workflow plan does not actually need to be saved as a csv file. Typically, it is super quick to programmatically generate in an R script before `make()` is called. I usually keep a `plan.R` file that uses `drake_plan()`, `plan_analyses()`, `plan_summaries()`, `expand_plan()`, and `evaluate_plan()` to generate the plan. The advantage here is that you can typically spin up thousands of targets with around 10-20 lines of human-readable code. (With `remake` or Make, you would have to write it all out by hand, which is often intractable.) Then, I `source("plan.R")` before calling `make()`.

### Input files and other imported objects

Examples include optional `report.Rmd` files, other knitr reports, user-defined CSV data files, and functions created in the R scripts. It's always a good idea to check for missing input files and imported functions/objects. You could either use the `missed()` function, or you could use `vis_drake_graph()` and look for the purple nodes, which indicate missing objects or files.

### R packages

Packages are tricker. Maintaining a reproducible package environment is beyond the scope of `drake`, and the documentation recommends [packrat](https://rstudio.github.io/packrat/) for this purpose (mentioned in the quickstart and caution vignettes). But if you do not use `packrat`, there is still hope. By default, each `make()` saves the `sessionInfo()` to the cache (use the `session_info` argument to deactivate). An independent researcher just needs to call `drake_session()` to learn about the required versions of R, packages, and other system requirements.

### The cache

Including the `drake` cache (Usually a hidden `.drake/` folder) lets independent researchers use `drake_session()` to learn about the package environment, and it allows them to call `make()` to check that everything is up to date (and thus re-creatable). This can be tricky: by design, the `drake` cache (which is just a `storr::storr_rds()`) divides its data over lots of tiny files, which is essential to avoid race conditions in parallel computing. However, this setup behaves poorly under version control, and it may even make it a bit annoying to use a shared file system like Dropbox or Google Drive.

For the version control bit, @noamross recently suggested that `drake` supply a flat text log file to give a list of targets and their fingerprints. That way, the git commit history of the log tracks changes to the targets over time. Right now, `drake` supports `make(..., cache_log_file = "my_log_file.txt")`,
 `make(..., cache_log_file = TRUE)`, and `drake_cache_log_file()`.
