### 2018-01-24 #156 @gothub
@wlandau-lilly Thanks for the explanation of `drake` persistence.

Yes, please include me as a listed reviewer of the package, and thanks!
