### 2018-01-24 #156 @wlandau-lilly
By the way, an [Appveyor build](https://ci.appveyor.com/project/wlandau-lilly/drake/build/1.0.1032#L225) failed just now because `listenv` is in the middle of an update on CRAN.

```
  There are binary versions available (and will be installed) but the
  source versions are later:
        binary source
listenv  0.6.0  0.7.0
tibble   1.4.1  1.4.2
Error in download.file(url, destfile, method, mode = "wb", ...) :
  cannot open URL 'https://cran.rstudio.com/bin/windows/contrib/3.4/listenv_0.6.0.zip
```

In my experience, this kind of error just blows over if I wait a bit. Would that be okay? I ask because my sense is that `drake` is about to be approved, so the timing is not ideal.
