### 2018-01-24 #156 @maelle
No problem, I'll test it on Windows later this week when I do the last checks!
