### 2018-01-25 d2eccff
Re-word the reproducibility section of the README
ropensci/onboarding#156
