### 2018-01-25 #156 @maelle
Approved! Thanks everyone for a productive and (I hope) enjoyable review process!

To-dos

- [ ] Transfer the repo to the rOpenSci organization under "Settings" in your repo.  I have invited you to one team that should allow you to do so.
- [ ] Fix any links in DESCRIPTION, and in badges for CI and coverage to point to the ropensci URL. (we'll turn on the services on our end.)
- [ ] Submit to JOSS using the Zenodo DOI. We will tag it for expedited review.

Welcome aboard! We'd also love a blog post about your package, either a short-form intro to it (https://ropensci.org/tech-notes/) or long-form post with more narrative about its development. ((https://ropensci.org/blog/). Let me know if you are interested, I'd ping rOpenSci community manager.
