### 2018-01-25 #156 @wlandau
Wonderful! I transferred the repo, which was super exciting. One small thing: I do not have access to the "Settings" tab, and I cannot change the link to the documentation website at the top. Am I just over-eager, or is there another step?

![capture](https://user-images.githubusercontent.com/1580860/35407264-5c78d112-01d9-11e8-9fdf-295b2c988d2a.PNG)
