### 2018-01-25 #156 @wlandau
I think I found the problem, and it is probably a common issue. @maelle, would you add me as a maintainer at https://github.com/orgs/ropensci/teams/drake/members?

![maintainer](https://user-images.githubusercontent.com/1580860/35407597-801dc536-01da-11e8-9681-7ec4939b2b25.png)
