### 2018-01-25 #156 @wlandau
Also, before I create a release for Zenodo, I think I should confirm with @jules32 and @benmarwick about being [listed as reviewers](https://github.com/ropensci/drake/blob/649780c2cbb0a7f1241b48926edf588d7c78bc05/DESCRIPTION#L28) in the package itself.
