### 2018-01-25 #156 @jules32
Hi All,

This is exciting to see happen! I am fine not being listed as a reviewer; I
hope that my suggestions were helpful but, as they were minimal, I don't
think I need to be included in the DESCRIPTION.

But thank you for asking!

On Thu, Jan 25, 2018 at 12:18 PM, Will Landau <notifications@github.com>
wrote:

> Also, before I create a release for Zenodo, I think I should confirm with
> @jules32 <https://github.com/jules32> and @benmarwick
> <https://github.com/benmarwick> about being listed as reviewers
> <https://github.com/ropensci/drake/blob/649780c2cbb0a7f1241b48926edf588d7c78bc05/DESCRIPTION#L28>
> in the package itself.
>
