### 2018-01-26 #156 @stefaniebutland
Hello @wlandau-lilly. I'm rOpenSci's Community Manager.
> I have prepared a narrative-form blog post, and I am excited to share it. Is there somewhere I can submit a PR?
We'd love to have a post about this. Please submit a PR according to: https://github.com/ropensci/roweb2#contributing-a-blog-post. I will review it, provide some feedback and a suggested date to publish it.

Thank you!
