### 2018-01-26 #156 @wlandau
Thanks, @stefaniebutland! I just submitted ropensci/roweb2#140. Really looking forward to the next steps.
