### 2018-01-26 6cbbe94
Add rOpenSci footer
ropensci/onboarding#156
