### 2018-01-26 #156 @wlandau
@maelle, I have added @benmarwick as a reviewer in the DESCRIPTION (permission granted by email), and I have completed the 5 to-dos from [here](https://github.com/ropensci/onboarding/issues/156#issuecomment-360559884). I also submitted `drake` 5.0.0 to CRAN, added an rOpenSci footer to the README, and [volunteered as a reviewer](https://ropensci.org/onboarding/). This is all super exciting!

Besides the separate processes for the blog entry and the JOSS paper, there are a couple minor loose ends that will probably resolve themselves.
1. [AppVeyor CI](https://ci.appveyor.com/project/ropensci/drake) is still deactivated.
2. The rOpenSci README badge still says "Under Review" in the [GitHub README](https://github.com/ropensci/drake/blob/master/README.md), but the same badge says "Peer Reviewed" on the [documentation website](https://ropensci.github.io/drake/).
