### 2018-01-26 #156 @maelle
1. Now activated, here is the badge code `[![Build status](https://ci.appveyor.com/api/projects/status/4ypc9xnmqt70j94e?svg=true)](https://ci.appveyor.com/project/ropensci/drake)`
2. Yes they take a while to propagate, but tell us if it's still the case on Monday for instance.
