### 2018-01-26 #156 @wlandau
Thanks, @maelle. The AppVeyor builds appear to be working. I will check back on the badge in a few days.
