### 2018-01-26 810b97d
Shorten a sentence
