### 2018-01-26 #156 @wlandau
Another minor thing: I am getting emails like this one on every commit:

>Build ropensci/drake@810b97d5ba successful: https://ropensci.ocpu.io/drake/
>
>NEW COMMITS:
[810b97d5ba]
  Time: 2018-01-26 08:39:16
>...

Is there a way to attenuate them? Are the other watchers receiving these emails, or just the maintainers? I do not see an OpenCPU webhook for drake, so I am guessing it is at the org level.
