### 2018-01-26 #156 @jeroen
Those emails are from the CI system, they are sent to the pusher only. You can ignore/filter them.
