### 2018-01-26 ropensci/drake#193 @wlandau
I have been thinking a lot more about this thread, and I have restructured `drake_example("basic")`, `drake_example("gsp")`, and `drake_example("packages")` to demonstrate how to set up the files. Didn't make it to CRAN v5.0.0, but there's always next time. The files are on GitHub now. If I were you, I would peruse the following to learn more about setting up the files for `drake` projects.

- https://github.com/ropensci/drake/tree/master/inst/examples/basic
- https://github.com/ropensci/drake/tree/master/inst/examples/gsp
- https://github.com/ropensci/drake/tree/master/inst/examples/packages
