### 2018-01-26 ropensci/drake#193 @pat-s
Always good to hear when a discussion ends fruitful! Haven't had time in the last days to take a look in detail, will do it later and report my thoughts!
