### 2018-01-26 ropensci/drake#193 @wlandau
And I look forward to learning what you think. This is such an important piece for `drake`'s docs to get right.
