### 2018-01-27 #156 @wlandau
The README badge now says "Peer Reviewed", and the blog entry at ropensci/roweb2#140 is going well. I think we can safely close this issue.
